### 2020-12-22 #2926 @whedon
[PRE REVIEW]: GeometricIntegrators.jl: Geometric Numerical Integration in Julia
**Submitting author:** @michakraus (<a href="http://orcid.org/0000-0001-7637-3979">Michael Kraus</a>)
**Repository:** <a href="https://github.com/JuliaGNI/GeometricIntegrators.jl" target ="_blank">https://github.com/JuliaGNI/GeometricIntegrators.jl</a>
**Version:** v0.7.0
