+++ Project.toml
@@ -39,7 +39,7 @@ Parameters = "0.12"
 Polynomials = "^1.1.13"
 ProgressMeter = "1"
 QuadratureRules = "0.1"
-Reexport = "0.2"
+Reexport = "0.2, 1.0"
 SimpleSolvers = "0.1"
 julia = "^1.4"
 
