### 2021-01-07 c4e739e
Add missing DOI to JOSS paper references.
