+++ docs/joss/paper.bib
@@ -23,7 +23,8 @@
 	Author = {Blanes, Sergio and Casas, Fernando},
 	Publisher = {CRC Press},
 	Title = {A Concise Introduction to Geometric Numerical Integration},
-	Year = {2016}}
+	Year = {2016},
+	doi = {10.1201/b21563}}
 
 
 @article{Bezanson:2017,
