### 2021-01-07 #2926 @michakraus
> Hi @michakraus could you check the missing DOI suggested by whedon?

Thanks, I added the DOI.

> Also, could you look at our [potential reviewer list](https://bit.ly/joss-reviewers) and maybe suggest reviewers (without tagging them with an @).? You can also suggest names outside of the list, just be aware of our [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)

Potential reviewers from the list are ChrisRackauckas, Datseris and ranocha.
