### 2021-01-22 JuliaGNI/GeometricIntegrators.jl#106 @ranocha
Example in the README.md fails
I get the following error using Julia v1.5.3.
```julia
pkg> st
Status `Project.toml`
  [dcce2d33] GeometricIntegrators v0.7.0
  [91a5bcdd] Plots v1.10.1

julia> using GeometricIntegrators

julia> ode = ODE((t, x, ẋ) -> ẋ[1] = x[1], [1.0]);

julia> int = Integrator(ode, TableauExplicitEuler(), 0.1);

julia> sol = integrate(ode, int, 10);

julia> using Plots

julia> plot(xlims=[0,1], xlab="t", ylab="x(t)", legend=:bottomright)

julia> plot!(sol.t, sol.q[1,:], label="numeric")
Error showing value of type Plots.Plot{Plots.GRBackend}:
ERROR: BoundsError: attempt to access 11-element OffsetArray(::Array{Float64,1}, 0:10) with eltype Float64 with indices 0:10 at index [0:11]
Stacktrace:
 [1] throw_boundserror(::OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}}, ::Tuple{UnitRange{Int64}}) at ./abstractarray.jl:541
 [2] checkbounds at ./abstractarray.jl:506 [inlined]
 [3] _getindex at ./multidimensional.jl:742 [inlined]
 [4] getindex(::OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}}, ::UnitRange{Int64}) at ./abstractarray.jl:1060
 [5] gr_draw_segments(::Plots.Series, ::OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}}, ::Array{Float64,1}, ::Nothing, ::Tuple{Float64,Float64}) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:1645
 [6] gr_add_series(::Plots.Subplot{Plots.GRBackend}, ::Plots.Series) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:1560
 [7] gr_display(::Plots.Subplot{Plots.GRBackend}, ::Measures.Length{:mm,Float64}, ::Measures.Length{:mm,Float64}, ::Array{Float64,1}) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:935
 [8] gr_display(::Plots.Plot{Plots.GRBackend}, ::String) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:626
 [9] gr_display at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:593 [inlined]
 [10] _display(::Plots.Plot{Plots.GRBackend}) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:1845
 [11] display(::Plots.PlotsDisplay, ::Plots.Plot{Plots.GRBackend}) at ~/.julia/packages/Plots/lmp2A/src/output.jl:150
 [12] display(::Any) at ./multimedia.jl:328
 [13] #invokelatest#1 at ./essentials.jl:710 [inlined]
 [14] invokelatest at ./essentials.jl:709 [inlined]
 [15] print_response(::IO, ::Any, ::Bool, ::Bool, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/REPL.jl:238
 [16] print_response(::REPL.AbstractREPL, ::Any, ::Bool, ::Bool) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/REPL.jl:223
 [17] (::REPL.var"#do_respond#54"{Bool,Bool,REPL.var"#64#73"{REPL.LineEditREPL,REPL.REPLHistoryProvider},REPL.LineEditREPL,REPL.LineEdit.Prompt})(::Any, ::Any, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/REPL.jl:822
 [18] (::REPL.var"#69#78")(::Any, ::Any, ::Vararg{Any,N} where N) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/REPL.jl:1086
 [19] #invokelatest#1 at ./essentials.jl:710 [inlined]
 [20] invokelatest at ./essentials.jl:709 [inlined]
 [21] (::REPL.LineEdit.var"#22#23"{REPL.var"#69#78",String})(::Any, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/LineEdit.jl:1364
 [22] prompt!(::REPL.Terminals.TextTerminal, ::REPL.LineEdit.ModalInterface, ::REPL.LineEdit.MIState) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/LineEdit.jl:2447
 [23] run_interface(::REPL.Terminals.TextTerminal, ::REPL.LineEdit.ModalInterface, ::REPL.LineEdit.MIState) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/LineEdit.jl:2350
 [24] run_frontend(::REPL.LineEditREPL, ::REPL.REPLBackendRef) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/REPL.jl:1144
 [25] (::REPL.var"#38#42"{REPL.LineEditREPL,REPL.REPLBackendRef})() at ./task.jl:356

julia> plot!(sol.t, exp.(sol.t), label="exact")
Error showing value of type Plots.Plot{Plots.GRBackend}:
ERROR: BoundsError: attempt to access 11-element OffsetArray(::Array{Float64,1}, 0:10) with eltype Float64 with indices 0:10 at index [0:11]
Stacktrace:
 [1] throw_boundserror(::OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}}, ::Tuple{UnitRange{Int64}}) at ./abstractarray.jl:541
 [2] checkbounds at ./abstractarray.jl:506 [inlined]
 [3] _getindex at ./multidimensional.jl:742 [inlined]
 [4] getindex(::OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}}, ::UnitRange{Int64}) at ./abstractarray.jl:1060
 [5] gr_draw_segments(::Plots.Series, ::OffsetArrays.OffsetArray{Float64,1,Array{Float64,1}}, ::Array{Float64,1}, ::Nothing, ::Tuple{Float64,Float64}) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:1645
 [6] gr_add_series(::Plots.Subplot{Plots.GRBackend}, ::Plots.Series) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:1560
 [7] gr_display(::Plots.Subplot{Plots.GRBackend}, ::Measures.Length{:mm,Float64}, ::Measures.Length{:mm,Float64}, ::Array{Float64,1}) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:935
 [8] gr_display(::Plots.Plot{Plots.GRBackend}, ::String) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:626
 [9] gr_display at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:593 [inlined]
 [10] _display(::Plots.Plot{Plots.GRBackend}) at ~/.julia/packages/Plots/lmp2A/src/backends/gr.jl:1845
 [11] display(::Plots.PlotsDisplay, ::Plots.Plot{Plots.GRBackend}) at ~/.julia/packages/Plots/lmp2A/src/output.jl:150
 [12] display(::Any) at ./multimedia.jl:328
 [13] #invokelatest#1 at ./essentials.jl:710 [inlined]
 [14] invokelatest at ./essentials.jl:709 [inlined]
 [15] print_response(::IO, ::Any, ::Bool, ::Bool, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/REPL.jl:238
 [16] print_response(::REPL.AbstractREPL, ::Any, ::Bool, ::Bool) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/REPL.jl:223
 [17] (::REPL.var"#do_respond#54"{Bool,Bool,REPL.var"#64#73"{REPL.LineEditREPL,REPL.REPLHistoryProvider},REPL.LineEditREPL,REPL.LineEdit.Prompt})(::Any, ::Any, ::Any) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/REPL.jl:822
 [18] #invokelatest#1 at ./essentials.jl:710 [inlined]
 [19] invokelatest at ./essentials.jl:709 [inlined]
 [20] run_interface(::REPL.Terminals.TextTerminal, ::REPL.LineEdit.ModalInterface, ::REPL.LineEdit.MIState) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/LineEdit.jl:2355
 [21] run_frontend(::REPL.LineEditREPL, ::REPL.REPLBackendRef) at /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.5/REPL/src/REPL.jl:1144
 [22] (::REPL.var"#38#42"{REPL.LineEditREPL,REPL.REPLBackendRef})() at ./task.jl:356
```
