### 2021-01-22 #2954 @ranocha
Thanks for creating this nice package, @michakraus. The brief summary looks very interesting to me. If the package satisfies the expectations generated there, it will be a nice contribution to the community. Please find below some first comments after looking at the README and the  manuscript. I'm looking forward to reading the documentation and trying out your package.

- Based on the commit history on GitHub, it looks like Tomasz Tyranowski made significant contributions to the software. However, he is not an author of the submitted manuscript (but mentioned in the acknowledgments).
- [ ] You should consider reformatting the license file `LICENSE.md` such that GitHub recognizes the license type automatically, cf. https://opensource.org/licenses/MIT
- [ ] The example in the `README.md` does not work, cf. https://github.com/JuliaGNI/GeometricIntegrators.jl/issues/106. It would be nice to run this kind of example automatically in CI (maybe using `UnicodePlots`?).
- [ ] It would be nice if the CI badge in the `README.md` linked directly to the CI GitHub action - which fails right now (commit 79282e0cc475433170de66367817aff14b3da432).
- [ ] The reference "Kraus, M. (2020). SPARK methods for degenerate Lagrangian systems." seems to be incomplete.
