### 2021-01-22 459ea39
Rename and reformat LICENSE file.
