### 2021-01-22 8865b13
Add Julia v1.6 to CI test matrix.
