+++ .github/workflows/CI.yml
@@ -12,6 +12,7 @@ jobs:
       matrix:
         version:
           - '1.5'
+          - '1.6'
         os:
           - ubuntu-latest
           - macOS-latest
