+++ src/solutions/solution_hdf5.jl
@@ -309,7 +309,7 @@ end
 
 function copy_timeteps_to_hdf5(sol::Solution, h5::HDF5.File, j1, j2, n1, n2)
     if size(h5["t"],1) < j2
-        HDF5.set_dims!(h5["t"], (j2,))
+        HDF5.set_extent_dims(h5["t"], (j2,))
     end
     h5["t"][j1:j2] = timesteps(sol)[n1:n2]
 end
@@ -317,7 +317,7 @@ end
 
 function copy_solution_to_hdf5(solution::Union{SolutionODE{DT,TT,1},SolutionDAE{DT,TT,1},SolutionSDE{DT,TT,1}}, h5::HDF5.File, j1, j2, n1, n2) where {DT <: Number, TT}
     if size(h5["q"])[end] < j2
-        HDF5.set_dims!(h5["q"], (j2,))
+        HDF5.set_extent_dims(h5["q"], (j2,))
     end
     h5["q"][j1:j2] = solution.q[n1:n2]
 end
@@ -325,7 +325,7 @@ end
 function copy_solution_to_hdf5(solution::Union{SolutionODE{AT,TT,1},SolutionDAE{AT,TT,1},SolutionSDE{AT,TT,1}}, h5::HDF5.File, j1, j2, n1, n2) where {DT, AT <: Array{DT}, TT}
     elaxes = axes(solution.q[begin])
     if size(h5["q"])[end] < j2
-        HDF5.set_dims!(h5["q"], (size(h5["q"])[1:end-1]..., j2))
+        HDF5.set_extent_dims(h5["q"], (size(h5["q"])[1:end-1]..., j2))
     end
     for i in eachindex(j1:j2, n1:n2)
         j = (j1:j2)[i]
@@ -336,7 +336,7 @@ end
 
 function copy_solution_to_hdf5(solution::Union{SolutionODE{DT,TT,2},SolutionDAE{DT,TT,2},SolutionSDE{DT,TT,2}}, h5::HDF5.File, j1, j2, n1, n2) where {DT <: Number, TT}
     if size(h5["q"],1) < j2
-        HDF5.set_dims!(h5["q"], (j2, size(h5["q"],2)))
+        HDF5.set_extent_dims(h5["q"], (j2, size(h5["q"],2)))
     end
     h5["q"][j1:j2, :] = solution.q[n1:n2, :]
 end
@@ -344,7 +344,7 @@ end
 function copy_solution_to_hdf5(solution::Union{SolutionODE{AT,TT,2},SolutionDAE{AT,TT,2},SolutionSDE{AT,TT,2}}, h5::HDF5.File, j1, j2, n1, n2) where {DT, AT <: Array{DT}, TT}
     elaxes = axes(solution.q[begin,begin])
     if size(h5["q"])[end-1] < j2
-        HDF5.set_dims!(h5["q"], (size(h5["q"])[begin:end-2]..., j2, size(h5["q"])[end]))
+        HDF5.set_extent_dims(h5["q"], (size(h5["q"])[begin:end-2]..., j2, size(h5["q"])[end]))
     end
     for k in 1:nsamples(solution.q)
         for i in eachindex(j1:j2, n1:n2)
@@ -358,10 +358,10 @@ end
 
 function copy_solution_to_hdf5(solution::Union{SolutionPODE{DT,TT,1},SolutionPDAE{DT,TT,1},SolutionPSDE{DT,TT,1}}, h5::HDF5.File, j1, j2, n1, n2) where {DT <: Number, TT}
     if size(h5["q"])[end] < j2
-        HDF5.set_dims!(h5["q"], (j2,))
[…]
