### 2021-01-22 5c40663
Fix Julia v1.6 CI entry.
