### 2021-01-22 006c4b1
Update CI badge to link to the CI GitHub action.
