### 2021-01-22 a77a30b
Fix DataSeries getindex() when called with Colon in time axis.
