### 2021-01-22 JuliaGNI/GeometricIntegrators.jl#106 @michakraus
Thanks a lot for catching this! Fixed in v0.7.1.
