### 2021-01-22 #2954 @michakraus
Thanks for the nice words and the great suggestions! Except for the test of the README example I have implemented all of them. The SPARK reference should have hit the arXiv by now, but there were some delays. Should be online within the next one or two weeks, though. I will add the missing details to the reference asap.

Tom Tyranowski has implemented the first version of stochastic integrators. We are in the process of restructuring the library. We have already moved out several parts of the library into standalone packages (e.g., basis functions, quadrature rules, solvers, Runge-Kutta methods). In the next step we want to put several of the specialised integrators into separate packages, most and foremost [StochasticIntegrators.jl](https://github.com/JuliaGNI/StochasticIntegrators.jl), which we plan to publish separately.
