### 2021-01-22 JuliaGNI/GeometricIntegrators.jl#108 @ranocha
Get references for algorithms
Currently, I do not know how to obtain a reference for a given algorithm that I should cite when using it in a research paper. For example, none of
```julia
julia> using GeometricIntegrators

julia> TableauQinZhang
TableauQinZhang (generic function with 1 method)

julia> TableauQinZhang()
TableauDIRK{Float64}(:qinzhang, 2, 2, Runge-Kutta Coefficients qinzhang with 2 stages and order 2  a = [0.25 0.0; 0.5 0.25]  b = [0.5, 0.5]  c = [0.25, 0.75])

help?> TableauQinZhang()
  Tableau of Qin and Zhang's symplectic two-stage, 2nd order method

help?> TableauQinZhang
search: TableauQinZhang

  Tableau of Qin and Zhang's symplectic two-stage, 2nd order method
```
works for this and I can't also find no reference in the source code.
