### 2021-01-22 JuliaGNI/GeometricIntegrators.jl#108 @michakraus
Thanks a lot! Very good point! I will start by adding all the references to the tableau docstrings in [RungeKutta.jl](https://github.com/JuliaGNI/RungeKutta.jl). Then I need to see what would be the best way to make them accessible in GeometricIntegrators.jl without just copying everything. Anyhow, we want to automatise importing methods defined in RungeKutta.jl into GeometricIntegrators.jl, so that we don't have to amend the latter every time we add a new algorithm to the former. This seems to go hand in hand.
