### 2021-01-22 JuliaGNI/GeometricIntegrators.jl#108 @ranocha
Could you just use Reexport.jl for this?
