### 2021-01-22 JuliaGNI/GeometricIntegrators.jl#108 @michakraus
Not right now as there is a conversion from the tableau type in RungeKutta.jl into different tableau types in GeometricIntegrators.jl (explicit, diagonally implicit, fully implicit). That was originally used for dispatch on the different integrators. However, that could and should be implemented via traits, which would simplify the implementation and should allow for a simple reexport.
