### 2021-01-22 JuliaGNI/GeometricIntegrators.jl#109 @ranocha
TestProblems not defined
Running the example from https://juliagni.github.io/GeometricIntegrators.jl/stable/integrators/
```julia
julia> using GeometricIntegrators

julia> ode = TestProblems.HarmonicOscillatorProblem.harmonic_oscillator_ode()
ERROR: UndefVarError: TestProblems not defined
Stacktrace:
 [1] top-level scope at REPL[2]:1
```
