### 2021-01-22 JuliaGNI/GeometricIntegrators.jl#109 @ranocha
Additionally, I get
```
julia> harmonic_oscillator_iode
ERROR: UndefVarError: harmonic_oscillator_iode not defined
```
following the same page of the docs
