### 2021-01-22 #2954 @ranocha
This is good work, @michakraus, and a nice contribution to the open source ecosystem. Here are some additional comments.

- [ ] As stated in the manuscript, the package "is designed to minimize overhead and maximize performance". However, there seem to be some caveats. For example, only LU decomposition seems to be implemented as linear solver, limiting the size of problems that can be solved. Hence, I would like to know the anticipated scope of applications and the size of ODE systems? GeometricIntegrators.jl looks to be good for small ODEs. Can I also apply the integrators to spatial semidiscretizations of PDEs? If so, is it okay for moderate 1D problems or also for large 3D problems on a single workstation? It would be nice to mention this range of applications in the manuscript.
- [ ] It would be nice to have links to all papers mentioned in the documentation at Home/References.
- [ ] When reading the manuscript, it is not really clear what kind of algorithms GeometricIntegrators.jl implements. It might be helpful to give some examples for each of the categories mentioned.
- [ ] It would be interesting to know whether some of the methods implemented in GeometricIntegrators.jl are also available from other sources and mention differences and/or relative advantages/disadvantages. For example, some geometric numerical integrators are available as Fortran codes from the website of Ernst Hairer. A lot of Runge-Kutta methods are also available in Julia via DifferentialEquations.jl. What is the reason for having another implementation in Julia of these schemes? How do the methods of GeometricIntegrators compare to that?
- [ ] The docs "Integrators/Splitting Methods" and "Integrators/Stochastic Integrators" are empty for v0.7.1.
- [ ] The docs for the modules "Basis Functions", "Linear Solvers", "Nonlinear Solvers", and "Quadratures" are nearly empty. It looks like these modules were moved to additional packages recently?
- [ ] It would be nice to improve docstrings since they don't really help to explain how to call some functions, cf. https://github.com/JuliaGNI/GeometricIntegrators.jl/issues/107.
- [ ] It would be nice if some standard `struct`s used many times could be printed nicely in the REPL by overloading the appropriate methods of `Base.show`.
- [ ] Is adaptive time-stepping or some way to estimate an appropriate step size supported?
- [ ] It should be made possible to get some reference for the algorithms implemented in this package to be able to use and cite them for research, cf. https://github.com/JuliaGNI/GeometricIntegrators.jl/issues/108.
- [ ] Another example from the docs does not work, cf. https://github.com/JuliaGNI/GeometricIntegrators.jl/issues/109
- [ ] It would be nice to see a few more examples for different problem types in the docs.
- [ ] It looks like not all problem types and methods are compatible with each other, e.g. `IDAE` and `TableauExplicitEuler`. It might be good to mention something like this in the docs.
- [ ] It is a bit confusing that an argument modified by a function is passed as last argument - the Julia standard convention is to pass modified arguments at first. What was the reason for this decision? It could be helpful to include a warning with an explanation in the docs.
- [ ] `P_{n,i}` is used twice for different things in the second equation of https://juliagni.github.io/GeometricIntegrators.jl/v0.7.1/integrators/vprk/
- [ ] Is there a specific reason for restricting the type of the right-hand sides to `Function`s? For example, I cannot use callable `struct`s, which would be my first choice when dealing with differential equations with parameters (or encapsulating more complex right-hand sides).
  julia> using GeometricIntegrators

  julia> function my_function(t, x, v)
            @. v = 1.0 * x # the 1.0 should actually be a parameter...
        end

  julia> ode = ODE(my_function, [1.0]) # this works

  julia> struct MyCallable
            parameter::Float64
        end

  julia> function (my_callable::MyCallable)(t, x, v)
            @. v = my_callable.parameter * x
        end

  julia> ode = ODE(MyCallable(1.0), [1.0])
  ERROR: MethodError: no method matching ODE(::MyCallable, ::Float64, ::Array{Array{Float64,1},1})
    ODE(::vType, ::tType, ::Array{arrayType,1}; h, parameters, periodicity) where {dType<:Number, tType<:Real, arrayType<:(AbstractArray{dType,N} where N), vType<:Function} at ~/.julia/packages/GeometricIntegrators/IFNyN/src/equations/ode.jl:43
    ODE(::Any, ::Any, ::AbstractArray{DT,N} where N where DT<:Number; kwargs...) at ~/.julia/packages/GeometricIntegrators/IFNyN/src/equations/ode.jl:58
    ODE(::Any, ::AbstractArray{var"#s16",1} where var"#s16"<:(AbstractArray{DT,N} where N) where DT; kwargs...) at ~/.julia/packages/GeometricIntegrators/IFNyN/src/equations/ode.jl:57
    ...
  [1] ODE(::MyCallable, ::Float64, ::Array{Float64,1}; kwargs::Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}}) at ~/.julia/packages/GeometricIntegrators/IFNyN/src/equations/ode.jl:58
  [2] ODE(::MyCallable, ::Float64, ::Array{Float64,1}) at ~/.julia/packages/GeometricIntegrators/IFNyN/src/equations/ode.jl:58
  [3] ODE(::MyCallable, ::Array{Float64,1}; kwargs::Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}}) at ~/.julia/packages/GeometricIntegrators/IFNyN/src/equations/ode.jl:59
  [4] ODE(::MyCallable, ::Array{Float64,1}) at ~/.julia/packages/GeometricIntegrators/IFNyN/src/equations/ode.jl:59
  [5] top-level scope at REPL[6]:1
  ```
  Hence, it looks like closures are the only way to deal with parameters right now.
- [ ] Quoting https://github.com/openjournals/joss-reviews/issues/2954#issuecomment-765312446
    > In the next step we want to put several of the specialised integrators into separate packages, most and foremost StochasticIntegrators.jl, which we plan to publish separately.

  I'm not sure whether it would be better to make the splitting before publishing this work. Otherwise, the stochastic integrators would be in two papers - but that's not my decision, just a comment you might consider.
