### 2021-01-22 #2954 @michakraus
Thanks a lot @ranocha for the thorough review. Working through your comments will improve the package and especially the documentation quite a bit. I will try to fix everything within next week, including the split of stochastic integrators. Let me comment on a few points right away, the rest I will just fix:

> * [ ]  It would be nice if the CI badge in the `README.md` linked directly to the CI GitHub action - which fails right now (commit 79282e0cc475433170de66367817aff14b3da432). It looks like CI is configured to fail if it fails on the nightly builds of Julia. It might be worth considering to change this behavior and show failing CI only if it fails on released stable versions of Julia.

Nightly builds are already allowed to fail without showing that CI failed. The Badge issue was fixed earlier today.

> * [ ] [...] Hence, I would like to know the anticipated scope of applications and the size of ODE systems? GeometricIntegrators.jl looks to be good for small ODEs. Can I also apply the integrators to spatial semidiscretizations of PDEs? If so, is it okay for moderate 1D problems or also for large 3D problems on a single workstation? It would be nice to mention this range of applications in the manuscript.

That's a good point. So far we have mostly targeted small to medium size ODE systems. We also use the library for PIC methods, but there the actual particle time steps decouple, i.e., one is again only solving small ODE systems. More recently, we thought a bit about FEM and DG methods, but we haven't performed any tests in that direction, yet. In principle, one can integrate arbitrary abstract arrays, warranted that a different linear solver is used for larger problems. We have a prototype wrapper around IterativeSolvers.jl, but that never got sufficient testing to make it into the repository.

> * [ ] When reading the manuscript, it is not really clear what kind of algorithms GeometricIntegrators.jl implements. It might be helpful to give some examples for each of the categories mentioned.

> * [ ] It would be interesting to know whether some of the methods implemented in GeometricIntegrators.jl are also available from other sources and mention differences and/or relative advantages/disadvantages. For example, some geometric numerical integrators are available as Fortran codes from the website of Ernst Hairer. A lot of Runge-Kutta methods are also available in Julia via DifferentialEquations.jl. What is the reason for having another implementation in Julia of these schemes? How do the methods of GeometricIntegrators compare to that?

Originally, the main purpose of the library was to implement our own integrators, that are not available elsewhere, and to use it both as a test bed and for providing reference implementation. We also added some standard methods for comparison / benchmarks. Recently, we have started to gather more geometric integrators (especially symplectic) from the literature and add them, in order to make GeometricIntegrators a truly comprehensive library for geometric numerical integration.
Some of the specific Runge-Kutta methods (those with names) are also available elsewhere, but for certain classes (Gauss, Lobatto, Radau) we feature tableaus with an arbitrary number of stages, which you can rarely find elsewhere.

> * [ ] The docs for the modules "Basis Functions", "Linear Solvers", "Nonlinear Solvers", and "Quadratures" are nearly empty. It looks like these modules were moved to additional packages recently?

Yes, all of these sub-modules have been extracted. Seems I forgot to remove the docs...

> * [ ] Is adaptive time-stepping or some way to estimate an appropriate step size supported?

We currently have no special infrastructure for adaptive time-stepping. However, we have used adaptive sub-cycling in special-purpose integrators that are not part of the main library. Here, the standard infrastructure is used to specify "target time steps", i.e., the adaptive integrator computes a solution for every point in the equidistant time series, but in between in adapts as it wishes. Whenever an adaptive time step would step over a "target time step", it is reduced to hit that target. For geometric integrators that often is the only sensible thing to do (and for most practical applications as well).
In all of this, the integrator has to take care of the sub-cycling and the rest of GeometricIntegrators doesn't really care about it. If for some reason you want to output the solution at the intermediate irregular time steps, this is relatively easily possible via the (still mostly undocumented) mid-level interface that is used to call the integrators.
If I am not mistaken, DifferentialEquations interpolates the solutions of the adaptive steps to get a regular output, but that would destroy any structure, so it is important that the integrator actually integrates to each regular time step.

> * [ ]  It is a bit confusing that an argument modified by a function is passed as last argument - the Julia standard convention is to pass modified arguments at first. What was the reason for this decision? It could be helpful to include a warning with an explanation in the docs.

I agree that this can be confusing. This convention was originally motivated by the usual notation of differential equations on jet bundles, where you have base space (time for ODEs) -> configuration space (the solution q) -> jet space (containing the derivatives, here v). We have thought about harmonising this with the Julia standard quite a bit, but we couldn't really come up with a satisfactory solution. We also thought about adopting the DifferentialEquations.jl convention. But in the end we found and still find the geometric convention most intuitive (at least if you live in that geometric integration bubble), although it clashes with the Julia convention.
In any case you are right in that this deserves a more prominent mentioning in the docs!

> * [ ] Is there a specific reason for restricting the type of the right-hand sides to Functions? For example, I cannot use callable structs, which would be my first choice when dealing with differential equations with parameters (or encapsulating more complex right-hand sides).

Again a very good point. I'll look into that. Callable structs, however, seem to work well only for ODEs, where only one function for the vector field is specified. All the other equation types require several functions to be specified, thus callable structs would not be a good option.
I guess the other equations types are yet another part of the library that needs a bit more documentation...

> Hence, it looks like closures are the only way to deal with parameters right now.

In fact, each equation type supports an optional argument `parameters` which is supposed to be a NamedTuple. I guess that is not really (well) documented... it would look like
```julia
function my_function(t, x, v, p)
          @. v = p.a * x
      end

julia> ode = ODE(my_function, [1.0]; parameters=(a=1.0,))
```

Thanks again for your suggestions. Your help is really very much appreciated!
