+++ test/integrators/stochastic_integrators_tests.jl
@@ -1,165 +1,0 @@
-
-using GeometricIntegrators.Config
-using GeometricIntegrators.Integrators
-using GeometricIntegrators.Integrators.Stochastic
-using GeometricIntegrators.Solutions
-using GeometricIntegrators.Tableaus
-using GeometricProblems.KuboOscillator
-using Test
-
-using GeometricProblems.KuboOscillator: Δt, nt
-
-include("utils.jl")
-
-ode    = kubo_oscillator_ode()
-sde1   = kubo_oscillator_sde_1()
-sde2   = kubo_oscillator_sde_2()
-sde3   = kubo_oscillator_sde_3()
-psde1  = kubo_oscillator_psde_1()
-psde2  = kubo_oscillator_psde_2()
-psde3  = kubo_oscillator_psde_3()
-spsde1 = kubo_oscillator_spsde_1()
-spsde2 = kubo_oscillator_spsde_2()
-spsde3 = kubo_oscillator_spsde_3()
-
-
-@testset "$(rpad("Stochastic integrator constructors",80))" begin
-
-    @test typeof(Integrator(sde1, TableauBurrageE1(), Δt)) <: IntegratorSERK
-    @test typeof(Integrator(sde1, TableauStochasticGLRK(1), Δt)) <: IntegratorSIRK
-    @test typeof(Integrator(sde1, TableauRosslerRS1(), Δt)) <: IntegratorWERK
-    @test typeof(Integrator(sde1, TableauSRKw1(), Δt)) <: IntegratorWIRK
-    @test typeof(Integrator(psde1, TableauStochasticStormerVerlet(), Δt)) <: IntegratorSIPRK
-    @test typeof(Integrator(spsde1, TableauModifiedStochasticStormerVerlet(), Δt)) <: IntegratorSISPRK
-
-end
-
-
-@testset "$(rpad("SERK integrators",80))" begin
-
-    int = Integrator(sde1, TableauBurrageE1(), Δt)
-    sol = Solution(sde1, Δt, nt, conv=:strong)
-    integrate!(int, sol)
-    @test rel_energy_err(sol) < 2E-6
-
-    int = Integrator(sde2, TableauBurrageE1(), Δt)
-    sol = Solution(sde2, Δt, nt, conv=:strong)
-    integrate!(int, sol)
-    @test rel_energy_err(sol) < 2E-6
-
-    int = Integrator(sde3, TableauBurrageE1(), Δt)
-    sol = Solution(sde3, Δt, nt, conv=:strong)
-    integrate!(int, sol)
-    @test rel_energy_err(sol) < 2E-6
-
-end
-
-
-@testset "$(rpad("SIRK integrators",80))" begin
-
-    int = Integrator(sde1, TableauStochasticGLRK(1), Δt)
-    sol = Solution(sde1, Δt, nt, conv=:strong)
-    integrate!(int, sol)
-    @test rel_energy_err(sol) < 1E-14
-
-    int = Integrator(sde2, TableauStochasticGLRK(1), Δt)
-    sol = Solution(sde2, Δt, nt, conv=:strong)
-    integrate!(int, sol)
-    @test rel_energy_err(sol) < 1E-14
-
-    int = Integrator(sde3, TableauStochasticGLRK(1), Δt)
-    sol = Solution(sde3, Δt, nt, conv=:strong)
-    integrate!(int, sol)
-    @test rel_energy_err(sol) < 1E-14
-
-    # check deterministic part of SIRK integrator to correspond to FIRK integrator
-    int_sto = Integrator(sde1, TableauStochasticGLRK(1), Δt)
[…]
