+++ test/solutions/solutions_tests.jl
@@ -6,5 +6,3 @@ using SafeTestsets
 @safetestset "Atomic Solutions                                                                " begin include("atomic_solutions_tests.jl") end
 @safetestset "Abstract Solution                                                               " begin include("abstract_solution_tests.jl") end
 @safetestset "Deterministic Solutions                                                         " begin include("deterministic_solutions_tests.jl") end
-# @safetestset "Stochastic Solutions                                                            " begin include("stochastic_solutions_tests.jl") end
-# TODO # reactivate
