+++ src/integrators/stochastic/integrators_werk.jl
@@ -1,265 +1,0 @@
-@doc raw"""
-Holds the tableau of a weak explicit Runge-Kutta method.
-
-    Reference: Andreas Rossler, "Second order Runge-Kutta methods for Stratonovich stochastic differential equations",
-    BIT Numerical Mathematics (2007) 47, equation (5.1).
-
-Order of the tableau is not included, because unlike in the deterministic
-setting, it depends on the properties of the noise (e.g., the dimension of
-the Wiener process and the commutativity properties of the diffusion matrix)
-
-Orders stored in `qdrift`, `qdiff` are irrelevant and set to 0.
-
-`qdrift0`, `qdrift1`, `qdrift2` correspond to A0, A1, A2 in the paper
-`qdiff0`, `qdiff1`, `qdiff2`, `qdiff3` correspond to B0, B1, B2, B3
-`qdrift0.b = alpha`
-`qdiff0.b  = beta1`
-`qdiff3.b  = beta2`
-`qdrift0.c = c0`
-`qdrift1.c = c1`
-`qdrift2.c = c2`
-"""
-struct TableauWERK{T} <: AbstractTableauERK{T}
-    name::Symbol
-    s::Int
-
-    qdrift0::CoefficientsRK{T}
-    qdrift1::CoefficientsRK{T}
-    qdrift2::CoefficientsRK{T}
-
-    qdiff0::CoefficientsRK{T}
-    qdiff1::CoefficientsRK{T}
-    qdiff2::CoefficientsRK{T}
-    qdiff3::CoefficientsRK{T}
-
-    function TableauWERK{T}(name, qdrift0, qdrift1, qdrift2, qdiff0, qdiff1, qdiff2, qdiff3) where {T}
-        @assert qdrift0.s == qdrift1.s == qdrift2.s == qdiff0.s == qdiff1.s == qdiff2.s == qdiff3.s
-        @assert qdrift0.c[1] == qdrift1.c[1] == qdrift2.c[1] == qdiff0.c[1] == qdiff1.c[1] == qdiff2.c[1] == qdiff3.c[1] == 0.
-        @assert istrilstrict(qdrift0.a)
-        @assert istrilstrict(qdrift1.a)
-        @assert istrilstrict(qdrift2.a)
-        @assert istrilstrict(qdiff0.a)
-        @assert istrilstrict(qdiff1.a)
-        @assert istrilstrict(qdiff2.a)
-        @assert istrilstrict(qdiff3.a)
-        @assert !(qdrift0.s==1. && qdrift0.a[1,1] ≠ 0.)
-        @assert !(qdrift1.s==1. && qdrift1.a[1,1] ≠ 0.)
-        @assert !(qdrift2.s==1. && qdrift2.a[1,1] ≠ 0.)
-        @assert !(qdiff0.s==1. && qdiff0.a[1,1] ≠ 0.)
-        @assert !(qdiff1.s==1. && qdiff1.a[1,1] ≠ 0.)
-        @assert !(qdiff2.s==1. && qdiff2.a[1,1] ≠ 0.)
-        @assert !(qdiff3.s==1. && qdiff3.a[1,1] ≠ 0.)
-        new(name, qdrift0.s, qdrift0, qdrift1, qdrift2, qdiff0, qdiff1, qdiff2, qdiff3)
-    end
-end
-
-function TableauWERK(name, qdrift0::CoefficientsRK{T}, qdrift1::CoefficientsRK{T}, qdrift2::CoefficientsRK{T},
[…]

[…]
