### 2021-01-25 961181c
Remove docs from extracted modules.
