+++ docs/src/modules/quadratures.md
@@ -1,7 +1,0 @@
-
-# Quadratures
-
-```@autodocs
-Modules = [GeometricIntegrators.Quadratures]
-Order   = [:constant, :type, :macro, :function]
-```
