+++ docs/src/modules/solvers_linear.md
@@ -1,10 +1,0 @@
-
-# Linear Solvers
-
-```@autodocs
-Modules = [GeometricIntegrators.Solvers]
-Pages   = ["solvers/linear/linear_solvers.jl",
-           "solvers/linear/lu_solver.jl",
-           "solvers/linear/lu_solver_lapack.jl"]
-Order   = [:constant, :type, :macro, :function]
-```
