### 2021-01-25 25e82f0
Clean up some leftovers of StochasticIntegrators.jl.
