+++ Project.toml
@@ -20,7 +20,6 @@ Polynomials = "f27b6e38-b328-58d1-80ce-0feddd5e7a45"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 QuadratureRules = "a08977f5-a20d-4b99-b11f-f5ea535e0575"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SharedArrays = "1a1011a3-84de-559e-8e89-a11a2f7dc383"
 SimpleSolvers = "36b790f5-6434-4fbe-b711-1f64a1e2f6a2"
