### 2021-01-28 #2954 @ranocha
I already posted some review comments above. When these are taken into account by modifying the package/docs/paper and @michakraus notifies me about this, I will finish the review.
