+++ src/integrators/spark/integrators_spark.jl
@@ -3,7 +3,7 @@
 const TableauSPARK = AbstractTableauSPARK{:spark}
 const ParametersSPARK = AbstractParametersSPARK{:spark}
 
-function Integrators.check_symplecticity(tab::TableauSPARK{T}; atol=16*eps(T), rtol=16*eps(T)) where {T}
+function RungeKutta.check_symplecticity(tab::TableauSPARK{T}; atol=16*eps(T), rtol=16*eps(T)) where {T}
     s_αa_qp̃ = [isapprox(tab.q.b[i] * tab.p.α[i,j] + tab.p.β[j] * tab.q̃.a[j,i], tab.q.b[i] * tab.p.β[j]; atol=atol, rtol=rtol) for i in 1:tab.s, j in 1:tab.r]
     s_α_q̃p̃  = [isapprox(tab.p.β[i] * tab.q̃.α[i,j] + tab.q.β[j] * tab.p̃.α[j,i], tab.p.β[i] * tab.q.β[j]; atol=atol, rtol=rtol) for i in 1:tab.r, j in 1:tab.r]
     s_b_qp  = isapprox.(tab.q.b, tab.p.b; atol=atol, rtol=rtol)
