### 2021-01-29 JuliaGNI/GeometricIntegrators.jl#109 @michakraus
I forgot to update the docs when extracting the example code into [GeometricProblems.jl](https://github.com/JuliaGNI/GeometricProblems.jl)
All examples should work now, cf. https://juliagni.github.io/GeometricIntegrators.jl/latest/integrators/
