### 2021-01-29 JuliaGNI/GeometricIntegrators.jl#108 @michakraus
The master branch of GeometricIntegrators.jl now uses [RungeKutta.jl](https://github.com/JuliaGNI/RungeKutta.jl) for the Tableaus. I added references to all tableaus in there. I still need to add references to the splitting methods...
