### 2021-02-01 f3e00d5
Update doctrings of VPRK tableaus and add missing TableauVPLobattoIIIC̄.
