+++ src/tableaus/tableaus_vprk.jl
@@ -1,65 +1,122 @@
 
-"Tableau for variational Gauss-Lobatto IIIA-IIIB method with s stages"
+@doc raw"""
+Tableau for variational Lobatto IIIA method with s stages
+
+Uses Lobatto IIIA for the coefficients $a_{ij}$ and its symplectic conjugate, Lobatto IIIB, for the coefficients $\bar{a}_{ij}$.
+"""
 function TableauVPLobattoIIIA(s)
     TableauVPRK(Symbol("LobattoIIIA$(s)"), 2s-2, CoefficientsLobattoIIIA(s), (-1)^(s+1), get_lobatto_d_vector(s))
 end
 
-"Tableau for variational Gauss-Lobatto IIIA-IIIB method with s stages"
+@doc raw"""
+Tableau for variational Lobatto IIIB method with s stages
+
+Uses Lobatto IIIB for the coefficients $a_{ij}$ and its symplectic conjugate, Lobatto IIIA, for the coefficients $\bar{a}_{ij}$.
+"""
 function TableauVPLobattoIIIB(s)
     TableauVPRK(Symbol("LobattoIIIB$(s)"), 2s-2, CoefficientsLobattoIIIB(s), (-1)^(s+1), get_lobatto_d_vector(s))
 end
 
-"Tableau for variational Gauss-Lobatto IIIC-III method with s stages"
+@doc raw"""
+Tableau for variational Lobatto IIIC method with s stages
+
+Uses Lobatto IIIC for the coefficients $a_{ij}$ and its symplectic conjugate, Lobatto IIIC̄, for the coefficients $\bar{a}_{ij}$.
+"""
 function TableauVPLobattoIIIC(s)
     TableauVPRK(Symbol("LobattoIIIC$(s)"), 2s-2, CoefficientsLobattoIIIC(s), (-1)^(s+1))
 end
 
-"Tableau for variational Gauss-Lobatto IIID method with s stages"
+@doc raw"""
+Tableau for variational Lobatto IIIC̄ method with s stages
+
+Uses Lobatto IIIC̄ for the coefficients $a_{ij}$ and its symplectic conjugate, Lobatto IIIC, for the coefficients $\bar{a}_{ij}$.
+"""
+function TableauVPLobattoIIIC̄(s)
+    TableauVPRK(Symbol("LobattoIIIC̄$(s)"), 2s-2, CoefficientsLobattoIIIC̄(s), (-1)^(s+1))
+end
+
+@doc raw"""
+Tableau for variational Lobatto IIID method with s stages
+
+Uses Lobatto IIID for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.
+"""
 function TableauVPLobattoIIID(s)
     TableauVPRK(Symbol("LobattoIIID$(s)"), 2s-s, CoefficientsLobattoIIID(s), (-1)^s)
 end
 
-"Tableau for variational Gauss-Lobatto IIIE method with s stages"
+@doc raw"""
+Tableau for variational Lobatto IIIE method with s stages
+
+Uses Lobatto IIIE for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.
+"""
 function TableauVPLobattoIIIE(s)
     TableauVPRK(Symbol("LobattoIIIE$(s)"), 2, CoefficientsLobattoIIIE(s), (-1)^s)
 end
 
-"Tableau for variational Gauss-Lobatto IIIF method with s stages"
+@doc raw"""
+Tableau for variational Lobatto IIIF method with s stages
+
[…]
