### 2021-02-01 2f62d65
Update Runge-Kutta tableau list in docs.
