### 2021-02-01 07b17fb
Add AUTHORS and CONTRIBUTING files and update LICENSE accordingly.
