+++ AUTHORS.md
@@ -1,0 +1,19 @@
+# Authors
+
+GeometricIntegrators' development is coordinated by a group of *principal developers*, who are also its main contributors and who can be contacted in case of questions about GeometricIntegrators. In addition, there are *contributors* who have provided substantial additions or modifications. Together, these two groups form "The GeometricIntegrators Authors" as mentioned in the [LICENSE](LICENSE) file.
+
+## Principal Developers
+
+* [Michael Kraus](https://www.michael-kraus.org/),
+  Max Planck Institute for Plasma Physics, Garching, Germany
+* Tomasz M. Tyranoski,
+  Max Planck Institute for Plasma Physics, Garching, Germany
+
+## Contributors
+
+The following people contributed to GeometricIntegrators and are listed in alphabetical order:
+
+* Christopher Albert
+* Michael Kraus
+* Christopher Rackauckas
+* Tomasz M. Tyranoski
