### 2021-02-01 928084e
Refactor splitting tableaus, add docstrings with references, and add new tableau type for simple methods.
