+++ src/tableaus/tableaus_splitting.jl
@@ -1,34 +1,171 @@
 
+@doc raw"""
+Lie-Trotter Splitting A
 
-function TableauLieA(T=Float64)
+For a vector field $\dot{x} = f_1 (t,x) +  f_2 (t,x)$, the splitting reads
+```math
+\Phi_{\Delta t} = \varphi^{2}_{\Delta t} \circ \varphi^{1}_{\Delta t}
+```
+
+Reference:
+
+    H. F. Trotter.
+    On the product of semi-groups of operators.
+    Proceedings of the American Mathematical Society, Volume 10, Pages 545-551, 1959.
+    doi: 10.1090/S0002-9939-1959-0108732-6.
+
+"""
+function TableauLieA(::Type{T}=Float64) where {T}
     a = Array{T}([ 1 ])
     b = Array{T}([ 0 ])
     TableauSplittingNS(:LieTrotterSplittingA, 1, a, b)
 end
 
 
-function TableauLieB(T=Float64)
+@doc raw"""
+Lie-Trotter Splitting B
+
+For a vector field $\dot{x} = f_1 (t,x) +  f_2 (t,x)$, the splitting reads
+```math
+\Phi_{\Delta t} = \varphi^{1}_{\Delta t} \circ \varphi^{2}_{\Delta t}
+```
+
+Reference:
+
+    H. F. Trotter.
+    On the product of semi-groups of operators.
+    Proceedings of the American Mathematical Society, Volume 10, Pages 545-551, 1959.
+    doi: 10.1090/S0002-9939-1959-0108732-6.
+
+"""
+function TableauLieB(::Type{T}=Float64) where {T}
     a = Array{T}([ 0 ])
     b = Array{T}([ 1 ])
     TableauSplittingNS(:LieTrotterSplittingB, 1, a, b)
 end
 
 
-function TableauStrang(T=Float64)
+@doc raw"""
+Strang Splitting
+
+For a vector field $\dot{x} = f_1 (t,x) + f_2 (t,x)$, the splitting reads
+```math
+\Phi_{\Delta t} = \varphi^{1}_{\Delta t / 2} \circ \varphi^{2}_{\Delta t / 2} \circ \varphi^{2}_{\Delta t / 2} \circ \varphi^{1}_{\Delta t / 2}
+```
+
+For vector fields with two components, this is not the most efficient implementation.
+For such cases [`TableauStrangA`](@ref) or [`TableauStrangB`](@ref) should be used instead.
+
+
+References:
+
+    Gilbert Strang.
+    On the construction and comparison of difference schemes.
+    SIAM Journal on Numerical Analysis, Volume 5, Pages 506-517, 1968.
+    doi: 10.1137/0705041.
+
+    Gurij Ivanovich Marchuk.
+    Some applications of splitting-up methods to the solution of mathematical physics problems.
+    Aplikace Matematiky, Volume 13, Pages 103-132, 1968.
+    doi: 10.21136/AM.1968.103142.
+
+"""
+function TableauStrang(::Type{T}=Float64) where {T}
     a = Array{T}([ 1//2 ])
     b = Array{T}([ 1//2 ])
     TableauSplittingNS(:StrangSplitting, 2, a, b)
 end
 
+"Alias for [`TableauStrang`](@ref)"
+TableauMarchuk = TableauStrang
 
-function TableauMcLachlan2(T=Float64; α=0.1932)
+
+@doc raw"""
+Strang Splitting A for a vector field $\dot{x} = f_1 (t,x) + f_2 (t,x)$.
+
+The splitting reads
+```math
+\Phi_{\Delta t} = \varphi^{1}_{\Delta t / 2} \circ \varphi^{2}_{\Delta t} \circ \varphi^{1}_{\Delta t / 2}
+```
+
+References:
+
+    Gilbert Strang.
[…]
