+++ docs/src/modules.md
@@ -2,18 +2,11 @@
 # Modules
 
 ```@contents
-Pages = ["modules/basis_functions.md",
-         "modules/quadratures.md",
-         "modules/discontinuities.md",
+Pages = ["modules/discontinuities.md",
          "modules/equations.md",
          "modules/solutions.md",
          "modules/integrators.md",
-         "modules/integrators_stochastic.md",
-         "modules/integrators_spark.md",
-         "modules/integrators_vprk.md",
          "modules/interpolation.md",
-         "modules/solvers_linear.md",
-         "modules/solvers_nonlinear.md",
          "modules/simulations.md",
          "modules/tableaus.md"]
 Depth = 3
