### 2021-02-02 b819abe
Add missing splitting files to modules/integrators.md page.
