+++ docs/GeometricIntegrators.bib
@@ -51,14 +51,14 @@
 
 
 @article{MarsdenWest:2001,
-	Author = {Marsden, Jerrold E. AND West, Matthew},
-	Doi = {10.1017/S096249290100006X},
-	Journal = {Acta Numerica},
-	Pages = {357--514},
-	Title = {Discrete Mechanics and Variational Integrators},
-	Volume = {10},
-	Year = {2001},
-	Bdsk-Url-1 = {http://dx.doi.org/10.1017/S096249290100006X}}
+    author  = {Marsden, Jerrold E. AND West, Matthew},
+    title   = {Discrete Mechanics and Variational Integrators},
+    journal = {Acta Numerica},
+    year    = {2001},
+    volume  = {10},
+    pages   = {357-514},
+    doi     = {10.1017/S096249290100006X}
+}
 
 @article{Fangzong:2016,
     author  = {Fangzong, Wang and Xiaobing, Liao},
@@ -119,17 +119,6 @@
     volume  = {204},
     pages   = {56-76},
     doi     = {10.1016/j.cam.2006.04.028}
-}
-
-
-@article{MarsdenWest:2001,
-    author  = {Marsden, Jerrold E. AND West, Matthew},
-    title   = {Discrete Mechanics and Variational Integrators},
-    journal = {Acta Numerica},
-    year    = {2001},
-    volume  = {10},
-    pages   = {357-514},
-    doi     = {10.1017/S096249290100006X}
 }
 
 
