### 2021-02-02 e854c91
Fixes in tutorial and manual.
