+++ docs/make.jl
@@ -13,12 +13,13 @@ makedocs(bib,
              "Integrators" => [
                 "Overview"    => "integrators.md",
                 "Runge-Kutta" => "integrators/rk.md",
+                "Splitting"   => "integrators/splitting.md",
                 "VPRK"        => "integrators/vprk.md",
                 "SPARK"       => "integrators/spark.md",
                #  "CGVI"        => "integrators/cgvi.md",
                #  "DGVI"        => "integrators/dgvi.md",
                #  "HPG"         => "integrators/hpg.md",
-                "Splitting"   => "integrators/splitting.md"],
+               ],
              "Modules"  => [
                 "Discontinuities"   => "modules/discontinuities.md",
                 "Equations"         => "modules/equations.md",
@@ -26,7 +27,9 @@ makedocs(bib,
                 "Interpolation"     => "modules/interpolation.md",
                 "Simulations"       => "modules/simulations.md",
                 "Solutions"         => "modules/solutions.md",
-                "Tableaus"          => "modules/tableaus.md"],
+                "Tableaus"          => "modules/tableaus.md",
+                "Runge-Kutta Methods"  => "modules/rungekutta.md",
+                ],
              "Release Notes" => "releasenotes.md",
              "Bibliography" => "bibliography.md",
              ]
