+++ docs/src/modules.md
@@ -8,6 +8,7 @@ Pages = ["modules/discontinuities.md",
          "modules/integrators.md",
          "modules/interpolation.md",
          "modules/simulations.md",
-         "modules/tableaus.md"]
+         "modules/tableaus.md",
+         "modules/rungekutta.md"]
 Depth = 3
 ```
