+++ docs/src/modules/rungekutta.md
@@ -1,0 +1,7 @@
+
+# Runge-Kutta Methods
+
+```@autodocs
+Modules = [RungeKutta]
+Order   = [:constant, :type, :macro, :function]
+```
