### 2021-02-02 30a582b
Add separate pages in docs for different tableau types.
