+++ docs/make.jl
@@ -20,16 +20,21 @@ makedocs(bib,
                #  "DGVI"        => "integrators/dgvi.md",
                #  "HPG"         => "integrators/hpg.md",
                ],
-             "Modules"  => [
+             "Modules" => [
                 "Discontinuities"   => "modules/discontinuities.md",
                 "Equations"         => "modules/equations.md",
                 "Integrators"       => "modules/integrators.md",
                 "Interpolation"     => "modules/interpolation.md",
                 "Simulations"       => "modules/simulations.md",
                 "Solutions"         => "modules/solutions.md",
-                "Tableaus"          => "modules/tableaus.md",
-                "Runge-Kutta Methods"  => "modules/rungekutta.md",
-                ],
+               ],
+             "Tableaus" => [
+                "Runge-Kutta Methods"  => "tableaus/rungekutta.md",
+                "Partitioned Runge-Kutta Methods"  => "tableaus/prk.md",
+                "Splitting Methods"  => "tableaus/splitting.md",
+                "VPRK Methods"  => "tableaus/vprk.md",
+                "SPARK Methods"  => "tableaus/spark.md",
+               ],
              "Release Notes" => "releasenotes.md",
              "Bibliography" => "bibliography.md",
              ]
