+++ docs/src/tableaus/spark.md
@@ -1,0 +1,16 @@
+# SPARK Tableaus
+
+```@autodocs
+Modules = [GeometricIntegrators.Tableaus]
+Pages   = ["tableaus/coefficients_glrk.jl",
+           "tableaus/coefficients_lob.jl",
+           "tableaus/tableaus_hpark.jl",
+           "tableaus/tableaus_hspark_primary.jl",
+           "tableaus/tableaus_hspark_secondary.jl",
+           "tableaus/tableaus_slrk.jl",
+           "tableaus/tableaus_spark.jl",
+           "tableaus/tableaus_vpark.jl",
+           "tableaus/tableaus_vspark_primary.jl",
+           "tableaus/tableaus_vspark_secondary.jl"]
+Order   = [:constant, :type, :macro, :function]
+```
