+++ docs/src/tableaus/splitting.md
@@ -1,0 +1,7 @@
+# Splitting Tableaus
+
+```@autodocs
+Modules = [GeometricIntegrators.Tableaus]
+Pages   = ["tableaus/tableaus_splitting.jl"]
+Order   = [:constant, :type, :macro, :function]
+```
