+++ docs/src/tableaus/prk.md
@@ -1,0 +1,8 @@
+# Partitioned Runge-Kutta Tableaus
+
+```@autodocs
+Modules = [GeometricIntegrators.Tableaus]
+Pages   = ["tableaus/tableaus_eprk.jl",
+           "tableaus/tableaus_iprk.jl"]
+Order   = [:constant, :type, :macro, :function]
+```
