+++ docs/src/modules/rungekutta.md
@@ -1,7 +1,0 @@
-
-# Runge-Kutta Methods
-
-```@autodocs
-Modules = [RungeKutta]
-Order   = [:constant, :type, :macro, :function]
-```
