+++ docs/src/tableaus/vprk.md
@@ -1,0 +1,7 @@
+# Variational Partitioned Runge-Kutta Tableaus
+
+```@autodocs
+Modules = [GeometricIntegrators.Tableaus]
+Pages   = ["tableaus/tableaus_vprk.jl"]
+Order   = [:constant, :type, :macro, :function]
+```
