### 2021-02-02 a3c61df
Add documentation for splitting integrators and update docstrings.
