+++ src/integrators/splitting/integrators_splitting.jl
@@ -1,5 +1,33 @@
+@doc raw"""
+Splitting integrator for the solution of initial value problems
+```math
+\dot{q} (t) = v(t, q(t)) , \qquad q(t_{0}) = q_{0} ,
+```
+whose vector field ``v`` is given as a sum of vector fields
+```math
+v (t) = v_1 (t) + ... + v_r (t) .
+```
 
-"Splitting integrator."
+`IntegratorSplitting` has two constructors:
+```julia
+IntegratorSplitting{DT,D}(solutions::Tuple, f::Vector{Int}, c::Vector, Δt)
+IntegratorSplitting(equation::SODE, tableau::AbstractTableauSplitting, Δt)
+```
+In the first constructor, `DT` is the data type of the state vector and `D`
+the dimension of the system. In the second constructor, this information
+is extracted from the equation. 
+The tuple `solutions` contains functions implementing the flow (exact solution)
+of the vector fields `v_i`. The vectors `f` and `c` define the actual splitting
+method: `f` is a vector of indices of the flows in the split equation to be
+solved and `c` is a vector of the same size `f` that contains the coefficients
+for each splitting step, i.e., the resulting integrator has the form
+```math
+\varphi_{\tau} = \phi_{c[s] \tau}^{v_{f[s]}} \circ \dotsc \circ \phi_{c[2] \tau}^{v_{f[2]}} \circ \phi_{c[1] \tau}^{v_{f[1]}} .
+```
+In the second constructor, these vectors are constructed from the tableau and
+the equation.
+
+"""
 struct IntegratorSplitting{DT, TT, D, S, QT <: Tuple} <: ODEIntegrator{DT,TT}
     q::QT
     f::NTuple{S,Int64}
@@ -14,7 +42,6 @@ struct IntegratorSplitting{DT, TT, D, S, QT <: Tuple} <: ODEIntegrator{DT,TT}
     end
 end
 
-"Construct splitting integrator."
 function IntegratorSplitting(equation::SODE{DT,TT}, tableau::ST, Δt::TT) where {DT, TT, ST <: AbstractTableauSplitting{TT}}
     @assert hassolution(equation)
     IntegratorSplitting{DT, ndims(equation)}(get_solution_tuple(equation), get_splitting_coefficients(length(equation.q), tableau)..., Δt)
