+++ src/integrators/splitting/integrators_composition.jl
@@ -1,5 +1,37 @@
+@doc raw"""
+Composition integrator for the solution of initial value problems
+```math
+\dot{q} (t) = v(t, q(t)) , \qquad q(t_{0}) = q_{0} ,
+```
+whose vector field ``v`` is given as a sum of vector fields
+```math
+v (t) = v_1 (t) + ... + v_r (t) .
+```
 
-"Composition integrator."
+`IntegratorComposition` has three constructors:
+```julia
+IntegratorComposition{DT,D}(integrators::Tuple, Δt)
+IntegratorComposition(equation::SODE, constructors::Tuple, tableau::AbstractTableauSplitting, Δt)
+IntegratorComposition(equation::SODE, tableau::AbstractTableauSplitting, Δt)
+```
+In the first constructor, `DT` is the data type of the state vector and `D`
+the dimension of the system. In the second and third constructor, this
+information is extracted from the equation. 
+The tuple `integrators` contains the integrators for each substep. Each integrator
+is instantiated with appropriately scaled time step size $\Delta t = c_i \tau$ to
+match the corresponding splitting scheme.
+In the second constructor, the tuple `constructors` contains constructors for the
+integrators of each step of the composition. The integrators are constructed
+according to the tableau and time step `\Delta t` and passed to the first
+constructor.
+The third constructor assumes that the exact solution is used for each splitting
+step. It thus constructs a composition method that is equivalent to a plain
+[`IntegratorSplitting`](@ref).
+
+In order to include exact solutions in the composition, the [`IntegratorExactODE`](@ref)
+implements the general integrator interface.
+
+"""
 struct IntegratorComposition{DT, TT, D, S, IT <: Tuple} <: ODEIntegrator{DT,TT}
     ints::IT
     Δt::TT
@@ -10,7 +42,6 @@ struct IntegratorComposition{DT, TT, D, S, IT <: Tuple} <: ODEIntegrator{DT,TT}
     end
 end
 
-"Construct composition integrator."
 function IntegratorComposition(equation::SODE{DT,TT}, integrators::Tuple, tableau::ST, Δt::TT) where {DT, TT, ST <: AbstractTableauSplitting{TT}}
     D = ndims(equation)
     R = length(equation.v)
