+++ docs/GeometricIntegrators.bib
@@ -49,6 +49,45 @@
     doi     = {10.5334/jors.151}
 }
 
+@article{Strang:1968,
+    author  = {Strang, Gilbert},
+    title   = {On the construction and comparison of difference schemes},
+    journal = {SIAM Journal on Numerical Analysis},
+    pages   = {506--517},
+    volume  = {5},
+    year    = {1968},
+    doi     = {10.1137/0705041}
+}
+
+@article{Marchuk:1968,
+    author  = {Marchuk, Gurij Ivanovich},
+    title   = {Some applications of splitting-up methods to the solution of mathematical physics problems},
+    journal = {Aplikace Matematiky},
+    pages   = {103--132},
+    volume  = {13},
+    year    = {1968},
+    doi     = {10.21136/AM.1968.103142}
+}
+
+@article{McLachlan:1995,
+    author        = {McLachlan, Robert I.},
+    title         = {On the Numerical Integration of Ordinary Differential Equations by Symmetric Composition Methods},
+    journal       = {SIAM Journal on Scientific Computing},
+    pages         = {151--168},
+    volume        = {16},
+    year          = {1995},
+    doi           = {10.1137/0916010}
+}
+
+@article{McLachlanQuispel:2002,
+    author    = {McLachlan, Robert I. and Quispel, G. Reinout W.},
+    title     = {Splitting methods},
+    journal   = {Acta Numerica},
+    volume    = {11},
+    year      = {2002},
+    pages     = {341–434},
+    doi       = {10.1017/S0962492902000053}
+}
 
 @article{MarsdenWest:2001,
     author  = {Marsden, Jerrold E. AND West, Matthew},

[…]
