### 2021-02-02 JuliaGNI/GeometricIntegrators.jl#108 @michakraus
All Runge-Kutta and Splitting methods now have references:
https://juliagni.github.io/GeometricIntegrators.jl/latest/tableaus/rungekutta/
https://juliagni.github.io/GeometricIntegrators.jl/latest/tableaus/splitting/

Also the splitting integrators themselves have obtained some documentation:
https://juliagni.github.io/GeometricIntegrators.jl/latest/integrators/splitting/

A few more examples would be nice, but it's a start... I will also add the possibility to integrate partitioned equations (PODEs) with the splitting and composition methods.
