### 2021-02-02 daf5294
Add missing equations to splitting/composition tableaus.
