### 2021-02-03 JuliaGNI/GeometricIntegrators.jl#108 @ranocha
Thanks, this is really helpful!

Although it's not related directly to this issue: On the first page you linked above, there are two `show` methods described. However, I have no clue how to call them since the signature is omitted, see below.
![image](https://user-images.githubusercontent.com/12693098/106703125-e5ac6d00-65e9-11eb-8dd1-33ccf00c5283.png)
