### 2021-02-03 JuliaGNI/GeometricIntegrators.jl#108 @michakraus
Thanks. There are quite a few of those in the library. Docstrings are a general problem, where the library doesn't shine (yet)... working on it!
