### 2021-02-03 d76cbe7
Add codecov to GitHub CI.
