+++ .github/workflows/CI.yml
@@ -53,3 +53,4 @@ jobs:
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
+      - uses: julia-actions/julia-uploadcodecov@latest
