### 2021-02-03 cd1af4c
Clean up docstrings for integrate_step!
