+++ src/integrators/vprk/integrators_vprk_degenerate.jl
@@ -121,7 +121,6 @@ function Integrators.function_stages!(x::Vector{ST}, b::Vector{ST},
 end
 
 
-"Integrate ODE with variational partitioned Runge-Kutta integrator."
 function Integrators.integrate_step!(int::IntegratorVPRKdegenerate{DT,TT}, sol::AtomicSolutionPODE{DT,TT},
                                      cache::IntegratorCacheVPRK{DT}=int.caches[DT]) where {DT,TT}
     # update nonlinear solver parameters from cache
