### 2021-02-03 c2e1879
Clean up partitioned Runge Kutta tableaus.
