+++ src/integrators/rk/coefficients.jl
@@ -1,12 +1,0 @@
-
-@define HeaderCoefficientsRK begin
-    name::Symbol
-    o::Int
-    s::Int
-end
-
-@define CoefficientsRK begin
-    a::Matrix{T}
-    b::Vector{T}
-    c::Vector{T}
-end
