+++ src/Tableaus.jl
@@ -28,15 +28,14 @@ module Tableaus
     include("tableaus/coefficients_lob.jl")
 
 
-    include("tableaus/tableaus_eprk.jl")
+    include("tableaus/tableaus_prk.jl")
 
-    export TableauSymplecticEulerA, TableauSymplecticEulerB,
-           TableauLobattoIIIAIIIB2, TableauLobattoIIIBIIIA2
+    export TableauSymplecticEulerA,
+           TableauSymplecticEulerB,
+           TableauLobattoIIIAIIIB,
+           TableauLobattoIIIBIIIA,
+           PartitionedTableauGauss
 
-    include("tableaus/tableaus_iprk.jl")
-
-    export TableauIPGLRK           
-    
     include("tableaus/tableaus_spark.jl")
 
     export TableauSPARKGLRK,
