+++ test/tableaus/rk_tableaus_tests.jl
@@ -2,16 +2,16 @@
 @testset "$(rpad("Runge-Kutta Tableaus",80))" begin
 
     # instatiate all partitioned tableaus
-    @test typeof(TableauSymplecticEulerA())     <: TableauEPRK
-    @test typeof(TableauSymplecticEulerB())     <: TableauEPRK
-    @test typeof(TableauLobattoIIIAIIIB2())     <: TableauEPRK
-    @test typeof(TableauLobattoIIIBIIIA2())     <: TableauEPRK
+    @test typeof(TableauSymplecticEulerA())     <: TableauPRK
+    @test typeof(TableauSymplecticEulerB())     <: TableauPRK
+    @test typeof(TableauLobattoIIIAIIIB(2))     <: TableauPRK
+    @test typeof(TableauLobattoIIIBIIIA(2))     <: TableauPRK
 
     # test instatiation of partioned tableau by composition of two RK tableaus
-    @test typeof(TableauEPRK(:PERK4, 4, TableauRK4(), TableauRK4())) <: TableauEPRK
-    @test TableauEPRK(:PERK4, 4, TableauRK4(), TableauRK4()) == TableauEPRK(:PERK4, 4, TableauRK4())
+    @test typeof(TableauPRK(:PERK4, TableauRK4(), TableauRK4())) <: TableauPRK
+    @test TableauPRK(:PERK4, TableauRK4(), TableauRK4()) == TableauPRK(:PERK4, TableauRK4())
 
 
-    # TODO Add tests for TableauIPRK, TableauSARK, TableauSPARK and TableauGLM.
+    # TODO Add tests for TableauSARK, TableauSPARK and TableauGLM.
 
 end
