### 2021-02-03 c941ed3
Clean up partitioned Runge Kutta integrators. Add tests for and fix integrators for implicit equations.
