+++ test/integrators/rk_integrators_tests.jl
@@ -75,8 +75,9 @@ end
     @test typeof(Integrator(ode, TableauExplicitMidpoint(), Δt)) <: IntegratorERK
     @test typeof(Integrator(ode, TableauCrouzeix(), Δt)) <: IntegratorDIRK
     @test typeof(Integrator(ode, TableauImplicitMidpoint(), Δt)) <: IntegratorFIRK
-    @test typeof(Integrator(pode, TableauEPRK(:eprk_midpoint, 2, TableauExplicitMidpoint()), Δt)) <: IntegratorEPRK
-    @test typeof(Integrator(pode, TableauIPRK(:iprk_midpoint, 2, TableauImplicitMidpoint()), Δt)) <: IntegratorIPRK
+    @test typeof(Integrator(pode, TableauPRK(:eprk_midpoint, TableauExplicitMidpoint()), Δt)) <: IntegratorEPRK
+    @test typeof(Integrator(pode, TableauPRK(:iprk_midpoint, TableauImplicitMidpoint()), Δt)) <: IntegratorIPRK
+    @test typeof(Integrator(iode, TableauPRK(:iprk_midpoint, TableauImplicitMidpoint()), Δt)) <: IntegratorPRKimplicit
 
     @test_logs (:warn, r"Initializing IntegratorDIRK with explicit tableau heun.*") IntegratorDIRK(ode, TableauHeun2(), Δt)
     @test_logs (:warn, r"Initializing IntegratorFIRK with explicit tableau heun.*") IntegratorFIRK(ode, TableauHeun2(), Δt)
@@ -85,8 +86,8 @@ end
     int_erk  = Integrator(ode, TableauExplicitMidpoint(), Δt)
     int_dirk = Integrator(ode, TableauCrouzeix(), Δt)
     int_firk = Integrator(ode, TableauImplicitMidpoint(), Δt)
-    int_eprk = Integrator(pode, TableauEPRK(:eprk_midpoint, 2, TableauExplicitMidpoint()), Δt)
-    int_iprk = Integrator(pode, TableauIPRK(:iprk_midpoint, 2, TableauImplicitMidpoint()), Δt)
+    int_eprk = Integrator(pode, TableauPRK(:eprk_midpoint, TableauExplicitMidpoint()), Δt)
+    int_iprk = Integrator(pode, TableauPRK(:iprk_midpoint, TableauImplicitMidpoint()), Δt)
 
     @test ndims(ode) == ndims(int_erk)
     @test ndims(ode) == ndims(int_dirk)
@@ -217,19 +218,19 @@ end
     @test rel_err(psol.q, refq) < 5E-2
     @test rel_err(psol.p, refp) < 1E-3
 
-    psol = integrate(pode, TableauEPRK(:prk4, 4, TableauRK4()), Δt, nt)
+    psol = integrate(pode, TableauPRK(:prk4, TableauRK4()), Δt, nt)
     @test rel_err(psol.q, refq) < 5E-7
     @test rel_err(psol.p, refp) < 5E-7
 
-    psol = integrate(pode, TableauIPRK(:pglrk, 2, TableauGauss(1)), Δt, nt)
+    psol = integrate(pode, TableauPRK(:pglrk, TableauGauss(1)), Δt, nt)
     @test rel_err(psol.q, refq) < 5E-4
     @test rel_err(psol.p, refp) < 5E-4
 
-    psol = integrate(pode, TableauIPRK(:pglrk, 4, TableauGauss(2)), Δt, nt)
+    psol = integrate(pode, TableauPRK(:pglrk, TableauGauss(2)), Δt, nt)
     @test rel_err(psol.q, refq) < 1E-7
[…]
