+++ src/integrators/rk/integrators_pglrk.jl
@@ -183,7 +183,7 @@ Projected Gauss-Legendre Runge-Kutta integrator.
 """
 struct IntegratorPGLRK{DT, TT, D, S, PT <: ParametersPGLRK{DT,TT},
                                      ST <: NonlinearSolver{DT},
-                                     IT <: InitialGuessODE{TT}} <: IntegratorRK{DT,TT}
+                                     IT <: InitialGuessODE{TT}} <: AbstractIntegratorRK{DT,TT}
     params::PT
     solver::ST
     iguess::IT
