### 2021-02-03 2797ac7
Rename Lobatto nullvector function.
