+++ src/tableaus/coefficients_lob.jl
@@ -32,7 +32,7 @@ function get_lobatto_glrk_coefficients(s, σ=s+1, T=Float64)
 end
 
 
-function get_lobatto_d_vector(s, T=Float64; normalize=false)
+function get_lobatto_nullvector(s, T=Float64; normalize=false)
     q = get_lobatto_nodes(s)
     l = Lagrange(q)
     v = [l'[x, j] for x in q, j in eachindex(l)]
