### 2021-02-03 1fc21a7
Fix some docstrings.
