### 2021-02-03 fedd669
Remove some obsolete tests.
