### 2021-02-04 89d62e5
Use RungeKutta.jl also for partitioned tableaus and Lobatto nullvector.
