### 2021-02-04 e9c0751
Bump version to v0.8.0.
