### 2021-02-10 5ea8a4f
Update docs for Runge-Kutta tableaus.
