+++ docs/src/tableaus/rungekutta_partitioned.md
@@ -1,8 +1,6 @@
 # Partitioned Runge-Kutta Tableaus
 
 ```@autodocs
-Modules = [GeometricIntegrators.Tableaus]
-Pages   = ["tableaus/tableaus_eprk.jl",
-           "tableaus/tableaus_iprk.jl"]
+Modules = [RungeKutta.PartitionedTableaus]
 Order   = [:constant, :type, :macro, :function]
 ```
