### 2021-02-10 a0779be
Simplify and generalise constructors of VPRK integrators.
