### 2021-02-10 c54f2d6
Simplify and generalise constructors of VPRK integrators.
