+++ docs/src/integrators/vprk.md
@@ -13,14 +13,16 @@ p       &= \dfrac{\partial L}{\partial \dot{q}} (q, \dot{q}) , &
 ```
 by the following scheme,
 ```math
+\begin{equation}\label{eq:vprk}
 \begin{aligned}
-P_{n,i} &= \dfrac{\partial L}{\partial \dot{q}} (Q_{n,i}, V_{n,i}) , &
-F_{n,i} &= \dfrac{\partial L}{\partial q}       (Q_{n,i}, V_{n,i}) , \\
-Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij}       \, V_{n,j} , &
-P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \bar{a}_{ij} \, F_{n,j} , \\
-q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i}        \, V_{n,i} , &
-p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i}  \, F_{n,i} .
+P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , &
+Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , &
+q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\
+F_{k,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , &
+P_{n,j} &= p_{n} + h  \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , &
+p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b_{i} \, F_{n,i} .
 \end{aligned}
+\end{equation}
 ```
 Here, $s$ denotes the number of internal stages, $a_{ij}$ and $\bar{a}_{ij}$ are the coefficients of the Runge-Kutta method and $b_{i}$ and $\bar{b}_{i}$ the corresponding weights.
 If the coefficients satisfy the symplecticity conditions,
@@ -33,32 +35,522 @@ b_{i} \bar{a}_{ij} + \bar{b}_{j} a_{ji} &= b_{i} \bar{b}_{j} &
 ```
 these methods correspond to the position-momentum form of the discrete Lagrangian [[MarsdenWest:2001](@cite)]
 ```math
+\begin{equation}\label{eq:vprk-discrete-lagrangian}
 L_{d} (q_{n}, q_{n+1}) = h \sum \limits_{i=1}^{s} b_{i} \, L \big( Q_{n,i}, V_{n,i} \big) .
+\end{equation}
 ```
 
 While these integrators show favourable properties for systems with regular Lagrangian, they are usually not applicable for degenerate Lagrangian systems, in particular those with Lagrangians of the form $L (q, \dot{q}) = \vartheta(q) \cdot \dot{q} - H(q)$.
-While variational integrators are still applicable in the case of $\vartheta$ being a linear function of $q$, they are often found to be unstable when $\vartheta$ is a nonlinear function of $q$ as is the case with Lotka-Volterra systems, guiding centre dynamics and various nonlinear oscillators.
-To mitigate this problem, projection methods have been developed, which when applied to variational integrators provide long-time stable integrators for general degenerate Lagrangian systems that maintain conservation of energy and momenta [[Kraus:2017](@cite)].
+While variational integrators are still applicable in the case of $\vartheta$ being a linear function of $q$, they are often found to be unstable when $\vartheta$ is a nonlinear function of $q$ as is the case with Lotka-Volterra systems, various nonlinear oscillators, guiding centre dynamics and other reduced charged particle models.
[…]
