+++ docs/GeometricIntegrators.bib
@@ -1,17 +1,28 @@
 
+@book{BlanesCasas:2016,
+    author    = {Blanes, Sergio and Casas, Fernando},
+    publisher = {CRC Press},
+    title     = {A Concise Introduction to Geometric Numerical Integration},
+    year      = {2016},
+    doi       = {10.1201/b21563}
+}
+
+@book{HairerLubichRoche:1989,
+    author    = {Hairer, Ernst and Lubich, Christian and Roche, Michel},
+    title     = {The Numerical Solution of Differential-Algebraic Systems by Runge-Kutta Methods},
+    publisher = {Springer},
+    series    = {Lecture Notes in Mathematics},
+    year      = {1989},
+    volume    = {1409},
+    doi       = {10.1007/BFb0093947}
+}
+
 @book{HairerLubichWanner:2006,
     author    = {Hairer, Ernst and Lubich, Christian and Wanner, Gerhard},
     publisher = {Springer},
     title     = {Geometric Numerical Integration},
     year      = {2006},
     doi       = {10.1007/3-540-30666-8}
-}
-
-@book{SanzSernaCalvo:1994,
-    author    = {Sanz-Serna, Jesús María and Calvo, Manuel P.},
-    publisher = {Chapman and Hall},
-    title     = {Numerical Hamiltonian Problems},
-    year      = {1994}
 }
 
 @book{LeimkuhlerReich:2004,
@@ -21,13 +32,23 @@
     year      = {2004}
 }
 
-@book{BlanesCasas:2016,
-    author    = {Blanes, Sergio and Casas, Fernando},
-    publisher = {CRC Press},
-    title     = {A Concise Introduction to Geometric Numerical Integration},
-    year      = {2016}
+@book{SanzSernaCalvo:1994,
+    author    = {Sanz-Serna, Jesús María and Calvo, Manuel P.},
+    publisher = {Chapman and Hall},
+    title     = {Numerical Hamiltonian Problems},
+    year      = {1994}
 }
 
+
+@article{Ascher:1991,
+    author  = {Ascher, Uri M. and Petzold, Linda R.},
+    title   = {Projected implicit Runge-Kutta methods for differential-algebraic equations},
+    journal = {SIAM Journal on Numerical Analysis},
+    year    = {1991},
+    volume  = {28},
+    pages   = {1097-1120},
+    doi     = {10.1137/0728059}
+}
 
 @article{Bezanson:2017,
     author  = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
@@ -39,64 +60,44 @@
     doi     = {10.1137/141000671}
 }
 
-@article{Rackauckas:2017,
-    author  = {Rackauckas, Christopher and Nie, Qing},
-    title   = {DifferentialEquations.jl - A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
-    journal = {Journal of Open Research Software},
-    year    = {2017},
-    volume  = {5},
-    pages   = {15},
-    doi     = {10.5334/jors.151}
+@incollection{Campos:2014,
+    author    = {Campos, Cédric M.},
+    title     = {High Order Variational Integrators: A Polynomial Approach},
+    booktitle = {Advances in Differential Equations and Applications},
+    publisher = {Springer International Publishing},
+    year      = {2014},
+    pages     = {249-258},
+    doi       = {10.1007/978-3-319-06953-1_24}
 }
 
-@article{Strang:1968,
-    author  = {Strang, Gilbert},
-    title   = {On the construction and comparison of difference schemes},
-    journal = {SIAM Journal on Numerical Analysis},
-    pages   = {506--517},
-    volume  = {5},
-    year    = {1968},
[…]
