### 2021-02-10 d740c0e
Add and revise integrator documentation.
