+++ docs/src/integrators/cgvi.md
@@ -1,2 +1,190 @@
+# [Continuous Galerkin Variational Integrators](@id cgvi)
 
-# Continuous Galerkin Variational Integrators
+In the following, we first present the Galerkin framework for variational integrators [[Leok:2004](@cite), [Leok:2011](@cite), [Campos:2014](@cite), [Campos:2015](@cite), [OberBlobaum:2014](@cite), [OberBloebaum:2016](@cite)] where the space of curves $\mf{Q}$ that connect two points in $\mf{M}$ is approximated by a finite-dimensional subspace (Galerkin integrators of 0th kind).
+We use Lagrange polynomials to approximate the trajectories albeit other choices are possible.
+Alternatively, we can approximate the generalised velocities 
+(Galerkin integrators of 1st kind), which leads us to variational-partitioned Runge-Kutta methods.
+Even though, in some cases correspondences between integrators of 0th kind and integrators of 1st kind might be found, in general they are different.
+
+In our treatment we distinguish between integer timesteps, which are the coordinates $q_{n}$ at time $t_{n}$, and internal stages (nodes), which are the coordinates $Q_{n,i}$ located between two integer timesteps $q_{n}$ and $q_{n+1}$ at consecutive points in time $t_{n}$ and $t_{n+1}$.
+In all of the following we assume that the timestep $h$ is constant.
+
+In the discrete variational principle, we have to consider variations at both, the integer timesteps and the internal stages.
+For Galerkin integrators of 0th kind, the coordinates $Q_{n,i}$ are varied, whereas for the Galerkin integrators of 1st kind, the velocities $\dot{Q}_{n,i}$ are varied. So the independent variables are $(q_{n}, Q_{n,i})$ and $(q_{n}, \dot{Q}_{n,i})$, respectively.
+
+
+## Space of Discrete Trajectories
+
+In order to construct the discrete space of curves from $q_{0}$ to $q_{N}$,
+```math
+\mf{Q}_{d} ( q_{0}, q_{N}, \{ t_{n} \}_{n=0}^{N} ) = \big\{ q_{d} : \{ t_{n} \}_{n=0}^{N} \rightarrow \mf{M} \; \big\vert \; q_{d} (t_{0}) = q_{0}, \, q_{d} (t_{N}) = q_{N} \big\} \subset \mf{Q}_{d} ( \mf{M} ) ,
+```
+we will construct a finite-dimensional subspace of
+```math
+\mf{Q} ( q_{0}, q_{N}, [0, T] ) = \big\{ q : [0, T] \rightarrow \mf{M} \; \big\vert \; q(0) = q_{0}, \, q(T) = q_{N} \big\} \subset \mf{Q} ( \mf{M} ) .
+```
+The subspace $\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] )$ is defined by
+```math
+\begin{equation}\label{eq:galerkin_vi_space_of_piecewise_polynomials}
+\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] ) = \big\{q_{h} : [0, T] \rightarrow \mf{M} \; \big\vert \; q_{h} \vert_{ [ t_{n} , t_{n+1} ] } \in \mathbb{P}_{s} ([t_{n}, t_{n+1}]) , \;  q_{h} \in C^{0} ([0, T]) \big\} ,
+\end{equation}
+```
[…]
