+++ docs/src/integrators/variational.md
@@ -1,0 +1,446 @@
+# Variational Integrators
+
+The basic idea of variational integrators is to construct a discrete counterpart to a particular mechanical system instead of directly discretising its equations of motion.
+This means that the fundamental building blocks of classical mechanics, namely the action functional, the Lagrangian, the variational principle, and the Noether theorem, all have discrete equivalents.
+The application of the discrete variational principle to the discrete action then leads to discrete Euler-Lagrange equations.
+The evolution map that corresponds to the discrete Euler-Lagrange equations is what is called a variational integrator.
+The discrete Noether theorem can be used to relate symmetries of the discretised system to discrete momenta that are in principle exactly preserved by this integrator.
+
+With standard numerical methods, one approximately solves the exact equations of some system. In a sense, the idea of variational integrators is to exactly solve the equations of an approximate system.
+
+The seminal work in the development of a discrete equivalent of classical mechanics was presented by [[Veselov:1988](@cite), [Veselov:1988](@cite)].
+His method, based on a discrete variational principle, leads to symplectic integration schemes that automatically preserve constants of motion [[WendlandtMarsden:1997](@cite), [MarsdenWendlandt:1997](@cite)].
+A comprehensive review of discrete mechanics can be found in [[MarsdenWest:2001](@cite)], including a thorough account on the historical development.
+The discrete version of Hamilton's phasespace action principle is presented in [[Leok:2011](@cite)].
+
+
+## Discretisation of the Action
+
+The derivation of the discrete theory follows along the lines of the derivation of the continuous theory.
+The starting point is the discretisation of the space of paths $\mathcal{Q} ( \mf{M} )$ that connect two points in $\mf{M}$,
+```math
+\mathcal{Q} ( \mf{M} ) = \big\{ q : \mathcal{I} \rightarrow \mf{M} \; \big\vert \; \mathcal{I} \subset \mathbb{R} \; \text{smooth and bounded} \big\} .
+```
+Therefore we divide each time interval $\mathcal{I}$ into an equidistant, monotonic sequence $\{ t_{n} \}_{n=0}^{N}$ and defined the discrete path space as
+```math
+\mathcal{Q}_{d} ( \mf{M} ) = \big\{ q_{d} : \{ t_{n} \}_{n=0}^{N} \rightarrow \mf{M} \big\} .
+```
+The space $\mathcal{Q}_{d} ( \mf{M} )$ contains all possible discrete trajectories $q_{d}$ in $\mf{M}$ and is isomorphic to $\mf{M} \times ... \times \mf{M}$ ($N+1$ copies),
+```math
+\mathcal{Q}_{d} ( \mf{M} ) \cong \bigtimes_{N+1} \mf{M} .
+```
[…]
