### 2021-02-10 4b08a93
Add some more background material on projected VPRK methods.
