+++ docs/src/integrators/vprk.md
@@ -194,6 +194,95 @@ Another approach that always works is to use directly compute the position-momen
 Such subtleties, which are easily overlooked, can be avoided by starting the discretisation of the action from a more fundamental point of view, namely by approximating the function spaces of the trajectories, which leads us to [Galerkin Variational Integrators](@ref cgvi).
 
 
+## [Degenerate Lagrangians](@id sec:degenerate-lagrangian-systems)
+
+Degenerate Lagrangian systems are relevant for the study of population models, point vortex dynamics or reduced charged particle models like the guiding centre system.
+Here, we consider degenerate Lagrangian systems characterized by a Lagrangian that is linear or singular in the velocities. In particular, we consider the class of systems whose Lagrangian is of the form
+```math
+\begin{equation}\label{eq:degenerate-lagrangian}
+L (q, v) = \vartheta(q) \cdot v - H(q) .
+\end{equation}
+```
+The Lagrangian $L$ is a function on the tangent bundle $\tb{\mf{M}}$,
+```math
+\begin{equation}\label{eq:lagrangian-tangent-bundle}
+L : \tb{\mf{M}} \rightarrow \mathbb{R} ,
+\end{equation}
+```
+where $\mf{M}$ denotes the configuration manifold of the system which is assumed to be of dimension $d$. The cotangent bundle of the configuration manifold $\mf{M}$ is denoted by $\cb{\mf{M}}$.
+Further, we denote the coordinates of a point $m \in \mf{M}$ by $q(m) = (q^{1} (m), \dotsc, q^{d} (m))$ and similarly coordinates of points in $\tb{\mf{M}}$ by $(q^{i}, v^{i})$ and coordinates of points in $\cb{\mf{M}}$ by $(q^{i}, p^{i})$.
+In the following, we will always assume the existence of a global coordinate chart, so that $\mf{M}$ can be identified with the Euclidean space $\mathbb{R}^{d}$. For simplicity, we often use short-hand notation where we write $q$ to refer to both a point in $\mf{M}$ as well as its coordinates. Similarly, we often denote points in the tangent bundle $\tb{\mf{M}}$ by $(q, v)$.
+In local coordinates, the Lagrangian \eqref{eq:lagrangian-tangent-bundle} is thus written as a map $(q, v) \mapsto L(q, v)$.
+
+In Equation \eqref{eq:degenerate-lagrangian}, $\vartheta = \vartheta_{i} (q) \, \ext q^{i}$ is a differential one-form $\vartheta : \mf{M} \rightarrow \cb{\mf{M}}$, whose components $\vartheta_{i} : \mf{M} \rightarrow \mathbb{R}$ are general, possibly nonlinear functions of $q$, some of which (but not all) could be identically zero.
[…]
