+++ docs/GeometricIntegrators.bib
@@ -1,4 +1,26 @@
 
+@book{AbrahamMarsden:1978,
+    author    = {Abraham, Ralph and Marsden, Jerrold E.},
+    publisher = {American Mathematical Society},
+    title     = {Foundations of Mechanics},
+    url       = {http://authors.library.caltech.edu/25029/},
+    year      = {1978}
+}
+
+@book{AbrahamMarsdenRatiu:1988,
+    author    = {Abraham, Ralph and Marsden, Jerrold E. and Ratiu, Tudor S.},
+    publisher = {Springer},
+    title     = {Manifolds, Tensor Analysis and Application},
+    year      = {1988}
+}
+
+@book{BaezMuniain:1994,
+    author    = {Baez, John and Muniain, Javier P.},
+    publisher = {World Scientific},
+    title     = {Gauge Fields, Knots and Gravity},
+    year      = {1994}
+}
+
 @book{BlanesCasas:2016,
     author    = {Blanes, Sergio and Casas, Fernando},
     publisher = {CRC Press},
@@ -7,6 +29,27 @@
     doi       = {10.1201/b21563}
 }
 
+@book{Darling:1994,
+    author    = {Darling, Richard W. R.},
+    title     = {Differential Forms and Connections},
+    publisher = {Cambridge University Press},
+    year      = {1994}
+}
+
+@book{Dray:2014,
+    author    = {Dray, Tevian},
+    title     = {Differential Forms and the Geometry of General Relativity},
+    publisher = {CRC Press},
+    year      = {2014}
+}
+
+@book{Frankel:2011,
+    author    = {Frankel, Theodore},
+    publisher = {Cambridge University Press},
+    title     = {The Geometry of Physics},
+    year      = {2011}
+}
+	
 @book{HairerLubichRoche:1989,
     author    = {Hairer, Ernst and Lubich, Christian and Roche, Michel},
     title     = {The Numerical Solution of Differential-Algebraic Systems by Runge-Kutta Methods},
@@ -25,11 +68,55 @@
     doi       = {10.1007/3-540-30666-8}
 }
 
+@book{Holm:2009,
+    author    = {Holm, Darryl D. and Schmah, Tanya and Stoica, Cristina},
+    publisher = {Oxford University Press},
+    title     = {Mechanics and Symmetry},
+    year      = {2009}
+}
+
+@book{Lee:2009,
+    author    = {Lee, Jeffrey M.},
+    publisher = {American Mathematical Society},
+    title     = {Manifolds and Differential Geometry},
+    year      = {2009}
+}
+
+@book{Lee:2012,
+    author    = {Lee, John M.},
+    publisher = {Springer},
+    title     = {Introduction to Smooth Manifolds},
+    year      = {2012}
+}
+
 @book{LeimkuhlerReich:2004,
     author    = {Leimkuhler, Benedict and Reich, Sebastian},
     publisher = {Cambridge University Press},
     title     = {Simulating Hamiltonian Dynamics},
     year      = {2004}
+}
+
+@book{MarsdenRatiu:2002,
+    author    = {Marsden, Jerrold E. and Ratiu, Tudor S.},
+    publisher = {Springer},
+    title     = {Introduction to Mechanics and Symmetry},
+    year      = {2002},
+    isbn      = {978-0-387-21792-5}
+}
+
+@book{Morita:2001,
+    author    = {Morita, Shigeyuki},
+    title     = {Geometry of Differential Forms},
+    publisher = {American Mathematical Society},
+    year      = {2001}
+}
+
+@book{SaletanCromer:1971,
+    author    = {Saletan, Eugene J. and Cromer, Alan H.},
[…]
