### 2021-02-10 eaf6792
Clean up VPRK theory.
