+++ docs/src/integrators/vprk.md
@@ -119,8 +119,18 @@ The terms of the variations which are multiplying $\delta q_{n}$ become
 ```math
 \lambda_{n+1} = \lambda_{n} + h \sum \limits_{i=1}^{s} b_{i} \, F_{n,i} .
 ```
-Similar to the position-momentum form, we can define
+Similar to classical [variational integrators](@ref variational-integrators), we can use the discrete fibre derivative
 ```math
+\begin{equation}\label{eq:discrete-fibre-derivative}
+\begin{aligned}
+\mathbb{F}^{-} L_{d} : (q_{n}, q_{n+1}) &\mapsto (q_{n},   p_{n})   = \big( q_{n}  , - D_{1} L_{d} (q_{n}, q_{n+1}) \big) , \\
+\mathbb{F}^{+} L_{d} : (q_{n}, q_{n+1}) &\mapsto (q_{n+1}, p_{n+1}) = \big( q_{n+1},   D_{2} L_{d} (q_{n}, q_{n+1}) \big) .
+\end{aligned}
+\end{equation}
+```
+to define the position-momentum form of the integrator,
+```math
+\begin{equation}\label{eq:vi-position-momentum-form}
 \begin{aligned}
 p_{n  }
 &=           -  D_{1} L_{d} (q_{n}, q_{n+1})
@@ -129,6 +139,7 @@ p_{n+1}
 &= \hphantom{-} D_{2} L_{d} (q_{n}, q_{n+1})
  = \lambda_{n+1} .
 \end{aligned}
+\end{equation}
 ```
 Replacing $\lambda_{n+1}$ in the second equation with its expression obtained from the first equation, we get
 ```math
@@ -190,7 +201,7 @@ accounting for the linear dependence of the $\dot{Q}_{n,i}$ and consequently als
 The particular values of $d_{i}$ depend on the number of stages $s$ and the definition of the $q_{n,i}$ [[OberBloebaum:2016](@cite)]. For two stages, we have $d_{1} = - d_{2}$, so that we can choose, for example, $d = (+1, -1)$, and \eqref{eq:vprk-lobatto} becomes equivalent to the variational integrator of the trapezoidal Lagrangian. For three stages, we can choose $d = (\tfrac{1}{2}, -1, \tfrac{1}{2})$, and for four stages we can use $d = (+1, -\sqrt{5}, +\sqrt{5}, -1)$.
 In GeometricIntegrators, these vectors can be obtained via the function [`get_lobatto_nullvector`](@ref) from [RungeKutta.jl](https://github.com/JuliaGNI/RungeKutta.jl).
 
-Another approach that always works is to use directly compute the position-momentum form of the variational integrator from the discrete Lagrangian \eqref{eq:vprk-discrete-lagrangian} instead of applying the discrete action principle.
+Another approach that always works is to use directly compute the position-momentum form \eqref{eq:vi-position-momentum-form} of the variational integrator for the discrete Lagrangian \eqref{eq:vprk-discrete-lagrangian} instead of applying the discrete action principle.
 Such subtleties, which are easily overlooked, can be avoided by starting the discretisation of the action from a more fundamental point of view, namely by approximating the function spaces of the trajectories, which leads us to [Galerkin Variational Integrators](@ref cgvi).
 
 
@@ -306,23 +317,23 @@ The general procedure is as follows.
 We start with initial conditions $q_{n}$ on $\Delta$ (recall that for the particular Lagrangian \eqref{eq:degenerate-lagrangian} considered here, the configuration manifold $\mf{M}$ and the constraint submanifold $\Delta$ are isomorphic, so that we can use the same coordinates on $\Delta$ as we use on $\mf{M}$).
[…]
