### 2021-02-10 c4a0af0
Add GeometricProblems to docs Project file.
