+++ docs/src/integrators/cgvi.md
@@ -25,19 +25,19 @@ we will construct a finite-dimensional subspace of
 ```
 The subspace $\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] )$ is defined by
 ```math
-\begin{equation}\label{eq:galerkin_vi_space_of_piecewise_polynomials}
+\begin{equation}\label{eq:galerkin-vi-space-of-piecewise-polynomials}
 \mf{Q}_{h} ( q_{0}, q_{N}, [0, T] ) = \big\{q_{h} : [0, T] \rightarrow \mf{M} \; \big\vert \; q_{h} \vert_{ [ t_{n} , t_{n+1} ] } \in \mathbb{P}_{s} ([t_{n}, t_{n+1}]) , \;  q_{h} \in C^{0} ([0, T]) \big\} ,
 \end{equation}
 ```
 where $\mathbb{P}_{s} ([t_{n}, t_{n+1}])$ is the space of polynomials of degree $s$ in the interval $[t_{n}, t_{n+1}] \subset [0, T]$.
-We see that $\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] ) \subset \mf{Q} ( q_{0}, q_{N}, [0, T] )$. In order for $\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] )$ to be an instance of $\mf{Q}_{d} ( q_{0}, q_{N}, \{ t_{n} \}_{n=0}^{N} )$ we have to require in addition to the definition~\eqref{eq:galerkin_vi_space_of_piecewise_polynomials} that  on the sequence $\{ t_{n} \}_{n=0}^{N}$, the curves $q_{h}$ in $\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] )$ satisfy
+We see that $\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] ) \subset \mf{Q} ( q_{0}, q_{N}, [0, T] )$. In order for $\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] )$ to be an instance of $\mf{Q}_{d} ( q_{0}, q_{N}, \{ t_{n} \}_{n=0}^{N} )$ we have to require in addition to the definition \eqref{eq:galerkin-vi-space-of-piecewise-polynomials} that  on the sequence $\{ t_{n} \}_{n=0}^{N}$, the curves $q_{h}$ in $\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] )$ satisfy
 ```math
 q_{h} \vert_{ [ t_{n} , t_{n+1} ] } (t_{n})   = q_{n}
 \qquad \text{and} \qquad
 q_{h} \vert_{ [ t_{n} , t_{n+1} ] } (t_{n+1}) = q_{n+1} ,
 ```
 where $q_{n}$ are the points of the discrete trajectories $q_{d} = \{ q_{n} \}_{n=0}^{N}$.
-However, it is often more convenient to enforce these continuity constraints weakly in the action rather than building them into the spaces, which implies dropping the condition $q_{h} \in C^{0} ([0, T])$ in~\eqref{eq:galerkin_vi_space_of_piecewise_polynomials}. This in turn means that the whole of $\mf{Q}_{h} ( q_{0}, q_{N}, [0, T] )$ is not a subspace of $\mf{Q} ( q_{0}, q_{N}, [0, T] )$ anymore.
[…]
