### 2021-02-11 9eb3e23
Update docstrings for atomic solutions.
