### 2021-02-11 657569f
Add docstring for get_internal_variables.
