### 2021-02-11 d0ca989
Update docstrings for solutions.
