+++ src/solutions/solution_ode.jl
@@ -16,6 +16,33 @@ Contains all fields necessary to store the solution of an ODE.
 * `counter`: counter for copied solution entries
 * `woffset`: counter for file offset
 * `h5`: HDF5 file for storage
+
+### Constructors
+
+```julia
+SSolutionODE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
+SSolutionODE(t::TimeSeries, q::SDataSeries, ntimesteps)
+SSolutionODE(file::String)
+PSolutionODE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
+PSolutionODE(t::TimeSeries, q::PDataSeries, ntimesteps)
+PSolutionODE(file::String)
+```
+
+The constructors `SSolutionODE` create a `SolutionODE` with internal data structures
+for serial simulations (i.e., standard arrays), while the constructors `PSolutionODE`
+create a `SolutionODE` with internal data structures for parallel simulations (i.e.,
+shared arrays).
+
+The usual way to initialise a `Solution` is by passing an equation, which for
+`SolutionODE` has to be an [`ODE`](@ref) or [`SODE`](@ref), a time step `Δt`
+and the number of time steps `ntimesteps`. The optional parameters `nsave` and
+`nwrite` determine the intervals for storing the solution and writing to file,
+i.e., if `nsave > 1` only every `nsave`'th solution is actually stored, and
+every `nwrite`'th time step the solution is stored to disk.
+
+The other constructors, either passing a `TimeSeries` and a `DataSeries` or a
+filename are used to read data from previous simulations.
+
 """
 abstract type SolutionODE{dType, tType, N} <: DeterministicSolution{dType, tType, N} end
 
