### 2021-02-11 bdafc4b
Add image generation to Documenter GitHub action.
