+++ .github/workflows/Documenter.yml
@@ -16,7 +16,9 @@ jobs:
         with:
           version: 1.5
       - name: Install dependencies
-        run: julia --project=docs -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate(); Pkg.build(); Pkg.precompile();'
+        run: |
+          julia --project=docs -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate(); Pkg.build(); Pkg.precompile();'
+          sudo apt-get install texlive-xetex
       - name: Build Tutorial
         run: julia --project tutorial.jl
         working-directory: docs
