### 2021-02-11 baf44e4
Extend docstrings of equations.
