### 2021-02-11 JuliaGNI/GeometricIntegrators.jl#107 @michakraus
I updated the docstrings for
```
Equations
Solutions
AtomicSolutions
Integrator
integrate
integrate!
```
to make them somewhat more useful.
