+++ dev/assets/documenter.js
@@ -4,10 +4,8 @@ requirejs.config({
     'highlight-julia': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.15.10/languages/julia.min',
     'headroom': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.10.3/headroom.min',
     'jqueryui': 'https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min',
-    'katex-auto-render': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.11.1/contrib/auto-render.min',
     'jquery': 'https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.1/jquery.min',
     'headroom-jquery': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.10.3/jQuery.headroom.min',
-    'katex': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.11.1/katex.min',
     'highlight': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.15.10/highlight.min',
     'highlight-julia-repl': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.15.10/languages/julia-repl.min',
   },
@@ -15,11 +13,6 @@ requirejs.config({
   "highlight-julia": {
     "deps": [
       "highlight"
-    ]
-  },
-  "katex-auto-render": {
-    "deps": [
-      "katex"
     ]
   },
   "headroom-jquery": {
@@ -36,32 +29,60 @@ requirejs.config({
 }
 });
 ////////////////////////////////////////////////////////////////////////////////
-require(['jquery', 'katex', 'katex-auto-render'], function($, katex, renderMathInElement) {
-$(document).ready(function() {
-  renderMathInElement(
-    document.body,
-    {
-  "delimiters": [
-    {
-      "left": "$",
-      "right": "$",
-      "display": false
+require([], function() {
+window.MathJax = {
+  "tex": {
+    "inlineMath": [
+      [
+        "$",
+        "$"
+      ],
+      [
+        "\\(",
+        "\\)"
+      ]
+    ],
+    "macros": {
+      "mf": [
+        "\\mathcal{#1}",
+        1
+      ],
+      "identity": "\\mathbb{I}",
+      "id": "\\mathop{id}",
+      "spn": "\\mathop{span}",
+      "bigtimes": "\\mathop{\\vcenter{\\huge\\times}}",
+      "cb": [
+        "\\mathsf{T}_{#1}^{*} #2",
+        2,
+        ""
+      ],
+      "tb": [
+        "\\mathsf{T}_{#1} #2",
+        2,
+        ""
+      ],
+      "rank": "\\mathop{rank}",
+      "ext": "\\mathsf{d}",
+      "abs": [
+        "\\left \\vert #1 \\right \\vert",
+        1
+      ]
     },
-    {
-      "left": "$$",
-      "right": "$$",
-      "display": true
-    },
-    {
-      "left": "\\[",
-      "right": "\\]",
-      "display": true
-    }
-  ]
+    "tags": "ams"
+  },
+  "options": {
+    "processHtmlClass": "tex2jax_process",
+    "ignoreHtmlClass": "tex2jax_ignore"
+  }
 }
+;
 
-  );
-})
+(function () {
+    var script = document.createElement('script');
+    script.src = 'https://cdnjs.cloudflare.com/ajax/libs/mathjax/3.0.5/es5/tex-svg.js';
+    script.async = true;
+    document.head.appendChild(script);
+})();
 
 })
 ////////////////////////////////////////////////////////////////////////////////
