### 2021-02-11 6da7568
Install ImageMagick in Documenter GitHub action.
