### 2021-02-11 d11bd85
Update image generation in Documenter GitHub action.
