### 2021-02-11 JuliaGNI/GeometricIntegrators.jl#107 @michakraus
I also added constructor descriptions to the docstrings of the tableaus in RungeKutta.jl.
