+++ src/equations/lode.jl
@@ -1,5 +1,5 @@
 @doc raw"""
-`VODE`: Variational Ordinary Differential Equation *EXPERIMENTAL*
+`LODE`: Variational Ordinary Differential Equation *EXPERIMENTAL*
 
 Defines an implicit initial value problem
 ```math
@@ -86,14 +86,14 @@ and
 ### Constructors
 
 ```julia
-VODE(ϑ, f, g, t₀, q₀, p₀, λ₀; v̄=(t,q,v)->nothing, f̄=f, h=nothing, Ω=nothing, ∇H=nothing, parameters=nothing, periodicity=zero(q₀[begin]))
-VODE(ϑ, f, g, q₀::StateVector, p₀::StateVector, λ₀::StateVector=zero(q₀); kwargs...) = VODE(ϑ, f, g, 0.0, q₀, p₀, λ₀; kwargs...)
-VODE(ϑ, f, g, t₀, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...) = VODE(ϑ, f, g, t₀, [q₀], [p₀], [λ₀]; kwargs...)
-VODE(ϑ, f, g, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...) = VODE(ϑ, f, g, 0.0, q₀, p₀, λ₀; kwargs...)
+LODE(ϑ, f, g, t₀, q₀, p₀, λ₀; v̄=(t,q,v)->nothing, f̄=f, h=nothing, Ω=nothing, ∇H=nothing, parameters=nothing, periodicity=zero(q₀[begin]))
+LODE(ϑ, f, g, q₀::StateVector, p₀::StateVector, λ₀::StateVector=zero(q₀); kwargs...) = LODE(ϑ, f, g, 0.0, q₀, p₀, λ₀; kwargs...)
+LODE(ϑ, f, g, t₀, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...) = LODE(ϑ, f, g, t₀, [q₀], [p₀], [λ₀]; kwargs...)
+LODE(ϑ, f, g, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...) = LODE(ϑ, f, g, 0.0, q₀, p₀, λ₀; kwargs...)
 ```
 
 """
-struct VODE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
+struct LODE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
             ϑType <: Function, fType <: Function, gType <: Function,
             v̄Type <: Function, f̄Type <: Function, hType <: OptionalFunction,
             ΩType <: OptionalFunction, ∇HType <: OptionalFunction,
@@ -116,7 +116,7 @@ struct VODE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
     parameters::pType
     periodicity::arrayType
 
-    function VODE(ϑ::ϑType, f::fType, g::gType, t₀::tType,
+    function LODE(ϑ::ϑType, f::fType, g::gType, t₀::tType,
                 q₀::Vector{arrayType}, p₀::Vector{arrayType}, λ₀::Vector{arrayType};
                 v̄::v̄Type=(t,q,v)->nothing, f̄::f̄Type=f, h::hType=nothing, Ω::ΩType=nothing, ∇H::∇HType=nothing,
                 parameters::pType=nothing, periodicity=zero(q₀[begin])) where {
[…]
