+++ test/integrators/spark_integrators_tests.jl
@@ -23,8 +23,8 @@ const parameters = (a₁=1.0, a₂=1.0, b₁=-1.0, b₂=-2.0)
 ode  = lotka_volterra_2d_ode(q₀; params=parameters)
 idae = lotka_volterra_2d_idae(q₀; params=parameters)
 pdae = lotka_volterra_2d_pdae(q₀; params=parameters)
-vdae = lotka_volterra_2d_vdae(q₀; params=parameters)
-vdae_slrk = lotka_volterra_2d_slrk(q₀; params=parameters)
+ldae = lotka_volterra_2d_ldae(q₀; params=parameters)
+ldae_slrk = lotka_volterra_2d_slrk(q₀; params=parameters)
 
 int  = IntegratorFIRK(ode, TableauGauss(8), Δt)
 sol  = integrate(ode, int, nt)
@@ -34,81 +34,81 @@ refx = sol.q[end]
 
 @testset "$(rpad("SLRK integrators",80))" begin
 
-    int = IntegratorSLRK(vdae_slrk, TableauSLRKLobattoIIIAB(2), Δt)
-    sol = integrate(vdae_slrk, int, nt)
+    int = IntegratorSLRK(ldae_slrk, TableauSLRKLobattoIIIAB(2), Δt)
+    sol = integrate(ldae_slrk, int, nt)
     @test rel_err(sol.q, refx) < 4E-6
 
-    int = IntegratorSLRK(vdae_slrk, TableauSLRKLobattoIIIAB(3), Δt)
-    sol = integrate(vdae_slrk, int, nt)
+    int = IntegratorSLRK(ldae_slrk, TableauSLRKLobattoIIIAB(3), Δt)
+    sol = integrate(ldae_slrk, int, nt)
     @test rel_err(sol.q, refx) < 2E-11
 
-    int = IntegratorSLRK(vdae_slrk, TableauSLRKLobattoIIIAB(4), Δt)
-    sol = integrate(vdae_slrk, int, nt)
+    int = IntegratorSLRK(ldae_slrk, TableauSLRKLobattoIIIAB(4), Δt)
+    sol = integrate(ldae_slrk, int, nt)
     @test rel_err(sol.q, refx) < 1E-15
 
 
-    int = IntegratorSLRK(vdae_slrk, TableauSLRKLobattoIIIBA(2), Δt)
-    sol = integrate(vdae_slrk, int, nt)
+    int = IntegratorSLRK(ldae_slrk, TableauSLRKLobattoIIIBA(2), Δt)
+    sol = integrate(ldae_slrk, int, nt)
     @test rel_err(sol.q, refx) < 4E-6
 
-    int = IntegratorSLRK(vdae_slrk, TableauSLRKLobattoIIIBA(3), Δt)
-    sol = integrate(vdae_slrk, int, nt)
+    int = IntegratorSLRK(ldae_slrk, TableauSLRKLobattoIIIBA(3), Δt)
+    sol = integrate(ldae_slrk, int, nt)
     @test rel_err(sol.q, refx) < 2E-11
 
-    int = IntegratorSLRK(vdae_slrk, TableauSLRKLobattoIIIBA(4), Δt)
-    sol = integrate(vdae_slrk, int, nt)
+    int = IntegratorSLRK(ldae_slrk, TableauSLRKLobattoIIIBA(4), Δt)
+    sol = integrate(ldae_slrk, int, nt)
     @test rel_err(sol.q, refx) < 1E-15
 
 
-    int = IntegratorSLRK(vdae_slrk, TableauSLRKLobattoIIICC̄(2), Δt)
-    sol = integrate(vdae_slrk, int, nt)
+    int = IntegratorSLRK(ldae_slrk, TableauSLRKLobattoIIICC̄(2), Δt)
[…]
