### 2021-02-11 8305dad
Update and clean up SPDAE.
