+++ src/equations/spdae.jl
@@ -18,11 +18,22 @@ taking values in ``\mathbb{R}^{d} \times \mathbb{R}^{d}`` and
 the algebraic variables ``(\lambda, \gamma)`` taking values in
 ``\mathbb{R}^{n} \times \mathbb{R}^{d}``.
 
+### Parameters
+
+* `DT <: Number`: data type
+* `TT <: Real`: time step type
+* `AT <: AbstractArray{DT}`: array type
+* `vType <: Function`: type of `v`
+* `fType <: Function`: type of `f`
+* `ϕType <: Function`: type of `ϕ`
+* `ψType <: Function`: type of `ψ`
+* `hType <: OptionalFunction`: type of `h`
+* `pType <: Union{NamedTuple,Nothing}`: parameters type
+
 ### Fields
 
 * `d`: dimension of dynamical variables ``q`` and ``p`` as well as the vector fields ``v`` and ``f``
 * `m`: dimension of algebraic variables ``\lambda`` and ``\gamma`` and the constraints ``\phi`` and ``\psi``
-* `n`: number of initial conditions
 * `v`: tuple of functions computing the vector fields ``v_i``, ``i = 1 ... 3``
 * `f`: tuple of functions computing the vector fields ``f_i``, ``i = 1 ... 3``
 * `ϕ`: primary constraints
@@ -35,65 +46,64 @@ the algebraic variables ``(\lambda, \gamma)`` taking values in
 ### Constructors
 
 ```julia
-SPDAE(DT, N, d, m, n, v, f, ϕ, ψ, t₀, q₀, p₀, λ₀; parameters=nothing, periodicity=zeros(DT,d))
-SPDAE(v, f, ϕ, ψ, t₀, q₀, p₀, λ₀=zero(q₀); kwargs...) = SPDAE(eltype(q₀), ndims(q₀), size(q₀,1), size(λ₀,1), size(q₀,2), v, f, ϕ, ψ, t₀, q₀, p₀, λ₀; kwargs...)
-SPDAE(v, f, ϕ, ψ, q₀, p₀, λ₀=zero(q₀); kwargs...) = SPDAE(v, f, ϕ, ψ, zero(eltype(q₀)), q₀, p₀, λ₀; kwargs...)
+SPDAE(v, f, ϕ, ψ, t₀, q₀, p₀, λ₀; parameters=nothing, periodicity=zeros(DT,d))
+SPDAE(v, f, ϕ, ψ, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...) = SPDAE(v, f, ϕ, ψ, 0.0, q₀, p₀, λ₀; kwargs...)
+SPDAE(v, f, ϕ, ψ, t₀, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...) = SPDAE(v, f, ϕ, ψ, t₀, [q₀], [p₀], [λ₀]; kwargs...)
+SPDAE(v, f, ϕ, ψ, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...) = SPDAE(v, f, ϕ, ψ, 0.0, q₀, p₀, λ₀; kwargs...)
 ```
 
 """
-struct SPDAE{dType <: Number, tType <: Real, vType <: Tuple, fType <: Tuple,
-             ϕType <: Function, ψType <: Function, pType <: Union{NamedTuple,Nothing},
-             N} <: AbstractEquationPDAE{dType, tType}
+struct SPDAE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
+             vType <: Tuple, fType <: Tuple, ϕType <: Function, ψType <: Function,
[…]
