+++ src/solutions/atomic_solution_dae.jl
@@ -11,16 +11,16 @@ Atomic solution for an DAE.
 ### Fields
 
 * `t`: time of current time step
-* `t̅`: time of previous time step
+* `t̄`: time of previous time step
 * `q`: current solution of q
-* `q̅`: previous solution of q
+* `q̄`: previous solution of q
 * `q̃`: compensated summation error of q
 * `λ`: current solution of λ
-* `λ̅`: previous solution of λ
+* `λ̄`: previous solution of λ
 * `v`: vector field of q
-* `v̅`: vector field of q̅
+* `v̄`: vector field of q̄
 * `u`: projective vector field of q
-* `u̅`: projective vector field of q̅
+* `ū`: projective vector field of q̄
 * `internal`: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)
 
 ### Constructors
@@ -38,20 +38,20 @@ AtomicSolutionDAE(t::TT, q::AT, λ::AT, internal::IT=NamedTuple())
 """
 mutable struct AtomicSolutionDAE{DT <: Number, TT <: Real, AT <: AbstractArray{DT}, IT <: NamedTuple} <: AtomicSolution{DT,TT,AT}
     t::TT
-    t̅::TT
+    t̄::TT
 
     q::AT
-    q̅::AT
+    q̄::AT
     q̃::AT
 
     λ::AT
-    λ̅::AT
+    λ̄::AT
 
     v::AT
-    v̅::AT
+    v̄::AT
 
     u::AT
-    u̅::AT
+    ū::AT
 
     internal::IT
 
@@ -99,11 +99,11 @@ function get_solution(asol::AtomicSolutionDAE)
 end
 
 function Common.reset!(asol::AtomicSolutionDAE, Δt)
-    asol.t̅  = asol.t
-    asol.q̅ .= asol.q
-    asol.λ̅ .= asol.λ
-    asol.v̅ .= asol.v
-    asol.u̅ .= asol.u
+    asol.t̄  = asol.t
+    asol.q̄ .= asol.q
+    asol.λ̄ .= asol.λ
+    asol.v̄ .= asol.v
+    asol.ū .= asol.u
     asol.t += Δt
 end
 
