### 2021-02-11 73a2ed0
Update equation docstrings for parameters and periodicity.
