+++ src/equations/hode.jl
@@ -38,6 +38,8 @@ taking values in ``\mathbb{R}^{d} \times \mathbb{R}^{d}``.
 * `t₀`: initial time
 * `q₀`: initial condition for dynamical variable ``q``
 * `p₀`: initial condition for dynamical variable ``p``
+* `parameters`: either a `NamedTuple` containing the equations parameters or `nothing`
+* `periodicity`: determines the periodicity of the state vector `q` for cutting periodic solutions
 
 ### Constructors
 
