+++ src/equations/ldae.jl
@@ -57,6 +57,8 @@ variables ``(q,p)`` and algebraic variables ``v``, ``\lambda`` and ``\mu``.
 * `p₀`: initial condition for `p`
 * `λ₀`: initial condition for `λ` (optional)
 * `μ₀`: initial condition for `μ` (optional)
+* `parameters`: either a `NamedTuple` containing the equations parameters or `nothing`
+* `periodicity`: determines the periodicity of the state vector `q` for cutting periodic solutions
 
 The functions `ϑ` and `f` must have the interface
 ```julia
