+++ src/equations/lode.jl
@@ -46,6 +46,8 @@ variables ``(q,p)`` and algebraic variable ``v``.
 * `q₀`: initial condition for `q`
 * `p₀`: initial condition for `p`
 * `λ₀`: initial condition for `λ` (optional)
+* `parameters`: either a `NamedTuple` containing the equations parameters or `nothing`
+* `periodicity`: determines the periodicity of the state vector `q` for cutting periodic solutions
 
 The functions `ϑ` and `f` must have the interface
 ```julia
