+++ src/equations/idae.jl
@@ -47,6 +47,8 @@ the algebraic variable ``\lambda`` taking values in ``\mathbb{R}^{n}``.
 * `q₀`: initial condition for dynamical variable ``q``
 * `p₀`: initial condition for dynamical variable ``p``
 * `λ₀`: initial condition for algebraic variable ``\lambda``
+* `parameters`: either a `NamedTuple` containing the equations parameters or `nothing`
+* `periodicity`: determines the periodicity of the state vector `q` for cutting periodic solutions
 
 ### Constructors
 
