### 2021-02-11 c06e21f
Generalise integrator constructors so that IODE integrators also accept LODE, PODE integrators also accept HODE, and analogously for the corersponding DAE equivalents.
