+++ src/integrators/spark/integrators_vspark.jl
@@ -77,7 +77,7 @@ struct IntegratorVSPARK{DT, TT, D, S, R, PT <: ParametersVSPARK{DT,TT,D,S,R},
         IntegratorVSPARK(params, solver, iguess, caches)
     end
 
-    function IntegratorVSPARK(equation::IDAE{DT,TT}, tableau::Union{TableauSPARK{TT},TableauVSPARK{TT}}, Δt::TT; kwargs...) where {DT,TT}
+    function IntegratorVSPARK(equation::Union{IDAE{DT,TT}, LDAE{DT,TT}}, tableau::Union{TableauSPARK{TT},TableauVSPARK{TT}}, Δt::TT; kwargs...) where {DT,TT}
         IntegratorVSPARK{DT, ndims(equation)}(get_function_tuple(equation), tableau, Δt; kwargs...)
     end
 end
