+++ src/integrators/dgvi/integrators_dgvi_experimental.jl
@@ -169,7 +169,7 @@ struct IntegratorDGVIEXP{DT, TT, D, S, R,
         IntegratorDGVIEXP(basis, quadrature, params, solver, iguess, caches)
     end
 
-    function IntegratorDGVIEXP(equation::IODE{DT,TT}, basis::Basis{TT}, quadrature::QuadratureRule{TT}, Δt::TT; kwargs...) where {DT,TT}
+    function IntegratorDGVIEXP(equation::Union{IODE{DT,TT}, LODE{DT,TT}}, basis::Basis{TT}, quadrature::QuadratureRule{TT}, Δt::TT; kwargs...) where {DT,TT}
         IntegratorDGVIEXP{DT, ndims(equation)}(get_function_tuple(equation), basis, quadrature, Δt; kwargs...)
     end
 end
