+++ src/integrators/spark/integrators_vspark_primary.jl
@@ -146,7 +146,7 @@ struct IntegratorVSPARKprimary{DT, TT, D, S, R, PT <: ParametersVSPARKprimary{DT
         IntegratorVSPARKprimary(params, solver, iguess, caches)
     end
 
-    function IntegratorVSPARKprimary(equation::IDAE{DT,TT}, tableau::TableauVSPARKprimary{TT}, Δt::TT; kwargs...) where {DT,TT}
+    function IntegratorVSPARKprimary(equation::Union{IDAE{DT,TT}, LDAE{DT,TT}}, tableau::TableauVSPARKprimary{TT}, Δt::TT; kwargs...) where {DT,TT}
         IntegratorVSPARKprimary{DT, ndims(equation)}(get_function_tuple(equation), tableau, Δt; kwargs...)
     end
 end
