### 2021-02-17 4204a8a
Implement convert function from PODE to ODE.
