### 2021-02-17 c9ea739
Clean up similar methods for equations.
