### 2021-02-17 bb5217a
Move equation conversion to separate file and add conversion from HODE to PODE.
