### 2021-02-17 e08b95d
Clean up equation tests.
