+++ test/equations/deterministic_equations_tests.jl
@@ -12,6 +12,25 @@ function ode_v(t, x, ẋ)
 end
 
 
+function v_sode_1(t, x, v)
+    v[1] = x[2]
+end
+
+function v_sode_2(t, x, v)
+    v[2] = 2x[1]
+end
+
+function q_sode_1(t, x̄, x)
+    x[1] = x̄[1]
+    x[2] = x̄[2]
+end
+
+function q_sode_2(t, x̄, x)
+    x[1] = x̄[1]
+    x[2] = x̄[2]
+end
+
+
 function pode_v(t, q, p, v)
     v[1] = p[1]
 end
@@ -43,23 +62,6 @@ end
 
 function iode_h(t, q, v)
     v[1]^2/2 + cos(q[1])
-end
-
-
-function f_sode_1(t, x, f)
-    f[1] = x[1]
-end
-
-function f_sode_2(t, x, f)
-    f[1] = x[1]^2
-end
-
-function q_sode_1(t, x̄, x)
-    x[1] = x̄[1]
-end
-
-function q_sode_2(t, x̄, x)
-    x[1] = x̄[1]
 end
 
 
@@ -128,10 +130,57 @@ end
     @test ode == ode2
     @test ode == ode3
 
-    @test hash(ode1) == hash(ode2) == hash(ode3)
+    @test hash(ode) == hash(ode1)
+    @test hash(ode) == hash(ode2)
+    @test hash(ode) == hash(ode3)
 
     @test ode == similar(ode, t₀, x₀)
     @test ode == similar(ode, x₀)
+
+end
+
+
+@testset "$(rpad("Split Ordinary Differential Equations (SODE)",80))" begin
+
+    v_sode = (v_sode_1, v_sode_2)
+    q_sode = (q_sode_1, q_sode_2)
+
+    sode  = SODE(v_sode, t₀, [x₀])
+    sode1 = SODE(v_sode, [x₀])
+    sode2 = SODE(v_sode, t₀, x₀)
+    sode3 = SODE(v_sode, x₀)
+
+    @test ndims(sode) == 2
+    @test nsamples(sode) == 1
+    @test periodicity(sode) == zero(x₀)
+
+    @test sode == sode1
+    @test sode == sode2
+    @test sode == sode3
+
+    @test hash(sode) == hash(sode1)
+    @test hash(sode) == hash(sode2)
+    @test hash(sode) == hash(sode3)
+
+    @test sode == similar(sode, t₀, x₀)
+    @test sode == similar(sode, x₀)
+
+
+    sode  = SODE(v_sode, q_sode, t₀, [x₀])
+    sode1 = SODE(v_sode, q_sode, [x₀])
+    sode2 = SODE(v_sode, q_sode, t₀, x₀)
+    sode3 = SODE(v_sode, q_sode, x₀)
+
+    @test sode == sode1
+    @test sode == sode2
+    @test sode == sode3
+
+    @test hash(sode) == hash(sode1)
+    @test hash(sode) == hash(sode2)
+    @test hash(sode) == hash(sode3)
+
+    @test sode == similar(sode, t₀, x₀)
+    @test sode == similar(sode, x₀)
 
 end
 
@@ -154,7 +203,9 @@ end
     @test pode == pode2
     @test pode == pode3
 
-    @test hash(pode1) == hash(pode2) == hash(pode3)
+    @test hash(pode) == hash(pode1)
+    @test hash(pode) == hash(pode2)
[…]
