### 2021-02-17 deb724e
Implement conversion from PODE and HODE to SODE.
