### 2021-02-22 #2954 @michakraus
@VivianePons Thanks for checking in on this.

> @michakraus have you worked on the different comments posted by @ranocha ?

I worked on most of the comments. However, in the process a few issues surfaced which are better fixed properly once and for all rather than quick and dirty. If things go well, I might finish this week, but depending on my workload it may also take a little longer...

Best,
Michael
