### 2021-03-04 #2954 @ChrisRackauckas
Hey sorry, got far behind. I'll start by saying I think GeometricIntegrators.jl is a great package and I've dug around it before, but now am digging around with a reviewer's eye. But in general, even before the review eye, I would think it's publication worthy.

```
>Is adaptive time-stepping or some way to estimate an appropriate step size supported?

We currently have no special infrastructure for adaptive time-stepping. However, we have used adaptive sub-cycling in special-purpose integrators that are not part of the main library. Here, the standard infrastructure is used to specify "target time steps", i.e., the adaptive integrator computes a solution for every point in the equidistant time series, but in between in adapts as it wishes. Whenever an adaptive time step would step over a "target time step", it is reduced to hit that target. For geometric integrators that often is the only sensible thing to do (and for most practical applications as well).
In all of this, the integrator has to take care of the sub-cycling and the rest of GeometricIntegrators doesn't really care about it. If for some reason you want to output the solution at the intermediate irregular time steps, this is relatively easily possible via the (still mostly undocumented) mid-level interface that is used to call the integrators.
If I am not mistaken, DifferentialEquations interpolates the solutions of the adaptive steps to get a regular output, but that would destroy any structure, so it is important that the integrator actually integrates to each regular time step.
```

I think this is worth commenting on because it's probably lost to the eyes of someone who's not deep in the field. There really aren't good ways to do geometric integration with adaptive time stepping in the general case. If you dig through OrdinaryDiffEq.jl where the DifferentialEquations.jl symplectic integrators are, you see the documentation mentions it's all fixed time step. There are a few ways to adapt time by incorporating it into the Hamiltonians, but they are more research at this stage and I haven't seen them work well. But, OrdinaryDiffEq.jl is all based around adaptive time stepping algorithms, since that's "generally" what users want... except in the geometric case. So having geometric integrators in OrdinaryDiffEq.jl is rather suboptimal because it could skip a lot of the code complexity if one's focus is fixed time step geometric integrators. We had no reason to implement them in the SciML ecosystem because GeometricIntegrators.jl exists as high performance implementations of these fixed time step algorithms, and the GeometricIntegratorsDiffEq.jl bridge then serves the purpose of filling that part of the ecosystem. And indeed interpolation-based approaches of `saveat` breaks the structure, and so you really need a fixed time implementation. So it fills a very complementary role in the ecosystem and it does it really well.

- [ ] I would say it would be good to add a note on this to the Statement of Need.

>Again a very good point. I'll look into that. Callable structs, however, seem to work well only for ODEs, where only one function for the vector field is specified. All the other equation types require several functions to be specified, thus callable structs would not be a good option.

It could make sense to give different callable structs to the two drift and diffusion functions of an SDE, each containing an appropriate cache. So I'm not quite convinced, but it's also not hard for the user to make the callable struct `<:Function` or wrap it in a `(args...) -> f(args...)` which would achieve the goal of making it a function with the callable.

- [ ] The developer documentation pages are empty.
- [ ] The interpolation docstring in https://juliagni.github.io/GeometricIntegrators.jl/latest/modules/interpolation/ doesn't mention how or when to use it.
- [ ] https://juliagni.github.io/GeometricIntegrators.jl/latest/modules/simulations/ isn't very informative.
- [ ] https://juliagni.github.io/GeometricIntegrators.jl/latest/equations/ is empty.
- [ ] It looks like everything was made intentionally generically typed, but that doesn't seem to be tested? It might be good to throw a few Dual number tests in there just to see if everything works out like you'd expect.
- [ ] I don't see a link to GeometricExamples in the docs? That might be good to mention.

All minor.
