### 2021-03-04 c3b25e8
Refurbish, clean up, simplify, generalise and unify equation interface.
