+++ src/equations/hode.jl
@@ -26,51 +26,72 @@ taking values in ``\mathbb{R}^{d} \times \mathbb{R}^{d}``.
 * `AT <: AbstractArray{DT}`: array type
 * `vType <: Function`: type of `v`
 * `fType <: Function`: type of `f`
-* `hType <: Function`: type of `h`
-* `pType <: Union{NamedTuple,Nothing}`: parameters type
+* `PType <: Function`: type of `P`
+* `hamType <: Function`: Hamiltonian type
+* `invType <: OptionalNamedTuple`: invariants type
+* `parType <: OptionalNamedTuple`: parameters type
+* `perType <: OptionalArray{AT}`: periodicity type
 
 ### Fields
 
 * `d`: dimension of dynamical variables ``q`` and ``p`` as well as the vector fields ``v`` and ``f``
 * `v`: function computing the vector field ``v``
 * `f`: function computing the vector field ``f``
-* `h`: function computing the Hamiltonian ``H``
+* `P`: function computing the Poisson matrix ``P``
 * `t₀`: initial time
 * `q₀`: initial condition for dynamical variable ``q``
 * `p₀`: initial condition for dynamical variable ``p``
-* `parameters`: either a `NamedTuple` containing the equations parameters or `nothing`
+* `hamiltonian`: function computing the Hamiltonian ``H``
+* `invariants`: either a `NamedTuple` containing the equation's invariants or `nothing`
+* `parameters`: either a `NamedTuple` containing the equation's parameters or `nothing`
 * `periodicity`: determines the periodicity of the state vector `q` for cutting periodic solutions
 
 ### Constructors
 
 ```julia
-HODE(v, f, h, t₀, q₀, p₀; parameters=nothing, periodicity=zero(q₀[begin]))
-HODE(v, f, h, q₀::StateVector, p₀::StateVector; kwargs...) = HODE(v, f, h, 0.0, q₀, p₀; kwargs...)
-HODE(v, f, h, t₀, q₀::State, p₀::State; kwargs...) = HODE(v, f, h, t₀, [q₀], [p₀]; kwargs...)
-HODE(v, f, h, q₀::State, p₀::State; kwargs...) = HODE(v, f, h, 0.0, q₀, p₀; kwargs...)
+HODE(v, f, poisson, t₀, q₀, p₀, hamiltonian, invariants, parameters, periodicity)
+
+HODE(v, f, h, t₀, q₀::StateVector, p₀::StateVector; kwargs...)
+HODE(v, f, h, q₀::StateVector, p₀::StateVector; kwargs...)
+HODE(v, f, h, t₀, q₀::State, p₀::State; kwargs...)
+HODE(v, f, h, q₀::State, p₀::State; kwargs...)
 ```
+
+### Keyword arguments:
+
+* `poisson = symplectic_matrix`
+* `invariants = nothing`
+* `parameters = nothing`
+* `periodicity = nothing`
 
 """
 struct HODE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
-            vType <: Function, fType <: Function, hType <: Function,
-            pType <: Union{NamedTuple,Nothing}} <: AbstractEquationPODE{dType, tType}
+            vType <: Function, fType <: Function, 
+            PType <: Function,
+            hamType <: Function,
+            invType <: OptionalNamedTuple,
+            parType <: OptionalNamedTuple,
+            perType <: OptionalArray{arrayType}} <: AbstractEquationPODE{dType, tType}
 
     d::Int
+
     v::vType
     f::fType
-    h::hType
+    P::PType
+
     t₀::tType
[…]
