+++ src/equations/pode.jl
@@ -20,19 +20,20 @@ with vector fields ``v`` and ``f``, initial conditions ``(q_{0}, p_{0})`` and th
 * `AT <: AbstractArray{DT}`: array type
 * `vType <: Function`: type of `v`
 * `fType <: Function`: type of `f`
-* `hType <: OptionalFunction`: type of `h`
-* `pType <: Union{NamedTuple,Nothing}`: parameters type
+* `invType <: OptionalNamedTuple`: invariants type
+* `parType <: OptionalNamedTuple`: parameters type
+* `perType <: OptionalArray{AT}`: periodicity type
 
 ### Fields
 
 * `d`: dimension of dynamical variables ``q`` and ``p`` as well as the vector fields ``v`` and ``f``
 * `v`: function computing the vector field ``v``
 * `f`: function computing the vector field ``f``
-* `h`: function computing the Hamiltonian (optional)
 * `t₀`: initial time
 * `q₀`: initial condition for `q`
 * `p₀`: initial condition for `p`
-* `parameters`: either a `NamedTuple` containing the equations parameters or `nothing`
+* `invariants`: either a `NamedTuple` containing the equation's invariants or `nothing`
+* `parameters`: either a `NamedTuple` containing the equation's parameters or `nothing`
 * `periodicity`: determines the periodicity of the state vector `q` for cutting periodic solutions
 
 The functions `v` and `f` must have the interface
@@ -58,33 +59,44 @@ vector fields ``v`` and ``f`` on `t`, `q` and `p`.
 ### Constructors
 
 ```julia
-PODE(v, f, t₀, q₀, p₀; h=nothing, parameters=nothing, periodicity=zero(q₀[begin]))
-PODE(v, f, q₀::StateVector, p₀::StateVector; kwargs...) = PODE(v, f, 0.0, q₀, p₀; kwargs...)
-PODE(v, f, t₀, q₀::State, p₀::State; kwargs...) = PODE(v, f, t₀, [q₀], [p₀]; kwargs...)
-PODE(v, f, q₀::State, p₀::State; kwargs...) = PODE(v, f, 0.0, q₀, p₀; kwargs...)
+PODE(v, f, t₀, q₀, p₀, invariants, parameters, periodicity)
+
+PODE(v, f, h, t₀, q₀::StateVector, p₀::StateVector; kwargs...)
+PODE(v, f, q₀::StateVector, p₀::StateVector; kwargs...)
+PODE(v, f, t₀, q₀::State, p₀::State; kwargs...)
+PODE(v, f, q₀::State, p₀::State; kwargs...)
 ```
+
+### Keyword arguments:
+
+* `invariants = nothing`
+* `parameters = nothing`
+* `periodicity = nothing`
 
 """
 struct PODE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
-            vType <: Function, fType <: Function, hType <: OptionalFunction,
-            pType <: Union{NamedTuple,Nothing}} <: AbstractEquationPODE{dType, tType}
+            vType <: Function, fType <: Function,
+            invType <: OptionalNamedTuple,
+            parType <: OptionalNamedTuple,
+            perType <: OptionalArray{arrayType}} <: AbstractEquationPODE{dType, tType}
 
     d::Int
+
     v::vType
     f::fType
-    h::hType
+
     t₀::tType
     q₀::Vector{arrayType}
     p₀::Vector{arrayType}
-    parameters::pType
-    periodicity::arrayType
 
-    function PODE(v::vType, f::fType, t₀::tType, q₀::Vector{arrayType}, p₀::Vector{arrayType};
[…]
