### 2021-03-04 ab7db00
Generalise integrator constructors.
