+++ src/integrators/rk/integrators_firk_implicit.jl
@@ -75,9 +75,9 @@ struct IntegratorFIRKimplicit{DT, TT, D, S, PT <: ParametersFIRKimplicit{DT,TT},
         # check if tableau is fully implicit
         if get_config(:verbosity) ≥ 1
             if isexplicit(tableau)
-                @warn "Initializing IntegratorFIRK with explicit tableau $(q.name).\nYou might want to use IntegratorERK instead."
+                @warn "Initializing IntegratorFIRK with explicit tableau $(tableau.name).\nYou might want to use IntegratorERK instead."
             elseif isdiagnonallyimplicit(tableau)
-                @warn "Initializing IntegratorFIRK with diagonally implicit tableau $(q.name).\nYou might want to use IntegratorDIRK instead."
+                @warn "Initializing IntegratorFIRK with diagonally implicit tableau $(tableau.name).\nYou might want to use IntegratorDIRK instead."
             end
         end
 
