### 2021-03-04 94b9c82
Generalise some solution methods.
