+++ src/equations/sode.jl
@@ -19,8 +19,9 @@ v (t) = v_1 (t) + ... + v_r (t) .
 * `AT <: AbstractArray{DT}`: array type
 * `vType <: Union{Tuple,Nothing}`: type of `v`
 * `qType <: Union{Tuple,Nothing}`: type of `q`
-* `hType <: OptionalFunction`: type of `h`
-* `pType <: Union{NamedTuple,Nothing}`: parameters type
+* `invType <: OptionalNamedTuple`: invariants type
+* `parType <: OptionalNamedTuple`: parameters type
+* `perType <: OptionalArray{AT}`: periodicity type
 
 ### Fields
 
@@ -29,7 +30,8 @@ v (t) = v_1 (t) + ... + v_r (t) .
 * `q`: tuple of functions computing the solution
 * `t₀`: initial time
 * `q₀`: initial condition
-* `parameters`: either a `NamedTuple` containing the equations parameters or `nothing`
+* `invariants`: either a `NamedTuple` containing the equation's invariants or `nothing`
+* `parameters`: either a `NamedTuple` containing the equation's parameters or `nothing`
 * `periodicity`: determines the periodicity of the state vector `q` for cutting periodic solutions
 
 The functions `v_i` providing the vector field must have the interface
@@ -60,59 +62,80 @@ methods, e.g., it allows to use another integrator for solving substeps.
 ### Constructors
 
 ```julia
-SODE(v, q, t₀, q₀; parameters=nothing, periodicity=zero(q₀[begin]))
+SODE(v, q, t₀, q₀, invariants, parameters, periodicity)
 
-SODE(v, q::Union{Tuple,Nothing}, t₀::Real, q₀::State; kwargs...) = SODE(v, q, t₀, [q₀]; kwargs...)
-SODE(v, q::Union{Tuple,Nothing}, q₀::StateVector; kwargs...) = SODE(v, q, 0.0, q₀; kwargs...)
-SODE(v, q::Union{Tuple,Nothing}, q₀::State; kwargs...) = SODE(v, q, 0.0, q₀; kwargs...)
+SODE(v, q::Union{Tuple,Nothing}, t₀::Real, q₀::StateVector; kwargs...)
+SODE(v, q::Union{Tuple,Nothing}, t₀::Real, q₀::State; kwargs...)
+SODE(v, q::Union{Tuple,Nothing}, q₀::StateVector; kwargs...)
+SODE(v, q::Union{Tuple,Nothing}, q₀::State; kwargs...)
 
-SODE(v, t₀::Real, q₀::StateVector; kwargs...) = SODE(v, nothing, t₀, q₀; kwargs...)
-SODE(v, t₀::Real, q₀::State; kwargs...) = SODE(v, nothing, t₀, [q₀]; kwargs...)
-SODE(v, q₀::StateVector; kwargs...) = SODE(v, nothing, 0.0, q₀; kwargs...)
-SODE(v, q₀::State; kwargs...) = SODE(v, nothing, 0.0, q₀; kwargs...)
+SODE(v, t₀::Real, q₀::StateVector; kwargs...)
+SODE(v, t₀::Real, q₀::State; kwargs...)
+SODE(v, q₀::StateVector; kwargs...)
+SODE(v, q₀::State; kwargs...)
 ```
+
+### Keyword arguments:
+
+* `invariants = nothing`
+* `parameters = nothing`
+* `periodicity = nothing`
 
 """
 struct SODE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
             vType <: Union{Tuple,Nothing}, qType <: Union{Tuple,Nothing},
-            pType <: Union{NamedTuple,Nothing}} <: AbstractEquationODE{dType, tType}
+            invType <: OptionalNamedTuple,
[…]
