+++ src/equations/spdae.jl
@@ -27,8 +27,9 @@ the algebraic variables ``(\lambda, \gamma)`` taking values in
 * `fType <: Function`: type of `f`
 * `ϕType <: Function`: type of `ϕ`
 * `ψType <: Function`: type of `ψ`
-* `hType <: OptionalFunction`: type of `h`
-* `pType <: Union{NamedTuple,Nothing}`: parameters type
+* `invType <: OptionalNamedTuple`: invariants type
+* `parType <: OptionalNamedTuple`: parameters type
+* `perType <: OptionalArray{AT}`: periodicity type
 
 ### Fields
 
@@ -42,66 +43,93 @@ the algebraic variables ``(\lambda, \gamma)`` taking values in
 * `q₀`: initial condition for dynamical variable ``q``
 * `p₀`: initial condition for dynamical variable ``p``
 * `λ₀`: initial condition for algebraic variable ``λ``
-* `parameters`: either a `NamedTuple` containing the equations parameters or `nothing`
+* `invariants`: either a `NamedTuple` containing the equation's invariants or `nothing`
+* `parameters`: either a `NamedTuple` containing the equation's parameters or `nothing`
 * `periodicity`: determines the periodicity of the state vector `q` for cutting periodic solutions
 
 ### Constructors
 
 ```julia
-SPDAE(v, f, ϕ, ψ, t₀, q₀, p₀, λ₀; parameters=nothing, periodicity=zeros(DT,d))
-SPDAE(v, f, ϕ, ψ, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...) = SPDAE(v, f, ϕ, ψ, 0.0, q₀, p₀, λ₀; kwargs...)
-SPDAE(v, f, ϕ, ψ, t₀, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...) = SPDAE(v, f, ϕ, ψ, t₀, [q₀], [p₀], [λ₀]; kwargs...)
-SPDAE(v, f, ϕ, ψ, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...) = SPDAE(v, f, ϕ, ψ, 0.0, q₀, p₀, λ₀; kwargs...)
+SPDAE(v, f, ϕ, ψ, t₀, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...)
+SPDAE(v, f, ϕ, ψ, q₀::StateVector, p₀::StateVector, λ₀::StateVector; kwargs...)
+SPDAE(v, f, ϕ, ψ, t₀, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...)
+SPDAE(v, f, ϕ, ψ, q₀::State, p₀::State, λ₀::State=zero(q₀); kwargs...)
 ```
+
+### Keyword arguments:
+
+* `invariants = nothing`
+* `parameters = nothing`
+* `periodicity = nothing`
 
 """
 struct SPDAE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
-             vType <: Tuple, fType <: Tuple, ϕType <: Function, ψType <: Function,
-             pType <: Union{NamedTuple,Nothing}} <: AbstractEquationPDAE{dType, tType}
+             vType <: Tuple, fType <: Tuple, ϕType <: Function, ψType <: OptionalFunction,
+             invType <: OptionalNamedTuple,
+             parType <: OptionalNamedTuple,
+             perType <: OptionalArray{arrayType}} <: AbstractEquationPDAE{dType, tType}
+
     d::Int
[…]
