### 2021-03-04 90a36f9
Add second Lagrange multiplier to HDAE.
