+++ src/equations/hdae.jl
@@ -110,6 +110,7 @@ struct HDAE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
     q₀::Vector{arrayType}
     p₀::Vector{arrayType}
     λ₀::Vector{arrayType}
+    μ₀::Vector{arrayType}
 
     hamiltonian::hamType
     invariants::invType
@@ -117,7 +118,7 @@ struct HDAE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
     periodicity::perType
 
     function HDAE(v, f, u, g, ϕ, ū, ḡ, ψ, v̄, f̄, P,
-                  t₀::tType, q₀::Vector{arrayType}, p₀::Vector{arrayType}, λ₀::Vector{arrayType},
+                  t₀::tType, q₀::Vector{arrayType}, p₀::Vector{arrayType}, λ₀::Vector{arrayType}, μ₀::Vector{arrayType},
                   hamiltonian, invariants, parameters, periodicity) where {
                         dType <: Number, tType <: Real, arrayType <: AbstractArray{dType}}
 
@@ -131,8 +132,9 @@ struct HDAE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
         @assert all(length(q) == d for q in q₀)
         @assert all(length(p) == d for p in p₀)
         @assert all(length(λ) == m for λ in λ₀)
+        @assert all(length(μ) == m for μ in μ₀)
 
-        @assert all([ndims(q) == ndims(p) == ndims(λ) for (q,p,λ) in zip(q₀,p₀,λ₀)])
+        @assert all([ndims(q) == ndims(p) == ndims(λ) == ndims(μ) for (q,p,λ,μ) in zip(q₀,p₀,λ₀,μ₀)])
 
         new{dType, tType, arrayType,
             typeof(v), typeof(f),
@@ -140,22 +142,22 @@ struct HDAE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
             typeof(ū), typeof(ḡ), typeof(ψ),
             typeof(v̄), typeof(f̄), typeof(P),
             typeof(hamiltonian), typeof(invariants), typeof(parameters), typeof(periodicity)}(
-                d, m, v, f, u, g, ϕ, ū, ḡ, ψ, v̄, f̄, P, t₀, q₀, p₀, λ₀,
+                d, m, v, f, u, g, ϕ, ū, ḡ, ψ, v̄, f̄, P, t₀, q₀, p₀, λ₀, μ₀,
                 hamiltonian, invariants, parameters, periodicity)
     end
 end
 
-_HDAE(v, f, u, g, ϕ, ū, ḡ, ψ, hamiltonian, t₀, q₀, p₀, λ₀; v̄=v, f̄=f, P=symplectic_matrix, invariants=nothing, parameters=nothing, periodicity=nothing) = HDAE(v, f, u, g, ϕ, ū, ḡ, ψ, v̄, f̄, P, t₀, q₀, p₀, λ₀, hamiltonian, invariants, parameters, periodicity)
[…]
