+++ src/equations/dae.jl
@@ -41,6 +41,7 @@ taking values in ``\mathbb{R}^{n}``.
 * `t₀`: initial time
 * `q₀`: initial condition for dynamical variable ``q``
 * `λ₀`: initial condition for algebraic variable ``\lambda``
+* `μ₀`: initial condition for algebraic variable ``μ`` (optional)
 * `invariants`: either a `NamedTuple` containing the equation's invariants or `nothing`
 * `parameters`: either a `NamedTuple` containing the equation's parameters or `nothing`
 * `periodicity`: determines the periodicity of the state vector `q` for cutting periodic solutions
@@ -58,15 +59,15 @@ on `t`, `q` and `λ`.
 ```julia
 DAE(v, u, ū, ϕ, ψ, v̄, t₀, q₀, λ₀, invariants, parameters, periodicity)
 
-DAE(v, u, ϕ, t₀, q₀::StateVector, λ₀::StateVector; kwargs...)
-DAE(v, u, ϕ, q₀::StateVector, λ₀::StateVector; kwargs...)
-DAE(v, u, ϕ, t₀, q₀::State, λ₀::State; kwargs...)
-DAE(v, u, ϕ, q₀::State, λ₀::State; kwargs...)
+DAE(v, u, ϕ, t₀, q₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
+DAE(v, u, ϕ, q₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
+DAE(v, u, ϕ, t₀, q₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
+DAE(v, u, ϕ, q₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
 
-DAE(v, u, ū, ϕ, ψ, t₀, q₀::StateVector, λ₀::StateVector; kwargs...)
-DAE(v, u, ū, ϕ, ψ, q₀::StateVector, λ₀::StateVector; kwargs...)
-DAE(v, u, ū, ϕ, ψ, t₀, q₀::State, λ₀::State; kwargs...)
-DAE(v, u, ū, ϕ, ψ, q₀::State, λ₀::State; kwargs...)
+DAE(v, u, ū, ϕ, ψ, t₀, q₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
+DAE(v, u, ū, ϕ, ψ, q₀::StateVector, λ₀::StateVector, μ₀::StateVector=zero(λ₀); kwargs...)
+DAE(v, u, ū, ϕ, ψ, t₀, q₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
+DAE(v, u, ū, ϕ, ψ, q₀::State, λ₀::State, μ₀::State=zero(λ₀); kwargs...)
 ```
 
 ### Example
@@ -106,21 +107,25 @@ struct DAE{dType <: Number, tType <: Real, arrayType <: AbstractArray{dType},
 
     d::Int
     m::Int
+
     v::vType
     u::uType
     ū::ūType
     ϕ::ϕType
     ψ::ψType
     v̄::v̄Type
+
     t₀::tType
     q₀::Vector{arrayType}
     λ₀::Vector{arrayType}
+    μ₀::Vector{arrayType}
+    
     invariants::invType
     parameters::parType
     periodicity::perType
 
     function DAE(v::vType, u::uType, ū::ūType, ϕ::ϕType, ψ::ψType, v̄::v̄Type,
[…]
