### 2021-03-04 63e4691
Clean up and add equation tests.
