+++ test/equations/deterministic_equations_tests.jl
@@ -160,6 +160,9 @@ end
 
     @test get_function_tuple(ode) == NamedTuple{(:v,)}((ode_v,))
 
+    @test ode == similar(ode, t₀, x₀)
+    @test ode == similar(ode, x₀)
+
     @test ode == ode1
     @test ode == ode2
     @test ode == ode3
@@ -170,9 +173,6 @@ end
     @test hash(ode) == hash(ode3)
     @test hash(ode) == hash(ode4)
 
-    @test ode == similar(ode, t₀, x₀)
-    @test ode == similar(ode, x₀)
-
 end
 
 
@@ -181,42 +181,48 @@ end
     v_sode = (v_sode_1, v_sode_2)
     q_sode = (q_sode_1, q_sode_2)
 
-    sode  = SODE(v_sode, t₀, [x₀])
-    sode1 = SODE(v_sode, [x₀])
-    sode2 = SODE(v_sode, t₀, x₀)
-    sode3 = SODE(v_sode, x₀)
+    sode  = SODE(v_sode, nothing, t₀, [x₀], nothing, nothing, nothing)
+    sode1 = SODE(v_sode, t₀, [x₀])
+    sode2 = SODE(v_sode, [x₀])
+    sode3 = SODE(v_sode, t₀, x₀)
+    sode4 = SODE(v_sode, x₀)
 
     @test ndims(sode) == 2
     @test nsamples(sode) == 1
     @test periodicity(sode) == zero(x₀)
 
+    @test sode == similar(sode, t₀, x₀)
+    @test sode == similar(sode, x₀)
+
     @test sode == sode1
     @test sode == sode2
     @test sode == sode3
+    @test sode == sode4
 
     @test hash(sode) == hash(sode1)
     @test hash(sode) == hash(sode2)
     @test hash(sode) == hash(sode3)
-
-    @test sode == similar(sode, t₀, x₀)
-    @test sode == similar(sode, x₀)
+    @test hash(sode) == hash(sode4)
 
 
     sode  = SODE(v_sode, q_sode, t₀, [x₀])
-    sode1 = SODE(v_sode, q_sode, [x₀])
-    sode2 = SODE(v_sode, q_sode, t₀, x₀)
-    sode3 = SODE(v_sode, q_sode, x₀)
+    sode1 = SODE(v_sode, q_sode, t₀, [x₀])
+    sode2 = SODE(v_sode, q_sode, [x₀])
+    sode3 = SODE(v_sode, q_sode, t₀, x₀)
+    sode4 = SODE(v_sode, q_sode, x₀)
+
+    @test sode == similar(sode, t₀, x₀)
+    @test sode == similar(sode, x₀)
 
     @test sode == sode1
     @test sode == sode2
     @test sode == sode3
+    @test sode == sode4
 
     @test hash(sode) == hash(sode1)
     @test hash(sode) == hash(sode2)
     @test hash(sode) == hash(sode3)
-
-    @test sode == similar(sode, t₀, x₀)
-    @test sode == similar(sode, x₀)
+    @test hash(sode) == hash(sode4)
 
 end
 
@@ -247,6 +253,9 @@ end
     @test functions.v == pode_v == pode.v
     @test functions.f == pode_f == pode.f
 
+    @test pode == similar(pode, t₀, q₀, p₀)
+    @test pode == similar(pode, q₀, p₀)
+
     @test pode == pode1
     @test pode == pode2
[…]
