### 2021-03-04 1c6eb50
Fix similar methods for DAEs.
