### 2021-03-04 ad19de0
Update release notes for v0.8.0.
