### 2021-03-04 9f9e80c
Add Base.show method for PGLRK integrator.
