+++ src/integrators/rk/integrators_pglrk.jl
@@ -222,6 +222,12 @@ struct IntegratorPGLRK{DT, TT, D, S, PT <: ParametersPGLRK{DT,TT},
     end
 end
 
+"Print Projected Gauss-Legendre Runge-Kutta integrator."
+function Base.show(io::IO, int::IntegratorPGLRK)
+    print(io, "Projected Gauss-Legendre Runge-Kutta Integrator with ")
+    print(io, "  timestep = ", int.params.Δt)
+    print(io, "  tableau  = ", int.params.tab)
+end
 
 @inline Base.ndims(::IntegratorPGLRK{DT,TT,D,S}) where {DT,TT,D,S} = D
 @inline nstages(::IntegratorPGLRK{DT,TT,D,S}) where {DT,TT,D,S} = S
