### 2021-03-04 6b01f16
Update release notes for v0.8.0.
