### 2021-03-04 a8fff44
Minor bugfixes and refactoring of equations.
