### 2021-03-16 9e4c6a7
Add missing docstrings to Runge-Kutta integrators.
