+++ src/integrators/rk/integrators_srk_implicit.jl
@@ -55,7 +55,34 @@ end
 @inline CacheType(ST, params::ParametersSRKimplicit{DT,TT,D,S}) where {DT,TT,D,S} = IntegratorCacheSRKimplicit{ST,D,S}
 
 
-"Fully implicit Runge-Kutta integrator."
+@doc raw"""
+Special implicit Runge-Kutta integrator for certain noncanonical
+symplectic equations solving the system
+```math
+\begin{aligned}
+P_{n,i} &= \vartheta (Q_{n,i}, V_{n,i}) , &
+Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , &
+q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\
+F_{n,i} &= f (Q_{n,i}, V_{n,i}) , &
+P_{n,i} &= p_{n} + h  \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , &
+p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} ,
+\end{aligned}
+```
+Usually we are interested in Lagrangian systems, where
+```math
+\begin{aligned}
+P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , &
+F_{n,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) ,
+\end{aligned}
+```
+and tableaus satisfying the symplecticity conditions
+```math
+\begin{aligned}
+b_{i} \bar{a}_{ij} + \bar{b}_{j} a_{ji} &= b_{i} \bar{b}_{j} , &
+\bar{b}_i &= b_i .
+\end{aligned}
+```
+"""
 struct IntegratorSRKimplicit{DT, TT, D, S, PT <: ParametersSRKimplicit{DT,TT},
                                     ST <: NonlinearSolver{DT},
                                     IT <: InitialGuessODE{TT}} <: AbstractIntegratorRK{DT,TT}
