+++ src/integrators/rk/integrators_flrk.jl
@@ -179,7 +179,7 @@ function compute_stages!(x::Vector{ST}, Q::Vector{Vector{ST}}, V::Vector{Vector{
     end
 end
 
-"Compute stages of formal Lagrangian Runge-Kutta methods."
+# Compute stages of formal Lagrangian Runge-Kutta methods.
 function function_stages!(x::Vector{ST}, b::Vector{ST}, params::ParametersFLRK{DT,TT,D,S},
                           caches::CacheDict) where {ST,DT,TT,D,S}
     # temporary variables
