### 2021-03-16 98448d9
Remove obsolete abstract RK tableaus.
