### 2021-05-17 fa8e7b1
Update docstring of Hermite interpolation.
