### 2021-05-17 5f50b1c
Add comments on scope of applications and fixed vs. adaptive time stepping in JOSS paper.
