+++ docs/joss/paper.md
@@ -35,12 +35,17 @@ Differential equations are ubiquitous in science and engineering. Many equations
 Robust, performant and structure-preserving solvers for different types of differential equations are thus needed across many disciplines. *GeometricIntegrators.jl* provides such solvers and makes them available for both direct use as well as integration into other codes. Furthermore, the implemented algorithms can also be used within the *DifferentialEquations.jl* ecosystem [@Rackauckas:2017], which is the defacto standard differential equation solver for the Julia programming language [@Bezanson:2017].
 
 *GeometricIntegrators.jl* provides a comprehensive library of existing geometric integration as well as non-geometric algorithms, such as explicit, implicit, partitioned and stochastic Runge-Kutta methods, SPARK methods, splitting methods, symplectic methods and variational integrators. Most integrators are implemented in an abstract way that allows for the flexible choice of tableaus, approximation spaces, basis functions, quadrature rules, and thus order of convergence.
+As most geometric integrators are not easily combined with time step adaption in a structure-preserving way, GeometricIntegrators.jl does not provide any general infrastructure for adaptive time stepping. Nonetheless, individual integrators can implement their own adaptivity strategies as long as they provide a solution at a predefined, equidistant time series.
+
 *GeometricIntegrators.jl* also serves as a testbed for the development and analysis of novel algorithms [@Kraus:2017; @Kraus:2020; @KrausTyranowski:2020]. Due to the modular structure and the use of the multiple dispatch paradigm, the library can easily be extended, e.g., towards new algorithms or new types of equations. The library is designed to minimize overhead and maximize performance in order to be able to perform simulations with millions or even billions of time steps, facilitating the study of the long-time behaviour of both numerical algorithms and dynamical systems.
+The current scope of applications is mainly small- to mid-size systems of differential equations, e.g., systems of ordinary differential equations or semidiscretisations of partial differential equations with a moderate number of degrees of freedom.
+It is envisaged that in the future GeometricIntegrators.jl will also be able to address larger problems, especially semidiscretisations of partial differential equations in higher dimensions. Many elements required for this are already in place, e.g., support for general solution data structures, but others such as interfaces to appropriate iterative and parallel linear solvers are still lacking.
 
 
 # Other Software
 
-A Julia package closely related to *GeometricIntegrators.jl* is *DifferentialEquations.jl* [@Rackauckas:2017]. Althouh the two libraries provide similar functionality, their scope is rather different. While *DifferentialEquations.jl* provides a feature-rich ecosystem for the solution of differential equations, the focus of *GeometricIntegrators.jl* is on algorithms. In fact, *GeometricIntegrators.jl* can be used as backend for *DifferentialEquations.jl* via [GeometricIntegratorsDiffEq.jl](https://github.com/JuliaDiffEq/GeometricIntegratorsDiffEq.jl).
+A Julia package closely related to *GeometricIntegrators.jl* is *DifferentialEquations.jl* [@Rackauckas:2017]. Althouh the two libraries provide similar functionality, their scope is rather different. While *DifferentialEquations.jl* provides a feature-rich ecosystem for the solution of differential equations, the focus of *GeometricIntegrators.jl* is on algorithms. 
+While *DifferentialEquations.jl* is based around adaptive time stepping algorithms, GeometricIntegrators.jl focuses on fixed time step methods, given that structure-preservation is not easily combined with time step adaptivity. Therefore both packages are rather complementary, and *GeometricIntegrators.jl* can in fact be used as backend for *DifferentialEquations.jl* via [GeometricIntegratorsDiffEq.jl](https://github.com/JuliaDiffEq/GeometricIntegratorsDiffEq.jl).
 
 
 # Quality control and contributions
