### 2021-05-17 9bf092b
Minor cleanup in package description in docs.
