+++ src/integrators/extrapolation/midpoint.jl
@@ -1,71 +1,5 @@
 
 """
-Compute p(x) where p is the unique polynomial of degree length(xi),
-such that p(x[i]) = y[i]) for all i.
-
-    ti: interpolation nodes
-    xi: interpolation values
-    t:  evaluation point
-    x:  evaluation value
-"""
-function aitken_neville(ti::Vector{TT}, xi::Matrix, t::TT, x::Vector) where {TT}
-    @assert length(ti) == size(xi,2)
-    @assert length(x)  == size(xi,1)
-
-    for j in axes(t)
-        for i in 1:(length(t)-j)
-            for k in axes(x,1)
-                xi[k,i] = xi[k,i+1] + (xi[k,i] - xi[k,i+1]) * (ti[i+j] - t) / (ti[i+j] - ti[i])
-            end
-        end
-    end
-    for k in eachindex(x)
-        x[k] = xi[k,1]
-    end
-end
-
-
-"""
-Euler extrapolation method with arbitrary order p.
-
-    v:  function to compute vector field
-    t₀: initial time
-    t₁: final   time
-    x₀: initial value
-    x₁: final   value
-    s:  number of interpolations (order p=s+1)
-
-# TODO This is probably broken!
-"""
-function euler_extrapolation(v::Function, t₀::TT, t₁::TT, x₀::Vector, x₁::Vector, s::Int) where {TT}
-    @assert size(x₀) == size(x₁)
-
-    F   = collect(1:(s+1))
-    Δt  = t₁ - t₀
-    σ   = Δt ./ F
-    pts = repmat(x₀, 1, s+1)
-
-    local xᵢ = zero(x₀)
-    local vᵢ = zero(x₀)
-
-    for i in 1:(s+1)
-        for j in 1:(F[i]-1)
-            tᵢ  = t₀ + σ[i]
-            for k in axes(pts,1)
-                xᵢ[k] = pts[k,i]
-            end
-            v(tᵢ, xᵢ, vᵢ)
-            for k in axes(pts,1)
-                pts[k,i] += σ[i] * vᵢ[k]
-            end
-        end
-    end
-
-    aitken_neville(σ, pts, zero(TT), x₁)
-end
-
-
-"""
 Midpoint extrapolation method with arbitrary order p.
 
     v:  function to compute vector field
