+++ src/integrators/rk/integrators_firk_implicit.jl
@@ -100,7 +100,7 @@ struct IntegratorFIRKimplicit{DT, TT, D, S, PT <: ParametersFIRKimplicit{DT,TT},
         solver = create_nonlinear_solver(DT, D*(S+1), params, caches)
 
         # create initial guess
-        iguess = InitialGuessODE(get_config(:ig_interpolation), equations[:v̄], Δt)
+        iguess = InitialGuessODE(get_config(:ig_extrapolation), equations[:v̄], Δt)
 
         # create integrator
         IntegratorFIRKimplicit(params, solver, iguess, caches)
