+++ src/integrators/dgvi/integrators_dgvi_projection_initial.jl
@@ -159,7 +159,7 @@ struct IntegratorDGVIP0{DT, TT, D, S, R,
     end
 
     function IntegratorDGVIP0{DT,D}(equations::NamedTuple, basis::Basis{TT}, quadrature::QuadratureRule{TT,R}, Δt::TT;
-                                  interpolation=HermiteInterpolation{DT}) where {DT,TT,D,R}
+                                  interpolation=HermiteExtrapolation{DT}) where {DT,TT,D,R}
 
         # get number of stages
         S = nbasis(basis)
@@ -174,7 +174,7 @@ struct IntegratorDGVIP0{DT, TT, D, S, R,
         solver = create_nonlinear_solver(DT, D*(S+1), params, caches)
 
         # create initial guess
-        iguess = InitialGuessODE(get_config(:ig_interpolation), equations[:v̄], Δt)
+        iguess = InitialGuessODE(get_config(:ig_extrapolation), equations[:v̄], Δt)
 
         # create integrator
         IntegratorDGVIP0(basis, quadrature, params, solver, iguess, caches)
