+++ docs/src/modules/interpolation.md
@@ -1,7 +1,0 @@
-
-# Interpolation
-
-```@autodocs
-Modules = [GeometricIntegrators.Interpolation]
-Order   = [:constant, :type, :macro, :function]
-```
