+++ src/integrators/extrapolation/euler.jl
@@ -9,7 +9,6 @@ Euler extrapolation method with arbitrary order p.
     x₁: final   value
     s:  number of interpolations (order p=s+1)
 
-# TODO This is probably broken!
 """
 function euler_extrapolation(v::Function, t₀::TT, t₁::TT, x₀::Vector, x₁::Vector, s::Int) where {TT}
     @assert size(x₀) == size(x₁)
@@ -17,14 +16,14 @@ function euler_extrapolation(v::Function, t₀::TT, t₁::TT, x₀::Vector, x₁
     F   = collect(1:(s+1))
     Δt  = t₁ - t₀
     σ   = Δt ./ F
-    pts = repmat(x₀, 1, s+1)
+    pts = repeat(x₀, outer = [1, s+1])
 
     local xᵢ = zero(x₀)
     local vᵢ = zero(x₀)
 
-    for i in 1:(s+1)
-        for j in 1:(F[i]-1)
-            tᵢ  = t₀ + σ[i]
+    for i in F
+        for _ in 1:(F[i]-1)
+            tᵢ = t₀ + σ[i]
             for k in axes(pts,1)
                 xᵢ[k] = pts[k,i]
             end
@@ -35,5 +34,5 @@ function euler_extrapolation(v::Function, t₀::TT, t₁::TT, x₀::Vector, x₁
         end
     end
 
-    aitken_neville(σ, pts, zero(TT), x₁)
+    aitken_neville!(σ, pts, zero(TT), x₁)
 end
