### 2021-05-18 5ea026b
Add tests for extrapolation methods.
