### 2021-05-18 d83b3bf
Remove type argument from abstract Extrapolator type.
