### 2021-05-18 222692c
Revise Euler extrapolation.
