+++ src/integrators/extrapolation/euler.jl
@@ -1,16 +1,32 @@
 
-"""
+@doc raw"""
 Euler extrapolation method with arbitrary order p.
 
-    v:  function to compute vector field
-    t₀: initial time
-    t₁: final   time
-    x₀: initial value
-    x₁: final   value
-    s:  number of interpolations (order p=s+1)
+Solves the ordinary differential equation
+```math
+\begin{aligned}
+\dot{x} &= v(t, x) , &
+x(t_0) &= x_0 ,
+\end{aligned}
+```
+for $x_1 = x(t_1)$.
+
+Call with
+```julia
+_euler_extrapolation_ode!(v, t₀, t₁, x₀, x₁, s)
+```
+
+where
+
+* `v`:  function to compute vector field with signature `v(t,x,ẋ)`
+* `t₀`: initial time
+* `t₁`: final   time
+* `x₀`: initial value $x_0 = x(t_0)$
+* `x₁`: final   value $x_1 = x(t_1)$
+* `s`:  number of interpolations (order $p=s+1$)
 
 """
-function euler_extrapolation(v::Function, t₀::TT, t₁::TT, x₀::Vector, x₁::Vector, s::Int) where {TT}
+function _euler_extrapolation_ode!(v::Function, t₀::TT, t₁::TT, x₀::Vector, x₁::Vector, s::Int) where {TT}
     @assert size(x₀) == size(x₁)
 
     F   = collect(1:(s+1))
@@ -35,4 +51,86 @@ function euler_extrapolation(v::Function, t₀::TT, t₁::TT, x₀::Vector, x₁
     end
 
     aitken_neville!(σ, pts, zero(TT), x₁)
+    return x₁
 end
+
+
+# struct EulerExtrapolation <: Extrapolation
+#     s::Int
+# end
+
+# function Common.evaluate!(extrap::EulerExtrapolation, v::Function, t₀::TT, t₁::TT, x₀::Vector{DT}, x₁::Vector{DT}) where {DT,TT}
+#     _euler_extrapolation_ode!(v, t₀, t₁, x₀, x₁, extrap.s)
+# end
+
+# function Common.evaluate!(extrap::EulerExtrapolation, v::Function, f::Function, t₀::TT, t₁::TT, q₀::Vector{DT}, q₁::Vector{DT}, p₀::Vector{DT}, p₁::Vector{DT}) where {DT,TT}
+#     _euler_extrapolation_pode!(v, t₀, t₁, q₀, q₁, p₀, p₁, extrap.s)
+# end
+
+
+struct EulerExtrapolationODE{VT} <: Extrapolation
+    s::Int
+    v::VT
+
+    function EulerExtrapolationODE{VT}(v, s) where {VT}
+        new(s,v)
+    end
+end
+
+function EulerExtrapolationODE(v::VT, s::Int) where {VT}
+    EulerExtrapolationODE{VT}(v, s)
+end
+
+function EulerExtrapolation(equ::ODE, s::Int)
+    EulerExtrapolationODE(_get_v(equ), s)
+end
+
+function Common.evaluate!(extrap::EulerExtrapolationODE, t₀::TT, t₁::TT, x₀::Vector{DT}, x₁::Vector{DT}) where {DT,TT}
+    _euler_extrapolation_ode!(extrap.v, t₀, t₁, x₀, x₁, extrap.s)
+end
+
+
+# struct EulerExtrapolationIODE{VT,FT} <: Extrapolation
+#     s::Int
+#     v::VT
+#     f::FT
+    
+#     function EulerExtrapolationIODE{VT,FT}(v, f, s) where {VT,FT}
[…]
