+++ src/integrators/extrapolation/midpoint.jl
@@ -1,15 +1,32 @@
 
-"""
+@doc raw"""
 Midpoint extrapolation method with arbitrary order p.
 
-    v:  function to compute vector field
-    t₀: initial time
-    t₁: final   time
-    x₀: initial value
-    x₁: final   value
-    s:  number of interpolations (order p=2s+2)
+Solves the ordinary differential equation
+```math
+\begin{aligned}
+\dot{x} &= v(t, x) , &
+x(t_0) &= x_0 ,
+\end{aligned}
+```
+for $x_1 = x(t_1)$.
+
+Call with
+```julia
+_midpoint_extrapolation_ode!(v, t₀, t₁, x₀, x₁, s)
+```
+
+where
+
+* `v`:  function to compute vector field with signature `v(t,x,ẋ)`
+* `t₀`: initial time
+* `t₁`: final   time
+* `x₀`: initial value $x_0 = x(t_0)$
+* `x₁`: final   value $x_1 = x(t_1)$
+* `s`:  number of interpolations (order $p=2s+2$)
+
 """
-function midpoint_extrapolation(v::Function, t₀::TT, t₁::TT, x₀::Vector{DT}, x₁::Vector{DT}, s::Int) where {DT,TT}
+function _midpoint_extrapolation_ode!(v::Function, t₀::TT, t₁::TT, x₀::Vector{DT}, x₁::Vector{DT}, s::Int) where {DT,TT}
     @assert size(x₀) == size(x₁)
 
     local F   = [2i*one(TT) for i in 1:(s+1)]
@@ -42,23 +59,65 @@ function midpoint_extrapolation(v::Function, t₀::TT, t₁::TT, x₀::Vector{DT
     end
 
     aitken_neville!(σ², pts, zero(TT), x₁)
+    return x₁
 end
 
 
-"""
+struct MidpointExtrapolationODE{VT} <: Extrapolation
+    s::Int
+    v::VT
+
+    function MidpointExtrapolationODE{VT}(v, s) where {VT}
+        new(s,v)
+    end
+end
+
+function MidpointExtrapolationODE(v::VT, s::Int) where {VT}
+    MidpointExtrapolationODE{VT}(v, s)
+end
+
+function MidpointExtrapolation(equ::ODE, s::Int)
+    MidpointExtrapolationODE(_get_v(equ), s)
+end
+
+function Common.evaluate!(extrap::MidpointExtrapolationODE, t₀::TT, t₁::TT, x₀::Vector{DT}, x₁::Vector{DT}) where {DT,TT}
+    _midpoint_extrapolation_ode!(extrap.v, t₀, t₁, x₀, x₁, extrap.s)
+end
+
+
+@doc raw"""
 Midpoint extrapolation method with arbitrary order p.
 
-    v:  function to compute vector field
-    f:  function to compute force  field
-    t₀: initial time
-    t₁: final   time
-    q₀: initial positions
-    p₀: initial momenta
-    q₁: final   positions
-    p₁: final   momenta
-    s:  number of interpolations (order p=2s+2)
+Solves the implicit differential equation
+```math
+\begin{aligned}
+\dot{q} &= v(t, q, p) , &
+q(t_0) &= q_0 , \\
+\dot{p} &= f(t, q, v) , &
+p(t_0) &= p_0 , 
+\end{aligned}
+```
[…]
