+++ test/integrators/extrapolation_tests.jl
@@ -6,114 +6,210 @@ using GeometricIntegrators.Tableaus
 using GeometricProblems.HarmonicOscillator
 using Test
 
-using GeometricIntegrators.Equations: _get_v
-using GeometricIntegrators.Integrators: _euler_extrapolation_ode!, midpoint_extrapolation
+using GeometricIntegrators.Equations: _get_v, _get_f, _get_v̄, _get_f̄
+using GeometricIntegrators.Integrators: _euler_extrapolation_ode!, _midpoint_extrapolation_ode!,
+                                        _midpoint_extrapolation_iode!, _midpoint_extrapolation_pode!
 
 const Δt = 0.1
 const t₀ = 0.0
 const t₁ = t₀ + Δt
 
 ode  = harmonic_oscillator_ode()
+pode = harmonic_oscillator_pode()
+iode = harmonic_oscillator_iode()
 
 
-# Create Solution Arrays
+# Create ODE Solution Arrays
 
-q₀ = ode.q₀[1]
-qᵢ = zero(q₀)
+x₀ = ode.q₀[1]
+xᵢ = zero(x₀)
 
-vₚ = zero(q₀)
-v₀ = zero(q₀)
-vₙ = zero(q₀)
-vᵢ = zero(q₀)
+ẋₚ = zero(x₀)
+ẋ₀ = zero(x₀)
+ẋₙ = zero(x₀)
+ẋᵢ = zero(x₀)
 
 
-# Compute Reference Solution
+# Compute Reference Solution for ODEs
 
 ref_prev = integrate(ode, TableauGauss(8), -Δt, 1)
 ref_next = integrate(ode, TableauGauss(8), +Δt, 1)
 
 tₚ = ref_prev.t[end]
-qₚ = ref_prev.q[end]
+xₚ = ref_prev.q[end]
 
 tₙ = ref_next.t[end]
+xₙ = ref_next.q[end]
+
+ode.v(tₚ, xₚ, ẋₚ, ode.parameters)
+ode.v(t₀, x₀, ẋ₀, ode.parameters)
+ode.v(tₙ, xₙ, ẋₙ, ode.parameters)
+
+
+# Create PODE Solution Arrays
+
+q₀ = pode.q₀[1]
+p₀ = pode.p₀[1]
+
+qᵢ = zero(q₀)
+pᵢ = zero(p₀)
+
+q̇ₚ = zero(q₀)
+q̇₀ = zero(q₀)
+q̇ₙ = zero(q₀)
+q̇ᵢ = zero(q₀)
+
+ṗₚ = zero(p₀)
+ṗ₀ = zero(p₀)
+ṗₙ = zero(p₀)
+ṗᵢ = zero(p₀)
+
+
+# Compute Reference Solution for PODEs
+
+ref_prev = integrate(pode, PartitionedTableauGauss(8), -Δt, 1)
+ref_next = integrate(pode, PartitionedTableauGauss(8), +Δt, 1)
+
+qₚ = ref_prev.q[end]
+pₚ = ref_prev.p[end]
+
 qₙ = ref_next.q[end]
+pₙ = ref_next.p[end]
 
+pode.v(tₚ, qₚ, pₚ, q̇ₚ, ode.parameters)
[…]
