### 2021-05-18 fed25fb
Generalise extrapolation function interfaces.
