+++ test/integrators/extrapolation_tests.jl
@@ -2,17 +2,19 @@
 using GeometricIntegrators.Common
 using GeometricIntegrators.Config
 using GeometricIntegrators.Integrators
+using GeometricIntegrators.Integrators.VPRK
 using GeometricIntegrators.Tableaus
 using GeometricProblems.HarmonicOscillator
 using Test
 
 using GeometricIntegrators.Equations: _get_v, _get_f, _get_v̄, _get_f̄
-using GeometricIntegrators.Integrators: _euler_extrapolation_ode!, _midpoint_extrapolation_ode!,
-                                        _midpoint_extrapolation_iode!, _midpoint_extrapolation_pode!
+using GeometricIntegrators.Integrators: _euler_extrapolation_ode!, _hermite_extrapolation!,
+                                        _midpoint_extrapolation_ode!, _midpoint_extrapolation_iode!, _midpoint_extrapolation_pode!
 
 const Δt = 0.1
 const t₀ = 0.0
 const t₁ = t₀ + Δt
+const t₂ = t₁ + Δt
 
 ode  = harmonic_oscillator_ode()
 pode = harmonic_oscillator_pode()
@@ -22,10 +24,14 @@ iode = harmonic_oscillator_iode()
 # Create ODE Solution Arrays
 
 x₀ = ode.q₀[1]
+x₁ = zero(x₀)
+x₂ = zero(x₀)
 xᵢ = zero(x₀)
 
 ẋₚ = zero(x₀)
 ẋ₀ = zero(x₀)
+ẋ₁ = zero(x₀)
+ẋ₂ = zero(x₀)
 ẋₙ = zero(x₀)
 ẋᵢ = zero(x₀)
 
@@ -41,55 +47,16 @@ xₚ = ref_prev.q[end]
 tₙ = ref_next.t[end]
 xₙ = ref_next.q[end]
 
+tᵢ = tₙ
+
 ode.v(tₚ, xₚ, ẋₚ, ode.parameters)
 ode.v(t₀, x₀, ẋ₀, ode.parameters)
 ode.v(tₙ, xₙ, ẋₙ, ode.parameters)
 
 
-# Create PODE Solution Arrays
-
-q₀ = pode.q₀[1]
-p₀ = pode.p₀[1]
-
-qᵢ = zero(q₀)
-pᵢ = zero(p₀)
-
-q̇ₚ = zero(q₀)
-q̇₀ = zero(q₀)
-q̇ₙ = zero(q₀)
-q̇ᵢ = zero(q₀)
-
-ṗₚ = zero(p₀)
-ṗ₀ = zero(p₀)
-ṗₙ = zero(p₀)
-ṗᵢ = zero(p₀)
-
-
-# Compute Reference Solution for PODEs
-
-ref_prev = integrate(pode, PartitionedTableauGauss(8), -Δt, 1)
-ref_next = integrate(pode, PartitionedTableauGauss(8), +Δt, 1)
-
-qₚ = ref_prev.q[end]
-pₚ = ref_prev.p[end]
-
-qₙ = ref_next.q[end]
-pₙ = ref_next.p[end]
-
[…]
