### 2021-05-18 247d66e
Revise Hermite extrapolation.
