+++ src/integrators/initial_guess/initial_guess_pode.jl
@@ -22,7 +22,7 @@ struct InitialGuessPODE{TT, VT, FT, IT <: Extrapolation}
 end
 
 function InitialGuessPODE(interp::Type{<:Extrapolation}, v::Function, f::Function, Δt)
-    int = interp(zero(Δt), one(Δt), Δt)
+    int = interp(zero(Δt), Δt)
     InitialGuessPODE(int, v, f, Δt)
 end
 
@@ -91,7 +91,7 @@ function Common.evaluate!(ig::InitialGuessPODE{TT},
         @warn "q₀ and q₁ in initial guess are identical! Setting q=q₁."
         guess_q .= q₁
     else
-        evaluate!(ig.int, q₀, q₁, v₀, v₁, one(TT)+c_q, guess_q)
+        evaluate!(ig.int, q₀, q₁, v₀, v₁, (one(TT)+c_q)*ig.Δt, guess_q)
     end
 end
 
@@ -115,7 +115,7 @@ function Common.evaluate!(ig::InitialGuessPODE{TT},
         guess_q .= q₁
         guess_v .= 0
     else
-        evaluate!(ig.int, q₀, q₁, v₀, v₁, one(TT)+c_q, guess_q, guess_v)
+        evaluate!(ig.int, q₀, q₁, v₀, v₁, (one(TT)+c_q)*ig.Δt, guess_q, guess_v)
     end
 end
 
@@ -138,14 +138,14 @@ function Common.evaluate!(ig::InitialGuessPODE{TT},
         @warn "q₀ and q₁ in initial guess are identical! Setting q=q₁."
         guess_q .= q₁
     else
-        evaluate!(ig.int, q₀, q₁, v₀, v₁, one(TT)+c_q, guess_q)
+        evaluate!(ig.int, q₀, q₁, v₀, v₁, (one(TT)+c_q)*ig.Δt, guess_q)
     end
 
     if p₀ == p₁
         @warn "p₀ and p₁ in initial guess are identical! Setting p=p₁."
         guess_p .= p₁
     else
-        evaluate!(ig.int, p₀, p₁, f₀, f₁, one(TT)+c_p, guess_p)
+        evaluate!(ig.int, p₀, p₁, f₀, f₁, (one(TT)+c_p)*ig.Δt, guess_p)
     end
 end
 
@@ -173,7 +173,7 @@ function Common.evaluate!(ig::InitialGuessPODE{TT},
         guess_q .= q₁
         guess_v .= 0
     else
-        evaluate!(ig.int, q₀, q₁, v₀, v₁, one(TT)+c_q, guess_q, guess_v)
+        evaluate!(ig.int, q₀, q₁, v₀, v₁, (one(TT)+c_q)*ig.Δt, guess_q, guess_v)
     end
 
     if p₀ == p₁
@@ -181,6 +181,6 @@ function Common.evaluate!(ig::InitialGuessPODE{TT},
         guess_p .= p₁
         guess_f .= 0
     else
-        evaluate!(ig.int, p₀, p₁, f₀, f₁, one(TT)+c_p, guess_p, guess_f)
+        evaluate!(ig.int, p₀, p₁, f₀, f₁, (one(TT)+c_p)*ig.Δt, guess_p, guess_f)
     end
 end
