+++ src/integrators/extrapolation/hermite.jl
@@ -1,11 +1,36 @@
 @doc raw"""
 # Hermite's Interpolating Polynomials
 
-Here, we implement a two point Hermite interpolation function which passes
+Implements a two point Hermite inter-/extrapolation function which passes
 through the function and its first derivative for the interval ``[0,1]``.
 The polynomial is determined by four constraint equations, matching the
 function and its derivative at the points ``0`` and ``1``.
 
+Call with one of the following methods
+```julia
+_hermite_extrapolation!(t₀, t₁, x₀, x₁, ẋ₀, ẋ₁, t, x)
+_hermite_extrapolation!(t₀, t₁, x₀, x₁, ẋ₀, ẋ₁, t, x, ẋ)
+_hermite_extrapolation!(v, t₀, t₁, x₀, x₁, t, x)
+_hermite_extrapolation!(v, t₀, t₁, x₀, x₁, t, x, ẋ)
+```
+
+where
+
+* `t₀`: first  sample time $t_0$
+* `t₁`: second sample time $t_1$
+* `x₀`: first  solution value $x_0 = x(t_0)$
+* `x₁`: second solution value $x_1 = x(t_1)$
+* `ẋ₀`: first  vector field value $ẋ_0 = v(t_0, x(t_0))$
+* `ẋ₁`: second vector field value $ẋ_1 = v(t_1, x(t_1))$
+* `v`:  function to compute vector field with signature `v(t,x,ẋ)`
+* `t`:  time $t$ to extrapolate
+* `x`:  extrapolated solution value $x(t)$
+* `v`:  extrapolated vector field value $ẋ(t)$
+
+
+#### Derivation
+
+The interpolation works as follows:
 Start by defining the 3rd degree polynomial and its derivative by
 ```math
 \begin{aligned}
@@ -89,59 +114,96 @@ g'(1) &= f'_1 .
 \end{aligned}
 ```
 """
-struct HermiteExtrapolation{T} <: Extrapolation
-    x₀::T
-    x₁::T
-    Δx::T
-
-    function HermiteExtrapolation{T}(x₀, x₁, Δx) where {T}
-        new(x₀, x₁, Δx)
-    end
-end
-
-function HermiteExtrapolation(x₀::T, x₁::T, Δx::T) where {T}
-    HermiteExtrapolation{T}(x₀, x₁, Δx)
-end
+function _hermite_extrapolation! end
 
 
-function Common.evaluate!(int::HermiteExtrapolation{TT}, y₀::AbstractArray{DT}, y₁::AbstractArray{DT}, f₀::AbstractArray{DT}, f₁::AbstractArray{DT}, x::TT, y::AbstractArray{DT}) where {DT,TT}
+function _hermite_extrapolation!(t₀::TT, t₁::TT, x₀::AbstractArray{DT}, x₁::AbstractArray{DT}, ẋ₀::AbstractArray{DT}, ẋ₁::AbstractArray{DT}, t::TT, x::AbstractArray{DT}) where {DT,TT}
     local a₀::TT
     local a₁::TT
     local b₀::TT
     local b₁::TT
+    local Δt = t₁ - t₀
+    local s = (t - t₀) / Δt
 
-    # Interpolate y values at required locations
-    if x == int.x₀
-        y .= y₀
[…]
