### 2021-05-18 e3a3bb0
Clean up and fix initial guess tests.
