+++ test/integrators/initial_guess_tests.jl
@@ -6,13 +6,13 @@ using GeometricIntegrators.Tableaus
 using GeometricProblems.LotkaVolterra2d
 using Test
 
-using GeometricIntegrators.Equations: _get_v̄, _get_f̄
+using GeometricIntegrators.Equations: _get_v, _get_f, _get_v̄, _get_f̄
 
 int = get_config(:ig_extrapolation)
 
-const Δt = 0.01
-const q₀ = [1.0, 1.0]
-const parameters = (a₁=1.0, a₂=1.0, b₁=-1.0, b₂=-2.0)
+Δt = 0.01
+q₀ = [1.0, 1.0]
+parameters = (a₁=1.0, a₂=1.0, b₁=-1.0, b₂=-2.0)
 
 ode  = lotka_volterra_2d_ode(q₀; params=parameters)
 pode = lotka_volterra_2d_pode(q₀; params=parameters)
@@ -66,7 +66,7 @@ ode.v(tₙ, qₙ, vₙ, ode.parameters)
 
 # InitialGuessODE
 
-igode = InitialGuessODE(int, (t,q,v) -> ode.v(t, q, v, ode.parameters), Δt)
+igode = InitialGuessODE(int, _get_v(ode), Δt)
 
 t₀ = ode.t₀
 q₀ = ode.q₀[begin]
@@ -85,16 +85,23 @@ ode.v(t₀, q₀, v₀, ode.parameters)
 initialize!(igode, t₀, q₀, v₀, t₁, q₁, v₁)
 evaluate!(igode, q₁, v₁, q₀, v₀, q₂, v₂, t₂)
 
-@test q₁ ≈ qₚ atol=1E-6
-@test v₁ ≈ vₚ atol=1E-6
+# println("IG-ODE")
+# println(q₁ .- qₚ)
+# println(v₁ .- vₚ)
+# println(q₂ .- qₙ)
+# println(v₂ .- vₙ)
+# println()
 
-@test q₂ ≈ qₙ atol=1E-6
-@test v₂ ≈ vₙ atol=1E-3
+@test q₁ ≈ qₚ atol=1E-14
+@test v₁ ≈ vₚ atol=1E-14
+
+@test q₂ ≈ qₙ atol=1E-8
+@test v₂ ≈ vₙ atol=1E-5
 
 
 # InitialGuessIODE
 
-igiode = InitialGuessIODE(int, (t,q,v) -> iode.v̄(t, q, v, iode.parameters), (t,q,p,v) -> iode.f̄(t, q, p, v, iode.parameters), Δt)
+igiode = InitialGuessIODE(int, _get_v̄(iode), _get_f̄(iode), Δt)
 
 t₀ = iode.t₀
 q₀ = iode.q₀[begin]
@@ -119,18 +126,23 @@ iode.v̄(t₀, q₀, v₀, iode.parameters)
 initialize!(igiode, t₀, q₀, p₀, v₀, f₀, t₁, q₁, p₁, v₁, f₁)
 evaluate!(igiode, q₁, p₁, v₁, f₁, q₀, p₀, v₀, f₀, q₂, v₂, t₂)
 
-@test q₁ ≈ qₚ atol=1E-6
-@test v₁ ≈ vₚ atol=1E-6
+# println("IG-IODE")
+# println(q₁ .- qₚ)
+# println(v₁ .- vₚ)
+# println(q₂ .- qₙ)
+# println(v₂ .- vₙ)
+# println()
 
-@test q₂ ≈ qₙ atol=1E-6
-@test v₂ ≈ vₙ atol=1E-3
+@test q₁ ≈ qₚ atol=1E-14
[…]
