### 2021-05-18 d756710
Update docstring for Aitken-Nevill algorithm.
