+++ src/integrators/dvi/integrators_cmdvi.jl
@@ -1,0 +1,216 @@
+
+import ForwardDiff
+
+"Parameters for right-hand side function of degenerate variational integrator."
+mutable struct ParametersCMDVI{DT, TT, D, ET <: NamedTuple} <: Parameters{DT,TT}
+    equs::ET
+    Δt::TT
+
+    t::TT
+    q::Vector{DT}
+    v::Vector{DT}
+    θ::Vector{DT}
+
+    function ParametersCMDVI{DT,D}(equs::ET, Δt::TT) where {DT, TT, D, ET <: NamedTuple}
+        new{DT, TT, D, ET}(equs, Δt, zero(TT), zeros(DT,D), zeros(DT,D), zeros(DT,D))
+    end
+end
+
+
+@doc raw"""
+Degenerate variational integrator cache.
+
+### Fields
+
+* `q`: internal stages of solution
+* `v`: internal stages of vector field
+* `Θ`: implicit function evaluated on solution
+* `f`: vector field of implicit function
+"""
+struct IntegratorCacheCMDVI{DT,D,S} <: ODEIntegratorCache{DT,D}
+    q::Vector{DT}
+    v::Vector{DT}
+    θ::Vector{DT}
+
+    q̄::Vector{DT}
+    θ̄::Vector{DT}
+    f̄::Vector{DT}
+
+    function IntegratorCacheCMDVI{DT,D,S}() where {DT,D,S}
+        new(zeros(DT,D), zeros(DT,D), zeros(DT,D),
+            zeros(DT,D), zeros(DT,D), zeros(DT,D))
+    end
+end
+
+function IntegratorCache{ST}(params::ParametersCMDVI{DT,TT,D,S}; kwargs...) where {ST,DT,TT,D,S}
+    IntegratorCacheCMDVI{ST,D,S}(; kwargs...)
+end
+
+@inline CacheType(ST, ::ParametersCMDVI{DT,TT,D,S}) where {DT,TT,D,S} = IntegratorCacheCMDVI{ST,D,S}
+
+
+"""
+Implicit midpoint Runge-Kutta integrator.
+"""
+struct IntegratorCMDVI{DT, TT, D, PT <: ParametersCMDVI{DT,TT},
+                                   ST <: NonlinearSolver{DT},
+                                   IT <: InitialGuessIODE{TT}} <: AbstractIntegratorRK{DT,TT}
+    params::PT
+    solver::ST
+    iguess::IT
+    caches::CacheDict{PT}
+
+    function IntegratorCMDVI(params::ParametersCMDVI{DT,TT,D}, solver::ST, iguess::IT, caches) where {DT,TT,D,ST,IT}
+        new{DT, TT, D, typeof(params), ST, IT}(params, solver, iguess, caches)
+    end
+
+    function IntegratorCMDVI{DT,D}(equations::NamedTuple, Δt::TT) where {DT,TT,D}
+        # create params
+        params = ParametersCMDVI{DT,D}(equations, Δt)
+
+        # create cache dict
+        caches = CacheDict(params)
+
+        # create solver
+        solver = create_nonlinear_solver(DT, 2D, params, caches)
+
+        # create initial guess
+        iguess = InitialGuessIODE(get_config(:ig_extrapolation), equations[:v̄], equations[:f̄], Δt)
+
+        # create integrator
+        IntegratorCMDVI(params, solver, iguess, caches)
+    end
+
+    function IntegratorCMDVI(equation::Union{IODE{DT}, LODE{DT}}, Δt::TT; kwargs...) where {DT,TT}
+        IntegratorCMDVI{DT, ndims(equation)}(get_functions(equation), Δt; kwargs...)
+    end
+end
+
+
[…]
