### 2021-06-24 1d79528
Add Julia v1.7 to CI.
