+++ .github/workflows/CI.yml
@@ -12,7 +12,8 @@ jobs:
       matrix:
         version:
           - '1.5'
-          - '^1.6.0-0'
+          - '1.6'
+          - '^1.7.0-0'
         os:
           - ubuntu-latest
           - macOS-latest
