### 2021-06-24 68e510b
Fixes in VPRK and VSPARK initialisation.
