### 2021-06-24 0f2177e
Fixes in TimeSeries getindex methods.
