### 2021-07-23 #2954 @michakraus
Hi @VivianePons, my apologies, this starts to become odd. Unfortunately, my last message still applies. I have been too busy with other things and hardly any time for coding. It seems I have the todo list of hydra: whenever I tick off one point there are at least two new ones.
However, I have been working on the documentation and some other issues for the last few days, and hope to finish the last missing points next week. This time for real.
In any case I will post a status report next week, stating which requests of the referees have been met and which are still open.
