### 2021-08-23 #2954 @michakraus
I have been off the grid for about two weeks myself, but am now back to work. A short update: we just released GeometricIntegrators v0.9 which addresses many requests of the reviewers as well as many other changes. The main points still missing are
- [ ] add more examples (different problem types)
- [ ] improve the documention of equations, solutions, simulations and add developer docs
- [ ] add an overview that shows which algorithms work for which problems
- [ ] add tests and examples for general types
- [ ] add more Base.show methods for important structures

We will work on these (and some other) issues in the next few weeks with the goal of releasing GeometricIntegrators v1.0 as the reference version for the paper.
