### 2021-11-24 #2954 @michakraus
Hi @VivianePons, sorry for the late reply. I was quite busy last week, but now, for the first time this year, I am free of any deadlines. I plan to spent most of my time until Christmas on software development. Unfortunately, in the last three months since my last status report (it's hard to comprehend that it has already been three months!), I did not have much time. Still, a few things that got done:

- We refactored our solution and equation modules, partially because they turned out to be a bit too complicated and over-engineered, and partially because of needs for other projects. I did not want to write complicated documentation for complicated modules, but rather simplify first. This is almost finished, but it requires some refactoring in the main library to account for modifications in interfaces.
- We did a lot of refactoring to adhere to standard Julia dev guidelines (e.g., argument order in functions), which were not yet in place when some parts of the library were originally written.
-  Adding tests and examples for general types, we realised that there are quite a few places in the library that do not yet support this. We started refactoring this, but this is ongoing.

To make a long story short: I hope to be done before Christmas... it is about time!
