### 2022-05-24 #2954 @michakraus
Dear @VivianePons and @arfon, apologies for the late reply. I have been off duty for the better part of the last 5~6 weeks.

We made some progress on the open issues, but not as much as I had hoped for. In the last months, we could not allocate a whole lot of resources to this project. However, most of the open issues need to be addressed in order to make progress on other projects or ours. So I have some hope that this will be taken care of soon. But we are currently fighting on several fronts, and it is hard to give an ETA. I will try to provide an update with a more concrete timeline by the end of June.
