### 2022-06-21 #2954 @arfon
:wave: folks. I'm going to be taking on this paper as @VivianePons is stepping down as a JOSS editor. Thanks for all of your help with JOSS @VivianePons!
