### 2022-07-24 #2954 @arfon
@michakraus - it's now nearly the end of July and we've not heard back from you in some time. Reviews in JOSS can sometimes take a while, but this submission looks to have been in review now for a very long time.

I propose we agree a firm deadline for you addressing the outstanding feedback (I propose one month from today), and if you're unable to meet this, we would proceed reject (or you could withdraw).

In this situation, a resubmission would be possible provided the feedback from the original review had been addressed.
