### 2022-07-28 #2954 @michakraus
@arfon - apologies for being silent for so long. I very much agree that this has taken much longer than it should have. As mentioned before, boundary conditions have not been in favour of working on this project in the last months.
Nonetheless, a few weeks back we merged a big PR leaving only one major problem open (refactoring solutions) in addition to some more documentation.
I will be out of office for most of August, so it would be impossible to meet the suggested deadline. However, if we can agree on end of September there is a fair chance we can finish until then.
