### 2022-09-30 #2954 @michakraus
> * [ ]  As stated in the manuscript, the package "is designed to minimize overhead and maximize performance". However, there seem to be some caveats. For example, only LU decomposition seems to be implemented as linear solver, limiting the size of problems that can be solved. Hence, I would like to know the anticipated scope of applications and the size of ODE systems? GeometricIntegrators.jl looks to be good for small ODEs. Can I also apply the integrators to spatial semidiscretizations of PDEs? If so, is it okay for moderate 1D problems or also for large 3D problems on a single workstation? It would be nice to mention this range of applications in the manuscript.

In the manuscript as well as the documentation, we added the following section:

_The library aims at providing efficient implementations of diverse algorithms in order to be able to perform simulations and benchmarks with millions or even billions of time steps that facilitate the study of the long-time behaviour of both numerical algorithms and dynamical systems.
The current scope of applications is mainly small- to mid-size systems of differential equations, e.g., systems of ordinary differential equations or semidiscretisations of partial differential equations with a moderate number of degrees of freedom.
It is envisaged that in the future GeometricIntegrators.jl will also be able to address larger problems, especially semidiscretisations of partial differential equations in higher dimensions.
In particular, this requires interfaces to appropriate iterative and parallel linear solvers, which are still lacking._

> * [ ]  It would be nice to have links to all papers mentioned in the documentation at Home/References.

We added a _Bibliography_ section to the docs using DocumenterCitations. It's not beautiful but does the job.

> * [ ]  When reading the manuscript, it is not really clear what kind of algorithms GeometricIntegrators.jl implements. It might be helpful to give some examples for each of the categories mentioned.

In the paper, we only list the various families of algorithms:

_The implemented algorithms include explicit, implicit and partitioned Runge-Kutta methods, SPARK methods, splitting methods, symplectic methods and variational integrators._

We will add specific examples in the docs, but that is still WIP.

* [ ] Add examples for the different integrator families to docs.

> * [ ]  It would be interesting to know whether some of the methods implemented in GeometricIntegrators.jl are also available from other sources and mention differences and/or relative advantages/disadvantages. For example, some geometric numerical integrators are available as Fortran codes from the website of Ernst Hairer. A lot of Runge-Kutta methods are also available in Julia via DifferentialEquations.jl. What is the reason for having another implementation in Julia of these schemes? How do the methods of GeometricIntegrators compare to that?

I tried to account for that in the following reformulation:

_*GeometricIntegrators.jl* provides a comprehensive library of geometric integration algorithms as well as some non-geometric algorithms.
It collects native Julia implementations of many known methods under a unified interface.
Once a problem is implemented in the GeometricIntegators framework, all of its algorithms can immediately be applied and their performance evaluated.
This facilitates numerical experiments with a wide variety of algorithms, simplifies benchmarking, and makes it easy to identify the best algorithm for a given problem._

> * [ ]  The docs "Integrators/Splitting Methods" and "Integrators/Stochastic Integrators" are empty for v0.7.1.

Docs for Splitting Methods have been added. Stochastic Integrators were moved to a different package.

> * [ ]  The docs for the modules "Basis Functions", "Linear Solvers", "Nonlinear Solvers", and "Quadratures" are nearly empty. It looks like these modules were moved to additional packages recently?

All of these modules were moved to separate packages in the JuliaGNI organisation.

> * [ ]  It would be nice to improve docstrings since they don't really help to explain how to call some functions, cf. [Improve docstrings JuliaGNI/GeometricIntegrators.jl#107](https://github.com/JuliaGNI/GeometricIntegrators.jl/issues/107).

Many docstrings have been updated and extended, but this is an ongoing endeavour.

> * [ ]  It would be nice if some standard `struct`s used many times could be printed nicely in the REPL by overloading the appropriate methods of `Base.show`.

Dito. `Base.show` methods have been added for various data structures, but there is certainly more that could be done.

If you have specific structures in mind, please let us know.

> * [ ]  Is adaptive time-stepping or some way to estimate an appropriate step size supported?

A paragraph has been added to the paper:

_As most geometric integrators are not easily combined with time step adaption in a structure-preserving way, GeometricIntegrators.jl does not provide any general infrastructure for adaptive time stepping. Nonetheless, individual integrators can implement their own adaptivity strategies as long as they provide a solution at a predefined, equidistant series of time steps._

There's also a page "Developer/Adaptive Time Stepping" that explains this in some more detail but still lacks an example.

* [ ] Add example for adaptive time stepping.

> * [ ]  It should be made possible to get some reference for the algorithms implemented in this package to be able to use and cite them for research, cf. [Get references for algorithms JuliaGNI/GeometricIntegrators.jl#108](https://github.com/JuliaGNI/GeometricIntegrators.jl/issues/108).

References to the various Runge-Kutta methods have been added to the docstrings of RungeKutta.jl. These docstrings can also be seen in the GeometricIntegrators docs under Tableaus -> Runge-Kutta. However, they do not appear in the Bibliography.
We still need to add some references for the other methods:

* [ ] Add references for non-Runge-Kutta methods.

> * [ ]  Another example from the docs does not work, cf. [TestProblems not defined JuliaGNI/GeometricIntegrators.jl#109](https://github.com/JuliaGNI/GeometricIntegrators.jl/issues/109)

This should be fixed, but the tutorial still needs to be extended a little bit further, also w.r.t. the next point.

* [ ] Extend Tutorial.

> * [ ]  It would be nice to see a few more examples for different problem types in the docs.

We started to work on this, but did not finish, yet.

* [ ] add more examples for different problem types

> * [ ]  It looks like not all problem types and methods are compatible with each other, e.g. `IDAE` and `TableauExplicitEuler`. It might be good to mention something like this in the docs.

I started to build a list of all methods (including all possible combinations of e.g. Integrators and Tableaus), where entries have traits for the problem type they are applicable to. As soon as this is finished, we can auto-generate this overview for the docs.

* [ ] add an overview that shows which algorithms work for which problems

> * [ ]  It is a bit confusing that an argument modified by a function is passed as last argument - the Julia standard convention is to pass modified arguments at first. What was the reason for this decision? It could be helpful to include a warning with an explanation in the docs.

This has been changed to comply with the Julia standard convention.

> * [ ]  `P_{n,i}` is used twice for different things in the second equation of https://juliagni.github.io/GeometricIntegrators.jl/v0.7.1/integrators/vprk/

This may seem confusing on first sight, but is actually correct, cf. e.g. Marsden & West, 2001.

> * [ ]  Is there a specific reason for restricting the type of the right-hand sides to `Function`s? For example, I cannot use callable `struct`s, which would be my first choice when dealing with differential equations with parameters (or encapsulating more complex right-hand sides).
>   Hence, it looks like closures are the only way to deal with parameters right now.

It should now be possible everywhere to use `Callable` instead of just `Function`.

Parameters are dealt with in a very similar way as in DifferentialEquations.jl.

> * [ ]  Quoting [[REVIEW]: GeometricIntegrators.jl: Geometric Numerical Integration in Julia #2954 (comment)](https://github.com/openjournals/joss-reviews/issues/2954#issuecomment-765312446)
>   > In the next step we want to put several of the specialised integrators into separate packages, most and foremost StochasticIntegrators.jl, which we plan to publish separately.
>
>   I'm not sure whether it would be better to make the splitting before publishing this work. Otherwise, the stochastic integrators would be in two papers - but that's not my decision, just a comment you might consider.

Meanwhile this has happened (some time ago).

> * [ ]  It would be good to have a top-level CONTRIBUTING file (mentioned also in the docs) so users see a link to it when they open a PR. It would also be good to note if you prefer users create a GitHub issue or use some other means for reporting problems.

A file `[CONTRIBUTING.md](https://github.com/JuliaGNI/GeometricIntegrators.jl/blob/master/CONTRIBUTING.md) has been added.

@ranocha Thanks once more for the helpful comments!
