### 2022-09-30 #2954 @michakraus
> Hey sorry, got far behind.

I guess I gave a new definition to "getting far behind".

> ```
> >Is adaptive time-stepping or some way to estimate an appropriate step size supported?
> ...
> ```
>
> I think this is worth commenting on because it's probably lost to the eyes of someone who's not deep in the field.
>
> * [ ]  I would say it would be good to add a note on this to the Statement of Need.

A paragraph has been added to the paper:

_As most geometric integrators are not easily combined with time step adaption in a structure-preserving way, GeometricIntegrators.jl does not provide any general infrastructure for adaptive time stepping. Nonetheless, individual integrators can implement their own adaptivity strategies as long as they provide a solution at a predefined, equidistant series of time steps._

There's also a page "Developer/Adaptive Time Stepping" that explains this in some more detail but still lacks an example.

> * [ ]  The developer documentation pages are empty.

Still mostly the case. Will be fixed asap.

* [ ] add developer docs.

> * [ ]  The interpolation docstring in https://juliagni.github.io/GeometricIntegrators.jl/latest/modules/interpolation/ doesn't mention how or when to use it.

This functionality has essentially been removed.

> * [ ]  https://juliagni.github.io/GeometricIntegrators.jl/latest/modules/simulations/ isn't very informative.

This feature and its docs has been disabled for now.

> * [ ]  https://juliagni.github.io/GeometricIntegrators.jl/latest/equations/ is empty.

This has been moved to GeometricEquations.jl. The docstrings there are rather complete, but still need a little love to account for recent refactorings.

* [ ] Revise GeometricEquations.jl docstrings.

> * [ ]  It looks like everything was made intentionally generically typed, but that doesn't seem to be tested? It might be good to throw a few Dual number tests in there just to see if everything works out like you'd expect.

Actually, some parts of the library (in particular GeometricEqations.jl) are generically typed, but in GeometricIntegrators itself, there is still a bit of refactoring needed to make this work with more than one or two integrators. Therefore we currently also have no tests for this.

> * [ ]  I don't see a link to GeometricExamples in the docs? That might be good to mention.

GeometricExamples holds scripts for some runs for papers. It would be more appropriate to mention GeometricProblems, which holds actual test problems. References to GeometricProblems have been added in some places.

@ChrisRackauckas Thanks for the comments!
