### 2022-09-30 #2954 @michakraus
@arfon To summarise, I think these are the open points, practically all concerning the documentation:

* [ ] revise and extend tutorial
* [ ] revise and extend documentation of integrators
* [ ] add examples for the different integrator families
* [ ] add overview that shows which algorithms work for which problems
* [ ] add developer docs
   * [ ] custom integrators
   * [ ] code integration (low-level interface)
   * [ ] example for adaptive time stepping
* [ ] add references for non-Runge-Kutta methods
* [ ] revise GeometricEquations.jl docstrings

I expect we need another week or two to take care of all of these points.
