### 2022-10-24 #2954 @michakraus
I couldn't agree more that this should be finished as soon as possible. I have been working on the docs for the better part of my available time in the past few weeks. Unfortunately everything got delayed once more as I have been sick in between. I pretty much finished the GeometricEquations.jl docs and the revision of the tutorial. I will spend the next days adding more examples and extending the tutorial. This should be done by the end of the week.
