### 2022-11-15 #2954 @michakraus
We added examples for all problem types, extended and fixed docstrings for problem types, extended and revised the tutorial. The latter made clear for which data structures `Base.show` methods are still missing. Adding more such methods was requested by one of the referees, but it was not quite clear, which core data structures are still lacking `Base.show` methods. Now it is, namely `GeometricProblem` and subtypes, `GeometricSolution` and subtypes, and some `Integrator` subtypes. We are working on this.

- [ ] add `Base.show` methods for
   - [ ] `GeometricProblem` and subtypes
   - [ ] `GeometricSolution` and subtypes
   - [ ] some `Integrator` subtypes

Further, we implemented a new `Method` type that makes selection of algorithms and setting their parameters easier, and simplifies documenting the compatibility of algorithms and problem types, as was also requested by one of the referees. This uncovered some issues with our current way of initialising variational integrators, which is overly complex and therefore also somewhat hard to document. We are currently trying to fix this.

- [ ] simplify selection and initialisation of Variational Partitioned Runge Kutta (VPRK) methods

I will try to provide another update before the weekend.
