### 2022-11-29 #2954 @michakraus
Well, as usual, things never turn out as expected. I did not fully finish before said weekend and was on conference last week.

I just ticket off a few of the points above: the tutorial, integrator and equation docs are in a good shape now. The algorithm overview is almost done. I am just lacking a good idea of how to propagate algorithm references to the newly added `Method` types. Once this is sorted, it's mostly a matter of adding a few more docstrings, `Base.show()` methods and the dev docs.

There's light at the end of the tunnel - and it is not a train!
