### 2022-12-02 #2954 @michakraus
Short update:
- the algorithm overview is done (for some integrators we still need to add method types)
- propagation of references is resolved and references are included in docstrings of most Method types
- `Base.show()` methods have been added for most Integrator and Method types
