### 2022-12-22 #2954 @michakraus
Unfortunately I have to report that I got stuck in some refactoring and that I was sick for about a week and thus was not able to take care of the open points. I will be off duty for a little more than two weeks starting tomorrow and continue in the new year.
