### 2023-09-04 #2954 @michakraus
As usual, I did not finish by the time I hoped for and have been traveling for the past five weeks. I cleaned up the docs and fixed many issues, but some empty or incomplete sections remain.

By now, I am somewhat hesitant to give any estimates. Compared to what has been done, it really is only an epsilon that is missing, but I can't give a reliable estimate of when I will be finished. I know that I will not have much time for software development in the next two weeks. After that, I should be fairly free, though. My best guess is that I can finish the remaining issues within one or two months (say mid~late October).

A recurrent theme in this project is that pretty much whenever I try to fix an issue raised by the referees or add missing documentation, I encounter serious issues in the code. I then try to fix those properly instead of just working around them or documenting the current state of affairs, which typically takes much more time than anticipated.
The most recent examples are the integrator and projection interfaces. When I tried to write the docs for these, I realized that the interfaces were neither defined nor implemented in a clean way. I fixed the integrator interface, but I am still working on the projection interface.

On the other hand, life is very nonlinear (four kids, a number of students, a book project, and a day job), time to work on this project is limited, and priorities change very dynamically. I need to satisfy external boundary conditions regarding prioritizing development goals, which may not align with the referees' requests. The package profited quite tremendously from the referees' comments, and I am very grateful for those, but often feature requests by our team pop up rather suddenly and need to be prioritized.

Apologies for the lengthy text. I just wanted to explain why all of this takes so long. But if you want to pull the ripcord on this submission, you have my complete understanding. I submitted the first draft more than 2.5 years ago, which is indeed a bit ridiculous. The fact of the matter is that we have somewhat opposing goals, and I fully realize this. You want to maintain reasonable publication times, while more often than not I need to work on aspects of the package that have nothing to do with the referee requests and consistently fail to provide a reliable timeline. In any case, I will continue to work towards meeting the referee's request.
