### 2023-09-27 #2954 @arfon
@michakraus – thank you for the detailed message, and acknowledging the 'somewhat opposing goals' here. At this point, having consulted with the other EiCs at JOSS, we're going to proceed to reject this submission. Should you find the time in the future to make the requested changes, we'd welcome a resubmission.

@ChrisRackauckas, @ranocha – thanks for your help here, sorry this didn't work out.
